Data.Vector

Description

A library for boxed vectors (that is, polymorphic arrays capable of holding any Haskell value). The vectors come in two flavours:

mutable

immutable

They support a rich interface of both list-like operations and bulk array operations.

For unboxed arrays, use Data.Vector.Unboxed.

Boxed vectors

data Vector a

Boxed vectors, supporting efficient slicing.

data MVector s a

Mutable boxed vectors keyed on the monad they live in (IO or ST s).

Accessors

Length information

length :: Vector a -> Int

O(1) Yield the length of the vector.

null :: Vector a -> Bool

O(1) Test whether a vector is empty.

Indexing

(!) :: Vector a -> Int -> a

O(1) Indexing.

(!?) :: Vector a -> Int -> Maybe a

O(1) Safe indexing.

head :: Vector a -> a

O(1) First element.

last :: Vector a -> a

O(1) Last element.

unsafeIndex :: Vector a -> Int -> a

O(1) Unsafe indexing without bounds checking.

unsafeHead :: Vector a -> a

O(1) First element, without checking if the vector is empty.

unsafeLast :: Vector a -> a

O(1) Last element, without checking if the vector is empty.

Monadic indexing

indexM :: Monad m => Vector a -> Int -> m a

O(1) Indexing in a monad.

The monad allows operations to be strict in the vector when necessary. Suppose vector copying is implemented like this:

copy mv v = ... write mv i (v ! i) ...

For lazy vectors, v ! i would not be evaluated which means that mv would unnecessarily retain a reference to v in each element written.

With indexM, copying can be implemented like this instead:

copy mv v = ... do
                  x <- indexM v i
                  write mv i x

Here, no references to v are retained because indexing (but not the element) is evaluated eagerly.

headM :: Monad m => Vector a -> m a

O(1) First element of a vector in a monad. See indexM for an explanation of why this is useful.

lastM :: Monad m => Vector a -> m a

O(1) Last element of a vector in a monad. See indexM for an explanation of why this is useful.

unsafeIndexM :: Monad m => Vector a -> Int -> m a

O(1) Indexing in a monad, without bounds checks. See indexM for an explanation of why this is useful.

unsafeHeadM :: Monad m => Vector a -> m a

O(1) First element in a monad, without checking for empty vectors. See indexM for an explanation of why this is useful.

unsafeLastM :: Monad m => Vector a -> m a

O(1) Last element in a monad, without checking for empty vectors. See indexM for an explanation of why this is useful.

Extracting subvectors (slicing)

slice

Arguments

:: Int
i starting index

-> Int
n length

-> Vector a

-> Vector a

O(1) Yield a slice of the vector without copying it. The vector must contain at least i+n elements.

init :: Vector a -> Vector a

O(1) Yield all but the last element without copying. The vector may not be empty.

tail :: Vector a -> Vector a

O(1) Yield all but the first element without copying. The vector may not be empty.

take :: Int -> Vector a -> Vector a

O(1) Yield at the first n elements without copying. The vector may contain less than n elements, in which case it is returned unchanged.

drop :: Int -> Vector a -> Vector a

O(1) Yield all but the first n elements without copying. The vector may contain less than n elements, in which case an empty vector is returned.

splitAt :: Int -> Vector a -> (Vector a, Vector a)

O(1) Yield the first n elements paired with the remainder, without copying.

Note that splitAt n v is equivalent to (take n v, drop n v), but slightly more efficient.

Since: 0.7.1

uncons :: Vector a -> Maybe (a, Vector a)

O(1) Yield the head and tail of the vector, or Nothing if the vector is empty.

Since: 0.12.2.0

unsnoc :: Vector a -> Maybe (Vector a, a)

O(1) Yield the last and init of the vector, or Nothing if the vector is empty.

Since: 0.12.2.0

unsafeSlice

Arguments

:: Int
i starting index

-> Int
n length

-> Vector a

-> Vector a

O(1) Yield a slice of the vector without copying. The vector must contain at least i+n elements, but this is not checked.

unsafeInit :: Vector a -> Vector a

O(1) Yield all but the last element without copying. The vector may not be empty, but this is not checked.

unsafeTail :: Vector a -> Vector a

O(1) Yield all but the first element without copying. The vector may not be empty, but this is not checked.

unsafeTake :: Int -> Vector a -> Vector a

O(1) Yield the first n elements without copying. The vector must contain at least n elements, but this is not checked.

unsafeDrop :: Int -> Vector a -> Vector a

O(1) Yield all but the first n elements without copying. The vector must contain at least n elements, but this is not checked.

Construction

Initialisation

empty :: Vector a

O(1) The empty vector.

singleton :: a -> Vector a

O(1) A vector with exactly one element.

replicate :: Int -> a -> Vector a

O(n) A vector of the given length with the same value in each position.

generate :: Int -> (Int -> a) -> Vector a

O(n) Construct a vector of the given length by applying the function to each index.

iterateN :: Int -> (a -> a) -> a -> Vector a

O(n) Apply the function \(\max(n - 1, 0)\) times to an initial value, producing a vector of length \(\max(n, 0)\). The 0th element will contain the initial value, which is why there is one less function application than the number of elements in the produced vector.

\( \underbrace{x, f (x), f (f (x)), \ldots}_{\max(0,n)\rm{~elements}} \)

Examples

Expand

>>> import qualified Data.Vector as V
>>> V.iterateN 0 undefined undefined :: V.Vector String
[]
>>> V.iterateN 4 (\x -> x <> x) "Hi"
["Hi","HiHi","HiHiHiHi","HiHiHiHiHiHiHiHi"]

Since: 0.7.1

Monadic initialisation

replicateM :: Monad m => Int -> m a -> m (Vector a)

O(n) Execute the monadic action the given number of times and store the results in a vector.

generateM :: Monad m => Int -> (Int -> m a) -> m (Vector a)

O(n) Construct a vector of the given length by applying the monadic action to each index.

iterateNM :: Monad m => Int -> (a -> m a) -> a -> m (Vector a)

O(n) Apply the monadic function \(\max(n - 1, 0)\) times to an initial value, producing a vector of length \(\max(n, 0)\). The 0th element will contain the initial value, which is why there is one less function application than the number of elements in the produced vector.

For a non-monadic version, see iterateN.

Since: 0.12.0.0

create :: (forall s. ST s (MVector s a)) -> Vector a

Execute the monadic action and freeze the resulting vector.

create (do { v <- new 2; write v 0 'a'; write v 1 'b'; return v }) = <a,b>

createT :: Traversable f => (forall s. ST s (f (MVector s a))) -> f (Vector a)

Execute the monadic action and freeze the resulting vectors.

Unfolding

unfoldr :: (b -> Maybe (a, b)) -> b -> Vector a

O(n) Construct a vector by repeatedly applying the generator function to a seed. The generator function yields Just the next element and the new seed or Nothing if there are no more elements.

unfoldr (\n -> if n == 0 then Nothing else Just (n,n-1)) 10
 = <10,9,8,7,6,5,4,3,2,1>

unfoldrN :: Int -> (b -> Maybe (a, b)) -> b -> Vector a

O(n) Construct a vector with at most n elements by repeatedly applying the generator function to a seed. The generator function yields Just the next element and the new seed or Nothing if there are no more elements.

unfoldrN 3 (\n -> Just (n,n-1)) 10 = <10,9,8>

unfoldrExactN :: Int -> (b -> (a, b)) -> b -> Vector a

O(n) Construct a vector with exactly n elements by repeatedly applying the generator function to a seed. The generator function yields the next element and the new seed.

unfoldrExactN 3 (\n -> (n,n-1)) 10 = <10,9,8>

Since: 0.12.2.0

unfoldrM :: Monad m => (b -> m (Maybe (a, b))) -> b -> m (Vector a)

O(n) Construct a vector by repeatedly applying the monadic generator function to a seed. The generator function yields Just the next element and the new seed or Nothing if there are no more elements.

unfoldrNM :: Monad m => Int -> (b -> m (Maybe (a, b))) -> b -> m (Vector a)

O(n) Construct a vector by repeatedly applying the monadic generator function to a seed. The generator function yields Just the next element and the new seed or Nothing if there are no more elements.

unfoldrExactNM :: Monad m => Int -> (b -> m (a, b)) -> b -> m (Vector a)

O(n) Construct a vector with exactly n elements by repeatedly applying the monadic generator function to a seed. The generator function yields the next element and the new seed.

Since: 0.12.2.0

constructN :: Int -> (Vector a -> a) -> Vector a

O(n) Construct a vector with n elements by repeatedly applying the generator function to the already constructed part of the vector.

constructN 3 f = let a = f <> ; b = f <a> ; c = f <a,b> in <a,b,c>

constructrN :: Int -> (Vector a -> a) -> Vector a

O(n) Construct a vector with n elements from right to left by repeatedly applying the generator function to the already constructed part of the vector.

constructrN 3 f = let a = f <> ; b = f<a> ; c = f <b,a> in <c,b,a>

Enumeration

enumFromN :: Num a => a -> Int -> Vector a

O(n) Yield a vector of the given length, containing the values x, x+1 etc. This operation is usually more efficient than enumFromTo.

enumFromN 5 3 = <5,6,7>

enumFromStepN :: Num a => a -> a -> Int -> Vector a

O(n) Yield a vector of the given length, containing the values x, x+y, x+y+y etc. This operations is usually more efficient than enumFromThenTo.

enumFromStepN 1 2 5 = <1,3,5,7,9>

enumFromTo :: Enum a => a -> a -> Vector a

O(n) Enumerate values from x to y.

WARNING: This operation can be very inefficient. If possible, use enumFromN instead.

enumFromThenTo :: Enum a => a -> a -> a -> Vector a

O(n) Enumerate values from x to y with a specific step z.

WARNING: This operation can be very inefficient. If possible, use enumFromStepN instead.

Concatenation

cons :: a -> Vector a -> Vector a

O(n) Prepend an element.

snoc :: Vector a -> a -> Vector a

O(n) Append an element.

(++) :: Vector a -> Vector a -> Vector a infixr 5

O(m+n) Concatenate two vectors.

concat :: [Vector a] -> Vector a

O(n) Concatenate all vectors in the list.

Restricting memory usage

force :: Vector a -> Vector a

O(n) Yield the argument, but force it not to retain any extra memory, possibly by copying it.

This is especially useful when dealing with slices. For example:

force (slice 0 2 <huge vector>)

Here, the slice retains a reference to the huge vector. Forcing it creates a copy of just the elements that belong to the slice and allows the huge vector to be garbage collected.

Modifying vectors

Bulk updates

(//)

Arguments

:: Vector a
initial vector (of length m)

-> [(Int, a)]
list of index/value pairs (of length n)

-> Vector a

O(m+n) For each pair (i,a) from the list of index/value pairs, replace the vector element at position i by a.

<5,9,2,7> // [(2,1),(0,3),(2,8)] = <3,9,8,7>

update

Arguments

:: Vector a
initial vector (of length m)

-> Vector (Int, a)
vector of index/value pairs (of length n)

-> Vector a

O(m+n) For each pair (i,a) from the vector of index/value pairs, replace the vector element at position i by a.

update <5,9,2,7> <(2,1),(0,3),(2,8)> = <3,9,8,7>

update_

Arguments

:: Vector a
initial vector (of length m)

-> Vector Int
index vector (of length n1)

-> Vector a
value vector (of length n2)

-> Vector a

O(m+min(n1,n2)) For each index i from the index vector and the corresponding value a from the value vector, replace the element of the initial vector at position i by a.

update_ <5,9,2,7>  <2,0,2> <1,3,8> = <3,9,8,7>

The function update provides the same functionality and is usually more convenient.

update_ xs is ys = update xs (zip is ys)

unsafeUpd :: Vector a -> [(Int, a)] -> Vector a

Same as (//), but without bounds checking.

unsafeUpdate :: Vector a -> Vector (Int, a) -> Vector a

Same as update, but without bounds checking.

unsafeUpdate_ :: Vector a -> Vector Int -> Vector a -> Vector a

Same as update_, but without bounds checking.

Accumulations

accum

Arguments

:: (a -> b -> a)
accumulating function f

-> Vector a
initial vector (of length m)

-> [(Int, b)]
list of index/value pairs (of length n)

-> Vector a

O(m+n) For each pair (i,b) from the list, replace the vector element a at position i by f a b.

Examples

Expand

>>> import qualified Data.Vector as V
>>> V.accum (+) (V.fromList [1000,2000,3000]) [(2,4),(1,6),(0,3),(1,10)]
[1003,2016,3004]

accumulate

Arguments

:: (a -> b -> a)
accumulating function f

-> Vector a
initial vector (of length m)

-> Vector (Int, b)
vector of index/value pairs (of length n)

-> Vector a

O(m+n) For each pair (i,b) from the vector of pairs, replace the vector element a at position i by f a b.

Examples

Expand

>>> import qualified Data.Vector as V
>>> V.accumulate (+) (V.fromList [1000,2000,3000]) (V.fromList [(2,4),(1,6),(0,3),(1,10)])
[1003,2016,3004]

accumulate_

Arguments

:: (a -> b -> a)
accumulating function f

-> Vector a
initial vector (of length m)

-> Vector Int
index vector (of length n1)

-> Vector b
value vector (of length n2)

-> Vector a

O(m+min(n1,n2)) For each index i from the index vector and the corresponding value b from the the value vector, replace the element of the initial vector at position i by f a b.

accumulate_ (+) <5,9,2> <2,1,0,1> <4,6,3,7> = <5+3, 9+6+7, 2+4>

The function accumulate provides the same functionality and is usually more convenient.

accumulate_ f as is bs = accumulate f as (zip is bs)

unsafeAccum :: (a -> b -> a) -> Vector a -> [(Int, b)] -> Vector a

Same as accum, but without bounds checking.

unsafeAccumulate :: (a -> b -> a) -> Vector a -> Vector (Int, b) -> Vector a

Same as accumulate, but without bounds checking.

unsafeAccumulate_ :: (a -> b -> a) -> Vector a -> Vector Int -> Vector b -> Vector a

Same as accumulate_, but without bounds checking.

Permutations

reverse :: Vector a -> Vector a

O(n) Reverse a vector.

backpermute :: Vector a -> Vector Int -> Vector a

O(n) Yield the vector obtained by replacing each element i of the index vector by xs!i. This is equivalent to map (xs!) is, but is often much more efficient.

backpermute <a,b,c,d> <0,3,2,3,1,0> = <a,d,c,d,b,a>

unsafeBackpermute :: Vector a -> Vector Int -> Vector a

Same as backpermute, but without bounds checking.

Safe destructive updates

modify :: (forall s. MVector s a -> ST s ()) -> Vector a -> Vector a

Apply a destructive operation to a vector. The operation may be performed in place if it is safe to do so and will modify a copy of the vector otherwise (see New for details).

Examples

Expand

>>> import qualified Data.Vector as V
>>> import qualified Data.Vector.Mutable as MV
>>> V.modify (\v -> MV.write v 0 'x') $ V.replicate 4 'a'
"xaaa"

Elementwise operations

Indexing

indexed :: Vector a -> Vector (Int, a)

O(n) Pair each element in a vector with its index.

Mapping

map :: (a -> b) -> Vector a -> Vector b

O(n) Map a function over a vector.

imap :: (Int -> a -> b) -> Vector a -> Vector b

O(n) Apply a function to every element of a vector and its index.

concatMap :: (a -> Vector b) -> Vector a -> Vector b

Map a function over a vector and concatenate the results.

Monadic mapping

mapM :: Monad m => (a -> m b) -> Vector a -> m (Vector b)

O(n) Apply the monadic action to all elements of the vector, yielding a vector of results.

imapM :: Monad m => (Int -> a -> m b) -> Vector a -> m (Vector b)

O(n) Apply the monadic action to every element of a vector and its index, yielding a vector of results.

mapM_ :: Monad m => (a -> m b) -> Vector a -> m ()

O(n) Apply the monadic action to all elements of a vector and ignore the results.

imapM_ :: Monad m => (Int -> a -> m b) -> Vector a -> m ()

O(n) Apply the monadic action to every element of a vector and its index, ignoring the results.

forM :: Monad m => Vector a -> (a -> m b) -> m (Vector b)

O(n) Apply the monadic action to all elements of the vector, yielding a vector of results. Equivalent to flip mapM.

forM_ :: Monad m => Vector a -> (a -> m b) -> m ()

O(n) Apply the monadic action to all elements of a vector and ignore the results. Equivalent to flip mapM_.

iforM :: Monad m => Vector a -> (Int -> a -> m b) -> m (Vector b)

O(n) Apply the monadic action to all elements of the vector and their indices, yielding a vector of results. Equivalent to flip imapM.

Since: 0.12.2.0

iforM_ :: Monad m => Vector a -> (Int -> a -> m b) -> m ()

O(n) Apply the monadic action to all elements of the vector and their indices and ignore the results. Equivalent to flip imapM_.

Since: 0.12.2.0

Zipping

zipWith :: (a -> b -> c) -> Vector a -> Vector b -> Vector c

O(min(m,n)) Zip two vectors with the given function.

zipWith3 :: (a -> b -> c -> d) -> Vector a -> Vector b -> Vector c -> Vector d

Zip three vectors with the given function.

zipWith4 :: (a -> b -> c -> d -> e) -> Vector a -> Vector b -> Vector c -> Vector d -> Vector e

zipWith5 :: (a -> b -> c -> d -> e -> f) -> Vector a -> Vector b -> Vector c -> Vector d -> Vector e -> Vector f

zipWith6 :: (a -> b -> c -> d -> e -> f -> g) -> Vector a -> Vector b -> Vector c -> Vector d -> Vector e -> Vector f -> Vector g

izipWith :: (Int -> a -> b -> c) -> Vector a -> Vector b -> Vector c

O(min(m,n)) Zip two vectors with a function that also takes the elements' indices.

izipWith3 :: (Int -> a -> b -> c -> d) -> Vector a -> Vector b -> Vector c -> Vector d

Zip three vectors and their indices with the given function.

izipWith4 :: (Int -> a -> b -> c -> d -> e) -> Vector a -> Vector b -> Vector c -> Vector d -> Vector e

izipWith5 :: (Int -> a -> b -> c -> d -> e -> f) -> Vector a -> Vector b -> Vector c -> Vector d -> Vector e -> Vector f

izipWith6 :: (Int -> a -> b -> c -> d -> e -> f -> g) -> Vector a -> Vector b -> Vector c -> Vector d -> Vector e -> Vector f -> Vector g

zip :: Vector a -> Vector b -> Vector (a, b)

O(min(m,n)) Zip two vectors.

zip3 :: Vector a -> Vector b -> Vector c -> Vector (a, b, c)

Zip together three vectors into a vector of triples.

zip4 :: Vector a -> Vector b -> Vector c -> Vector d -> Vector (a, b, c, d)

zip5 :: Vector a -> Vector b -> Vector c -> Vector d -> Vector e -> Vector (a, b, c, d, e)

zip6 :: Vector a -> Vector b -> Vector c -> Vector d -> Vector e -> Vector f -> Vector (a, b, c, d, e, f)

Monadic zipping

zipWithM :: Monad m => (a -> b -> m c) -> Vector a -> Vector b -> m (Vector c)

O(min(m,n)) Zip the two vectors with the monadic action and yield a vector of results.

izipWithM :: Monad m => (Int -> a -> b -> m c) -> Vector a -> Vector b -> m (Vector c)

O(min(m,n)) Zip the two vectors with a monadic action that also takes the element index and yield a vector of results.

zipWithM_ :: Monad m => (a -> b -> m c) -> Vector a -> Vector b -> m ()

O(min(m,n)) Zip the two vectors with the monadic action and ignore the results.

izipWithM_ :: Monad m => (Int -> a -> b -> m c) -> Vector a -> Vector b -> m ()

O(min(m,n)) Zip the two vectors with a monadic action that also takes the element index and ignore the results.

Unzipping

unzip :: Vector (a, b) -> (Vector a, Vector b)

O(min(m,n)) Unzip a vector of pairs.

unzip3 :: Vector (a, b, c) -> (Vector a, Vector b, Vector c)

unzip4 :: Vector (a, b, c, d) -> (Vector a, Vector b, Vector c, Vector d)

unzip5 :: Vector (a, b, c, d, e) -> (Vector a, Vector b, Vector c, Vector d, Vector e)

unzip6 :: Vector (a, b, c, d, e, f) -> (Vector a, Vector b, Vector c, Vector d, Vector e, Vector f)

Working with predicates

Filtering

filter :: (a -> Bool) -> Vector a -> Vector a

O(n) Drop all elements that do not satisfy the predicate.

ifilter :: (Int -> a -> Bool) -> Vector a -> Vector a

O(n) Drop all elements that do not satisfy the predicate which is applied to the values and their indices.

filterM :: Monad m => (a -> m Bool) -> Vector a -> m (Vector a)

O(n) Drop all elements that do not satisfy the monadic predicate.

uniq :: Eq a => Vector a -> Vector a

O(n) Drop repeated adjacent elements. The first element in each group is returned.

Examples

Expand

>>> import qualified Data.Vector as V
>>> V.uniq $ V.fromList [1,3,3,200,3]
[1,3,200,3]
>>> import Data.Semigroup
>>> V.uniq $ V.fromList [ Arg 1 'a', Arg 1 'b', Arg 1 'c']
[Arg 1 'a']

mapMaybe :: (a -> Maybe b) -> Vector a -> Vector b

O(n) Map the values and collect the Just results.

imapMaybe :: (Int -> a -> Maybe b) -> Vector a -> Vector b

O(n) Map the indices/values and collect the Just results.

mapMaybeM :: Monad m => (a -> m (Maybe b)) -> Vector a -> m (Vector b)

O(n) Apply the monadic function to each element of the vector and discard elements returning Nothing.

Since: 0.12.2.0

imapMaybeM :: Monad m => (Int -> a -> m (Maybe b)) -> Vector a -> m (Vector b)

O(n) Apply the monadic function to each element of the vector and its index. Discard elements returning Nothing.

Since: 0.12.2.0

catMaybes :: Vector (Maybe a) -> Vector a

O(n) Return a Vector of all the Just values.

Since: 0.12.2.0

takeWhile :: (a -> Bool) -> Vector a -> Vector a

O(n) Yield the longest prefix of elements satisfying the predicate. The current implementation is not copy-free, unless the result vector is fused away.

dropWhile :: (a -> Bool) -> Vector a -> Vector a

O(n) Drop the longest prefix of elements that satisfy the predicate without copying.

Partitioning

partition :: (a -> Bool) -> Vector a -> (Vector a, Vector a)

O(n) Split the vector in two parts, the first one containing those elements that satisfy the predicate and the second one those that don't. The relative order of the elements is preserved at the cost of a sometimes reduced performance compared to unstablePartition.

unstablePartition :: (a -> Bool) -> Vector a -> (Vector a, Vector a)

O(n) Split the vector in two parts, the first one containing those elements that satisfy the predicate and the second one those that don't. The order of the elements is not preserved, but the operation is often faster than partition.

partitionWith :: (a -> Either b c) -> Vector a -> (Vector b, Vector c)

O(n) Split the vector into two parts, the first one containing the Left elements and the second containing the Right elements. The relative order of the elements is preserved.

Since: 0.12.1.0

span :: (a -> Bool) -> Vector a -> (Vector a, Vector a)

O(n) Split the vector into the longest prefix of elements that satisfy the predicate and the rest without copying.

break :: (a -> Bool) -> Vector a -> (Vector a, Vector a)

O(n) Split the vector into the longest prefix of elements that do not satisfy the predicate and the rest without copying.

groupBy :: (a -> a -> Bool) -> Vector a -> [Vector a]

O(n) Split a vector into a list of slices, using a predicate function.

The concatenation of this list of slices is equal to the argument vector, and each slice contains only equal elements, as determined by the equality predicate function.

Does not fuse.

>>> import qualified Data.Vector as V
>>> import           Data.Char (isUpper)
>>> V.groupBy (\a b -> isUpper a == isUpper b) (V.fromList "Mississippi River")
["M","ississippi ","R","iver"]

See also groupBy, group.

Since: 0.13.0.1

group :: Eq a => Vector a -> [Vector a]

O(n) Split a vector into a list of slices of the input vector.

The concatenation of this list of slices is equal to the argument vector, and each slice contains only equal elements.

Does not fuse.

This is the equivalent of 'groupBy (==)'.

>>> import qualified Data.Vector as V
>>> V.group (V.fromList "Mississippi")
["M","i","ss","i","ss","i","pp","i"]

See also group.

Since: 0.13.0.1

Searching

elem :: Eq a => a -> Vector a -> Bool infix 4

O(n) Check if the vector contains an element.

notElem :: Eq a => a -> Vector a -> Bool infix 4

O(n) Check if the vector does not contain an element (inverse of elem).

find :: (a -> Bool) -> Vector a -> Maybe a

O(n) Yield Just the first element matching the predicate or Nothing if no such element exists.

findIndex :: (a -> Bool) -> Vector a -> Maybe Int

O(n) Yield Just the index of the first element matching the predicate or Nothing if no such element exists.

findIndexR :: (a -> Bool) -> Vector a -> Maybe Int

O(n) Yield Just the index of the last element matching the predicate or Nothing if no such element exists.

Does not fuse.

findIndices :: (a -> Bool) -> Vector a -> Vector Int

O(n) Yield the indices of elements satisfying the predicate in ascending order.

elemIndex :: Eq a => a -> Vector a -> Maybe Int

O(n) Yield Just the index of the first occurrence of the given element or Nothing if the vector does not contain the element. This is a specialised version of findIndex.

elemIndices :: Eq a => a -> Vector a -> Vector Int

O(n) Yield the indices of all occurrences of the given element in ascending order. This is a specialised version of findIndices.

Folding

foldl :: (a -> b -> a) -> a -> Vector b -> a

O(n) Left fold.

foldl1 :: (a -> a -> a) -> Vector a -> a

O(n) Left fold on non-empty vectors.

foldl' :: (a -> b -> a) -> a -> Vector b -> a

O(n) Left fold with strict accumulator.

foldl1' :: (a -> a -> a) -> Vector a -> a

O(n) Left fold on non-empty vectors with strict accumulator.

foldr :: (a -> b -> b) -> b -> Vector a -> b

O(n) Right fold.

foldr1 :: (a -> a -> a) -> Vector a -> a

O(n) Right fold on non-empty vectors.

foldr' :: (a -> b -> b) -> b -> Vector a -> b

O(n) Right fold with a strict accumulator.

foldr1' :: (a -> a -> a) -> Vector a -> a

O(n) Right fold on non-empty vectors with strict accumulator.

ifoldl :: (a -> Int -> b -> a) -> a -> Vector b -> a

O(n) Left fold using a function applied to each element and its index.

ifoldl' :: (a -> Int -> b -> a) -> a -> Vector b -> a

O(n) Left fold with strict accumulator using a function applied to each element and its index.

ifoldr :: (Int -> a -> b -> b) -> b -> Vector a -> b

O(n) Right fold using a function applied to each element and its index.

ifoldr' :: (Int -> a -> b -> b) -> b -> Vector a -> b

O(n) Right fold with strict accumulator using a function applied to each element and its index.

foldMap :: Monoid m => (a -> m) -> Vector a -> m

O(n) Map each element of the structure to a monoid and combine the results. It uses the same implementation as the corresponding method of the Foldable type class. Note that it's implemented in terms of foldr and won't fuse with functions that traverse the vector from left to right (map, generate, etc.).

Since: 0.12.2.0

foldMap' :: Monoid m => (a -> m) -> Vector a -> m

O(n) Like foldMap, but strict in the accumulator. It uses the same implementation as the corresponding method of the Foldable type class. Note that it's implemented in terms of foldl', so it fuses in most contexts.

Since: 0.12.2.0

Specialised folds

all :: (a -> Bool) -> Vector a -> Bool

O(n) Check if all elements satisfy the predicate.

Examples

Expand

>>> import qualified Data.Vector as V
>>> V.all even $ V.fromList [2, 4, 12]
True
>>> V.all even $ V.fromList [2, 4, 13]
False
>>> V.all even (V.empty :: V.Vector Int)
True

any :: (a -> Bool) -> Vector a -> Bool

O(n) Check if any element satisfies the predicate.

Examples

Expand

>>> import qualified Data.Vector as V
>>> V.any even $ V.fromList [1, 3, 7]
False
>>> V.any even $ V.fromList [3, 2, 13]
True
>>> V.any even (V.empty :: V.Vector Int)
False

and :: Vector Bool -> Bool

O(n) Check if all elements are True.

Examples

Expand

>>> import qualified Data.Vector as V
>>> V.and $ V.fromList [True, False]
False
>>> V.and V.empty
True

or :: Vector Bool -> Bool

O(n) Check if any element is True.

Examples

Expand

>>> import qualified Data.Vector as V
>>> V.or $ V.fromList [True, False]
True
>>> V.or V.empty
False

sum :: Num a => Vector a -> a

O(n) Compute the sum of the elements.

Examples

Expand

>>> import qualified Data.Vector as V
>>> V.sum $ V.fromList [300,20,1]
321
>>> V.sum (V.empty :: V.Vector Int)
0

product :: Num a => Vector a -> a

O(n) Compute the product of the elements.

Examples

Expand

>>> import qualified Data.Vector as V
>>> V.product $ V.fromList [1,2,3,4]
24
>>> V.product (V.empty :: V.Vector Int)
1

maximum :: Ord a => Vector a -> a

O(n) Yield the maximum element of the vector. The vector may not be empty. In case of a tie, the first occurrence wins.

Examples

Expand

>>> import qualified Data.Vector as V
>>> V.maximum $ V.fromList [2, 1]
2
>>> import Data.Semigroup
>>> V.maximum $ V.fromList [Arg 1 'a', Arg 2 'b']
Arg 2 'b'
>>> V.maximum $ V.fromList [Arg 1 'a', Arg 1 'b']
Arg 1 'a'

maximumBy :: (a -> a -> Ordering) -> Vector a -> a

O(n) Yield the maximum element of the vector according to the given comparison function. The vector may not be empty. In case of a tie, the first occurrence wins. This behavior is different from maximumBy which returns the last tie.

Examples

Expand

>>> import Data.Ord
>>> import qualified Data.Vector as V
>>> V.maximumBy (comparing fst) $ V.fromList [(2,'a'), (1,'b')]
(2,'a')
>>> V.maximumBy (comparing fst) $ V.fromList [(1,'a'), (1,'b')]
(1,'a')

maximumOn :: Ord b => (a -> b) -> Vector a -> a

O(n) Yield the maximum element of the vector by comparing the results of a key function on each element. In case of a tie, the first occurrence wins. The vector may not be empty.

Examples

Expand

>>> import qualified Data.Vector as V
>>> V.maximumOn fst $ V.fromList [(2,'a'), (1,'b')]
(2,'a')
>>> V.maximumOn fst $ V.fromList [(1,'a'), (1,'b')]
(1,'a')

Since: 0.13.0.0

minimum :: Ord a => Vector a -> a

O(n) Yield the minimum element of the vector. The vector may not be empty. In case of a tie, the first occurrence wins.

Examples

Expand

>>> import qualified Data.Vector as V
>>> V.minimum $ V.fromList [2, 1]
1
>>> import Data.Semigroup
>>> V.minimum $ V.fromList [Arg 2 'a', Arg 1 'b']
Arg 1 'b'
>>> V.minimum $ V.fromList [Arg 1 'a', Arg 1 'b']
Arg 1 'a'

minimumBy :: (a -> a -> Ordering) -> Vector a -> a

O(n) Yield the minimum element of the vector according to the given comparison function. The vector may not be empty. In case of a tie, the first occurrence wins.

Examples

Expand

>>> import Data.Ord
>>> import qualified Data.Vector as V
>>> V.minimumBy (comparing fst) $ V.fromList [(2,'a'), (1,'b')]
(1,'b')
>>> V.minimumBy (comparing fst) $ V.fromList [(1,'a'), (1,'b')]
(1,'a')

minimumOn :: Ord b => (a -> b) -> Vector a -> a

O(n) Yield the minimum element of the vector by comparing the results of a key function on each element. In case of a tie, the first occurrence wins. The vector may not be empty.

Examples

Expand

>>> import qualified Data.Vector as V
>>> V.minimumOn fst $ V.fromList [(2,'a'), (1,'b')]
(1,'b')
>>> V.minimumOn fst $ V.fromList [(1,'a'), (1,'b')]
(1,'a')

Since: 0.13.0.0

minIndex :: Ord a => Vector a -> Int

O(n) Yield the index of the minimum element of the vector. The vector may not be empty.

minIndexBy :: (a -> a -> Ordering) -> Vector a -> Int

O(n) Yield the index of the minimum element of the vector according to the given comparison function. The vector may not be empty.

Examples

Expand

>>> import Data.Ord
>>> import qualified Data.Vector as V
>>> V.minIndexBy (comparing fst) $ V.fromList [(2,'a'), (1,'b')]
1
>>> V.minIndexBy (comparing fst) $ V.fromList [(1,'a'), (1,'b')]
0

maxIndex :: Ord a => Vector a -> Int

O(n) Yield the index of the maximum element of the vector. The vector may not be empty.

maxIndexBy :: (a -> a -> Ordering) -> Vector a -> Int

O(n) Yield the index of the maximum element of the vector according to the given comparison function. The vector may not be empty. In case of a tie, the first occurrence wins.

Examples

Expand

>>> import Data.Ord
>>> import qualified Data.Vector as V
>>> V.maxIndexBy (comparing fst) $ V.fromList [(2,'a'), (1,'b')]
0
>>> V.maxIndexBy (comparing fst) $ V.fromList [(1,'a'), (1,'b')]
0

Monadic folds

foldM :: Monad m => (a -> b -> m a) -> a -> Vector b -> m a

O(n) Monadic fold.

ifoldM :: Monad m => (a -> Int -> b -> m a) -> a -> Vector b -> m a

O(n) Monadic fold using a function applied to each element and its index.

foldM' :: Monad m => (a -> b -> m a) -> a -> Vector b -> m a

O(n) Monadic fold with strict accumulator.

ifoldM' :: Monad m => (a -> Int -> b -> m a) -> a -> Vector b -> m a

O(n) Monadic fold with strict accumulator using a function applied to each element and its index.

fold1M :: Monad m => (a -> a -> m a) -> Vector a -> m a

O(n) Monadic fold over non-empty vectors.

fold1M' :: Monad m => (a -> a -> m a) -> Vector a -> m a

O(n) Monadic fold over non-empty vectors with strict accumulator.

foldM_ :: Monad m => (a -> b -> m a) -> a -> Vector b -> m ()

O(n) Monadic fold that discards the result.

ifoldM_ :: Monad m => (a -> Int -> b -> m a) -> a -> Vector b -> m ()

O(n) Monadic fold that discards the result using a function applied to each element and its index.

foldM'_ :: Monad m => (a -> b -> m a) -> a -> Vector b -> m ()

O(n) Monadic fold with strict accumulator that discards the result.

ifoldM'_ :: Monad m => (a -> Int -> b -> m a) -> a -> Vector b -> m ()

O(n) Monadic fold with strict accumulator that discards the result using a function applied to each element and its index.

fold1M_ :: Monad m => (a -> a -> m a) -> Vector a -> m ()

O(n) Monadic fold over non-empty vectors that discards the result.

fold1M'_ :: Monad m => (a -> a -> m a) -> Vector a -> m ()

O(n) Monadic fold over non-empty vectors with strict accumulator that discards the result.

Monadic sequencing

sequence :: Monad m => Vector (m a) -> m (Vector a)

Evaluate each action and collect the results.

sequence_ :: Monad m => Vector (m a) -> m ()

Evaluate each action and discard the results.

Scans

prescanl :: (a -> b -> a) -> a -> Vector b -> Vector a

O(n) Left-to-right prescan.

prescanl f z = init . scanl f z

Examples

Expand

>>> import qualified Data.Vector as V
>>> V.prescanl (+) 0 (V.fromList [1,2,3,4])
[0,1,3,6]

prescanl' :: (a -> b -> a) -> a -> Vector b -> Vector a

O(n) Left-to-right prescan with strict accumulator.

postscanl :: (a -> b -> a) -> a -> Vector b -> Vector a

O(n) Left-to-right postscan.

postscanl f z = tail . scanl f z

Examples

Expand

>>> import qualified Data.Vector as V
>>> V.postscanl (+) 0 (V.fromList [1,2,3,4])
[1,3,6,10]

postscanl' :: (a -> b -> a) -> a -> Vector b -> Vector a

O(n) Left-to-right postscan with strict accumulator.

scanl :: (a -> b -> a) -> a -> Vector b -> Vector a

O(n) Left-to-right scan.

scanl f z <x1,...,xn> = <y1,...,y(n+1)>
  where y1 = z
        yi = f y(i-1) x(i-1)

Examples

Expand

>>> import qualified Data.Vector as V
>>> V.scanl (+) 0 (V.fromList [1,2,3,4])
[0,1,3,6,10]

scanl' :: (a -> b -> a) -> a -> Vector b -> Vector a

O(n) Left-to-right scan with strict accumulator.

scanl1 :: (a -> a -> a) -> Vector a -> Vector a

O(n) Initial-value free left-to-right scan over a vector.

scanl f <x1,...,xn> = <y1,...,yn>
  where y1 = x1
        yi = f y(i-1) xi

Note: Since 0.13, application of this to an empty vector no longer results in an error; instead it produces an empty vector.

Examples

Expand

>>> import qualified Data.Vector as V
>>> V.scanl1 min $ V.fromListN 5 [4,2,4,1,3]
[4,2,2,1,1]
>>> V.scanl1 max $ V.fromListN 5 [1,3,2,5,4]
[1,3,3,5,5]
>>> V.scanl1 min (V.empty :: V.Vector Int)
[]

scanl1' :: (a -> a -> a) -> Vector a -> Vector a

O(n) Initial-value free left-to-right scan over a vector with a strict accumulator.

Note: Since 0.13, application of this to an empty vector no longer results in an error; instead it produces an empty vector.

Examples

Expand

>>> import qualified Data.Vector as V
>>> V.scanl1' min $ V.fromListN 5 [4,2,4,1,3]
[4,2,2,1,1]
>>> V.scanl1' max $ V.fromListN 5 [1,3,2,5,4]
[1,3,3,5,5]
>>> V.scanl1' min (V.empty :: V.Vector Int)
[]

iscanl :: (Int -> a -> b -> a) -> a -> Vector b -> Vector a

O(n) Left-to-right scan over a vector with its index.

Since: 0.12.0.0

iscanl' :: (Int -> a -> b -> a) -> a -> Vector b -> Vector a

O(n) Left-to-right scan over a vector (strictly) with its index.

Since: 0.12.0.0

prescanr :: (a -> b -> b) -> b -> Vector a -> Vector b

O(n) Right-to-left prescan.

prescanr f z = reverse . prescanl (flip f) z . reverse

prescanr' :: (a -> b -> b) -> b -> Vector a -> Vector b

O(n) Right-to-left prescan with strict accumulator.

postscanr :: (a -> b -> b) -> b -> Vector a -> Vector b

O(n) Right-to-left postscan.

postscanr' :: (a -> b -> b) -> b -> Vector a -> Vector b

O(n) Right-to-left postscan with strict accumulator.

scanr :: (a -> b -> b) -> b -> Vector a -> Vector b

O(n) Right-to-left scan.

scanr' :: (a -> b -> b) -> b -> Vector a -> Vector b

O(n) Right-to-left scan with strict accumulator.

scanr1 :: (a -> a -> a) -> Vector a -> Vector a

O(n) Right-to-left, initial-value free scan over a vector.

Note: Since 0.13, application of this to an empty vector no longer results in an error; instead it produces an empty vector.

Examples

Expand

>>> import qualified Data.Vector as V
>>> V.scanr1 min $ V.fromListN 5 [3,1,4,2,4]
[1,1,2,2,4]
>>> V.scanr1 max $ V.fromListN 5 [4,5,2,3,1]
[5,5,3,3,1]
>>> V.scanr1 min (V.empty :: V.Vector Int)
[]

scanr1' :: (a -> a -> a) -> Vector a -> Vector a

O(n) Right-to-left, initial-value free scan over a vector with a strict accumulator.

Note: Since 0.13, application of this to an empty vector no longer results in an error; instead it produces an empty vector.

Examples

Expand

>>> import qualified Data.Vector as V
>>> V.scanr1' min $ V.fromListN 5 [3,1,4,2,4]
[1,1,2,2,4]
>>> V.scanr1' max $ V.fromListN 5 [4,5,2,3,1]
[5,5,3,3,1]
>>> V.scanr1' min (V.empty :: V.Vector Int)
[]

iscanr :: (Int -> a -> b -> b) -> b -> Vector a -> Vector b

O(n) Right-to-left scan over a vector with its index.

Since: 0.12.0.0

iscanr' :: (Int -> a -> b -> b) -> b -> Vector a -> Vector b

O(n) Right-to-left scan over a vector (strictly) with its index.

Since: 0.12.0.0

Comparisons

eqBy :: (a -> b -> Bool) -> Vector a -> Vector b -> Bool

O(n) Check if two vectors are equal using the supplied equality predicate.

Since: 0.12.2.0

cmpBy :: (a -> b -> Ordering) -> Vector a -> Vector b -> Ordering

O(n) Compare two vectors using the supplied comparison function for vector elements. Comparison works the same as for lists.

cmpBy compare == compare

Since: 0.12.2.0

Conversions

Lists

toList :: Vector a -> [a]

O(n) Convert a vector to a list.

fromList :: [a] -> Vector a

O(n) Convert a list to a vector.

fromListN :: Int -> [a] -> Vector a

O(n) Convert the first n elements of a list to a vector. It's expected that the supplied list will be exactly n elements long. As an optimization, this function allocates a buffer for n elements, which could be used for DoS-attacks by exhausting the memory if an attacker controls that parameter.

fromListN n xs = fromList (take n xs)

Arrays

toArray :: Vector a -> Array a

O(n) Convert a vector to an array.

Since: 0.12.2.0

fromArray :: Array a -> Vector a

O(1) Convert an array to a vector.

Since: 0.12.2.0

toArraySlice :: Vector a -> (Array a, Int, Int)

O(1) Extract the underlying Array, offset where vector starts and the total number of elements in the vector. Below property always holds:

let (array, offset, len) = toArraySlice v
v === unsafeFromArraySlice len offset array

Since: 0.13.0.0

unsafeFromArraySlice

Arguments

:: Array a
Immutable boxed array.

-> Int
Offset

-> Int
Length

-> Vector a

O(1) Convert an array slice to a vector. This function is very unsafe, because constructing an invalid vector can yield almost all other safe functions in this module unsafe. These are equivalent:

unsafeFromArraySlice len offset === unsafeTake len . unsafeDrop offset . fromArray

Since: 0.13.0.0

Other vector types

convert :: (Vector v a, Vector w a) => v a -> w a

O(n) Convert between different vector types.

Mutable vectors

freeze :: PrimMonad m => MVector (PrimState m) a -> m (Vector a)

O(n) Yield an immutable copy of the mutable vector.

thaw :: PrimMonad m => Vector a -> m (MVector (PrimState m) a)

O(n) Yield a mutable copy of an immutable vector.

copy :: PrimMonad m => MVector (PrimState m) a -> Vector a -> m ()

O(n) Copy an immutable vector into a mutable one. The two vectors must have the same length.

unsafeFreeze :: PrimMonad m => MVector (PrimState m) a -> m (Vector a)

O(1) Unsafely convert a mutable vector to an immutable one without copying. The mutable vector may not be used after this operation.

unsafeThaw :: PrimMonad m => Vector a -> m (MVector (PrimState m) a)

O(1) Unsafely convert an immutable vector to a mutable one without copying. Note that this is a very dangerous function and generally it's only safe to read from the resulting vector. In this case, the immutable vector could be used safely as well.

Problems with mutation happen because GHC has a lot of freedom to introduce sharing. As a result mutable vectors produced by unsafeThaw may or may not share the same underlying buffer. For example:

foo = do
  let vec = V.generate 10 id
  mvec <- V.unsafeThaw vec
  do_something mvec

Here GHC could lift vec outside of foo which means that all calls to do_something will use same buffer with possibly disastrous results. Whether such aliasing happens or not depends on the program in question, optimization levels, and GHC flags.

All in all, attempts to modify a vector produced by unsafeThaw fall out of domain of software engineering and into realm of black magic, dark rituals, and unspeakable horrors. The only advice that could be given is: "Don't attempt to mutate a vector produced by unsafeThaw unless you know how to prevent GHC from aliasing buffers accidentally. We don't."

unsafeCopy :: PrimMonad m => MVector (PrimState m) a -> Vector a -> m ()

O(n) Copy an immutable vector into a mutable one. The two vectors must have the same length. This is not checked.