Data.Vector.Generic

Description

Generic interface to immutable vectors.

Immutable vectors

class MVector (Mutable v) a => Vector v a where

Class of immutable vectors. Every immutable vector is associated with its mutable version through the Mutable type family. Methods of this class should not be used directly. Instead, Data.Vector.Generic and other Data.Vector modules provide safe and fusible wrappers.

Minimum complete implementation:

basicUnsafeFreeze

basicUnsafeThaw

basicLength

basicUnsafeSlice

basicUnsafeIndexM

Minimal complete definition

basicUnsafeFreeze, basicUnsafeThaw, basicLength, basicUnsafeSlice, basicUnsafeIndexM

Methods

basicUnsafeFreeze :: Mutable v s a -> ST s (v a)

Assumed complexity: O(1)

Unsafely convert a mutable vector to its immutable version without copying. The mutable vector may not be used after this operation.

basicUnsafeThaw :: v a -> ST s (Mutable v s a)

Assumed complexity: O(1)

Unsafely convert an immutable vector to its mutable version without copying. The immutable vector may not be used after this operation.

basicLength :: v a -> Int

Assumed complexity: O(1)

Yield the length of the vector.

basicUnsafeSlice

Arguments

:: Int
starting index

-> Int
length

-> v a

-> v a

Assumed complexity: O(1)

Yield a slice of the vector without copying it. No range checks are performed.

basicUnsafeIndexM :: v a -> Int -> Box a

Assumed complexity: O(1)

Yield the element at the given position in a monad. No range checks are performed.

The monad allows us to be strict in the vector if we want. Suppose we had

unsafeIndex :: v a -> Int -> a

instead. Now, if we wanted to copy a vector, we'd do something like

copy mv v ... = ... unsafeWrite mv i (unsafeIndex v i) ...

For lazy vectors, the indexing would not be evaluated, which means that we would retain a reference to the original vector in each element we write. This is not what we want!

With basicUnsafeIndexM, we can do

copy mv v ... = ... case basicUnsafeIndexM v i of
                      Box x -> unsafeWrite mv i x ...

which does not have this problem, because indexing (but not the returned element!) is evaluated immediately.

basicUnsafeCopy :: Mutable v s a -> v a -> ST s ()

Assumed complexity: O(n)

Copy an immutable vector into a mutable one. The two vectors must have the same length, but this is not checked.

Instances of Vector should redefine this method if they wish to support an efficient block copy operation.

Default definition: copying based on basicUnsafeIndexM and basicUnsafeWrite.

elemseq :: v a -> a -> b -> b

Evaluate a as far as storing it in a vector would and yield b. The v a argument only fixes the type and is not touched. This method is only used for optimisation purposes. Thus, it is safe for instances of Vector to evaluate a less than it would be when stored in a vector, although this might result in suboptimal code.

elemseq v x y = (singleton x `asTypeOf` v) `seq` y

Default defintion: a is not evaluated at all.

type family Mutable (v :: Type -> Type) = (mv :: Type -> Type -> Type) | mv -> v

Mutable v s a is the mutable version of the immutable vector type v a with the state token s. It is injective on GHC 8 and newer.

Accessors

Length information

length :: Vector v a => v a -> Int

O(1) Yield the length of the vector.

null :: Vector v a => v a -> Bool

O(1) Test whether a vector is empty.

Indexing

(!) :: (HasCallStack, Vector v a) => v a -> Int -> a infixl 9

O(1) Indexing.

(!?) :: Vector v a => v a -> Int -> Maybe a infixl 9

O(1) Safe indexing.

head :: Vector v a => v a -> a

O(1) First element.

last :: Vector v a => v a -> a

O(1) Last element.

unsafeIndex :: Vector v a => v a -> Int -> a

O(1) Unsafe indexing without bounds checking.

unsafeHead :: Vector v a => v a -> a

O(1) First element, without checking if the vector is empty.

unsafeLast :: Vector v a => v a -> a

O(1) Last element, without checking if the vector is empty.

Monadic indexing

indexM :: (HasCallStack, Vector v a, Monad m) => v a -> Int -> m a

O(1) Indexing in a monad.

The monad allows operations to be strict in the vector when necessary. Suppose vector copying is implemented like this:

copy mv v = ... write mv i (v ! i) ...

For lazy vectors, v ! i would not be evaluated which means that mv would unnecessarily retain a reference to v in each element written.

With indexM, copying can be implemented like this instead:

copy mv v = ... do
                  x <- indexM v i
                  write mv i x

Here, no references to v are retained because indexing (but not the element) is evaluated eagerly.

headM :: (Vector v a, Monad m) => v a -> m a

O(1) First element of a vector in a monad. See indexM for an explanation of why this is useful.

lastM :: (Vector v a, Monad m) => v a -> m a

O(1) Last element of a vector in a monad. See indexM for an explanation of why this is useful.

unsafeIndexM :: (Vector v a, Monad m) => v a -> Int -> m a

O(1) Indexing in a monad, without bounds checks. See indexM for an explanation of why this is useful.

unsafeHeadM :: (Vector v a, Monad m) => v a -> m a

O(1) First element in a monad, without checking for empty vectors. See indexM for an explanation of why this is useful.

unsafeLastM :: (Vector v a, Monad m) => v a -> m a

O(1) Last element in a monad, without checking for empty vectors. See indexM for an explanation of why this is useful.

Extracting subvectors (slicing)

slice

Arguments

:: (HasCallStack, Vector v a)

=> Int
i starting index

-> Int
n length

-> v a

-> v a

O(1) Yield a slice of the vector without copying it. The vector must contain at least i+n elements.

init :: Vector v a => v a -> v a

O(1) Yield all but the last element without copying. The vector may not be empty.

tail :: Vector v a => v a -> v a

O(1) Yield all but the first element without copying. The vector may not be empty.

take :: Vector v a => Int -> v a -> v a

O(1) Yield the first n elements without copying. The vector may contain less than n elements, in which case it is returned unchanged.

drop :: Vector v a => Int -> v a -> v a

O(1) Yield all but the first n elements without copying. The vector may contain less than n elements, in which case an empty vector is returned.

splitAt :: Vector v a => Int -> v a -> (v a, v a)

O(1) Yield the first n elements paired with the remainder, without copying.

Note that splitAt n v is equivalent to (take n v, drop n v), but slightly more efficient.

Since: 0.7.1

uncons :: Vector v a => v a -> Maybe (a, v a)

O(1) Yield the head and tail of the vector, or Nothing if the vector is empty.

Since: 0.12.2.0

unsnoc :: Vector v a => v a -> Maybe (v a, a)

O(1) Yield the last and init of the vector, or Nothing if the vector is empty.

Since: 0.12.2.0

unsafeSlice

Arguments

:: Vector v a

=> Int
i starting index

-> Int
n length

-> v a

-> v a

O(1) Yield a slice of the vector without copying. The vector must contain at least i+n elements, but this is not checked.

unsafeInit :: Vector v a => v a -> v a

O(1) Yield all but the last element without copying. The vector may not be empty, but this is not checked.

unsafeTail :: Vector v a => v a -> v a

O(1) Yield all but the first element without copying. The vector may not be empty, but this is not checked.

unsafeTake :: Vector v a => Int -> v a -> v a

O(1) Yield the first n elements without copying. The vector must contain at least n elements, but this is not checked.

unsafeDrop :: Vector v a => Int -> v a -> v a

O(1) Yield all but the first n elements without copying. The vector must contain at least n elements, but this is not checked.

Construction

Initialisation

empty :: Vector v a => v a

O(1) The empty vector.

singleton :: forall v a. Vector v a => a -> v a

O(1) A vector with exactly one element.

replicate :: forall v a. Vector v a => Int -> a -> v a

O(n) A vector of the given length with the same value in each position.

generate :: Vector v a => Int -> (Int -> a) -> v a

O(n) Construct a vector of the given length by applying the function to each index.

iterateN :: Vector v a => Int -> (a -> a) -> a -> v a

O(n) Apply the function \(\max(n - 1, 0)\) times to an initial value, producing a vector of length \(\max(n, 0)\). The 0th element will contain the initial value, which is why there is one less function application than the number of elements in the produced vector.

\( \underbrace{x, f (x), f (f (x)), \ldots}_{\max(0,n)\rm{~elements}} \)

Since: 0.7.1

Monadic initialisation

replicateM :: (Monad m, Vector v a) => Int -> m a -> m (v a)

O(n) Execute the monadic action the given number of times and store the results in a vector.

generateM :: (Monad m, Vector v a) => Int -> (Int -> m a) -> m (v a)

O(n) Construct a vector of the given length by applying the monadic action to each index.

iterateNM :: (Monad m, Vector v a) => Int -> (a -> m a) -> a -> m (v a)

O(n) Apply the monadic function \(\max(n - 1, 0)\) times to an initial value, producing a vector of length \(\max(n, 0)\). The 0th element will contain the initial value, which is why there is one less function application than the number of elements in the produced vector.

For a non-monadic version, see iterateN.

Since: 0.12.0.0

create :: Vector v a => (forall s. ST s (Mutable v s a)) -> v a

Execute the monadic action and freeze the resulting vector.

create (do { v <- new 2; write v 0 'a'; write v 1 'b'; return v }) = <a,b>

createT :: (Traversable f, Vector v a) => (forall s. ST s (f (Mutable v s a))) -> f (v a)

Execute the monadic action and freeze the resulting vectors.

Unfolding

unfoldr :: Vector v a => (b -> Maybe (a, b)) -> b -> v a

O(n) Construct a vector by repeatedly applying the generator function to a seed. The generator function yields Just the next element and the new seed or Nothing if there are no more elements.

unfoldr (\n -> if n == 0 then Nothing else Just (n,n-1)) 10
 = <10,9,8,7,6,5,4,3,2,1>

unfoldrN :: Vector v a => Int -> (b -> Maybe (a, b)) -> b -> v a

O(n) Construct a vector with at most n elements by repeatedly applying the generator function to a seed. The generator function yields Just the next element and the new seed or Nothing if there are no more elements.

unfoldrN 3 (\n -> Just (n,n-1)) 10 = <10,9,8>

unfoldrExactN :: Vector v a => Int -> (b -> (a, b)) -> b -> v a

O(n) Construct a vector with exactly n elements by repeatedly applying the generator function to a seed. The generator function yields the next element and the new seed.

unfoldrExactN 3 (\n -> (n,n-1)) 10 = <10,9,8>

Since: 0.12.2.0

unfoldrM :: (Monad m, Vector v a) => (b -> m (Maybe (a, b))) -> b -> m (v a)

O(n) Construct a vector by repeatedly applying the monadic generator function to a seed. The generator function yields Just the next element and the new seed or Nothing if there are no more elements.

unfoldrNM :: (Monad m, Vector v a) => Int -> (b -> m (Maybe (a, b))) -> b -> m (v a)

O(n) Construct a vector by repeatedly applying the monadic generator function to a seed. The generator function yields Just the next element and the new seed or Nothing if there are no more elements.

unfoldrExactNM :: (Monad m, Vector v a) => Int -> (b -> m (a, b)) -> b -> m (v a)

O(n) Construct a vector with exactly n elements by repeatedly applying the monadic generator function to a seed. The generator function yields the next element and the new seed.

Since: 0.12.2.0

constructN :: forall v a. Vector v a => Int -> (v a -> a) -> v a

O(n) Construct a vector with n elements by repeatedly applying the generator function to the already constructed part of the vector.

constructN 3 f = let a = f <> ; b = f <a> ; c = f <a,b> in <a,b,c>

constructrN :: forall v a. Vector v a => Int -> (v a -> a) -> v a

O(n) Construct a vector with n elements from right to left by repeatedly applying the generator function to the already constructed part of the vector.

constructrN 3 f = let a = f <> ; b = f<a> ; c = f <b,a> in <c,b,a>

Enumeration

enumFromN :: (Vector v a, Num a) => a -> Int -> v a

O(n) Yield a vector of the given length, containing the values x, x+1 etc. This operation is usually more efficient than enumFromTo.

enumFromN 5 3 = <5,6,7>

enumFromStepN :: forall v a. (Vector v a, Num a) => a -> a -> Int -> v a

O(n) Yield a vector of the given length, containing the values x, x+y, x+y+y etc. This operations is usually more efficient than enumFromThenTo.

enumFromStepN 1 2 5 = <1,3,5,7,9>

enumFromTo :: (Vector v a, Enum a) => a -> a -> v a

O(n) Enumerate values from x to y.

WARNING: This operation can be very inefficient. If possible, use enumFromN instead.

enumFromThenTo :: (Vector v a, Enum a) => a -> a -> a -> v a

O(n) Enumerate values from x to y with a specific step z.

WARNING: This operation can be very inefficient. If possible, use enumFromStepN instead.

Concatenation

cons :: forall v a. Vector v a => a -> v a -> v a

O(n) Prepend an element.

snoc :: forall v a. Vector v a => v a -> a -> v a

O(n) Append an element.

(++) :: Vector v a => v a -> v a -> v a infixr 5

O(m+n) Concatenate two vectors.

concat :: Vector v a => [v a] -> v a

O(n) Concatenate all vectors in the list.

concatNE :: Vector v a => NonEmpty (v a) -> v a

O(n) Concatenate all vectors in the non-empty list.

Restricting memory usage

force :: Vector v a => v a -> v a

O(n) Yield the argument, but force it not to retain any extra memory, possibly by copying it.

This is especially useful when dealing with slices. For example:

force (slice 0 2 <huge vector>)

Here, the slice retains a reference to the huge vector. Forcing it creates a copy of just the elements that belong to the slice and allows the huge vector to be garbage collected.

Modifying vectors

Bulk updates

(//)

Arguments

:: Vector v a

=> v a
initial vector (of length m)

-> [(Int, a)]
list of index/value pairs (of length n)

-> v a

O(m+n) For each pair (i,a) from the list of index/value pairs, replace the vector element at position i by a.

<5,9,2,7> // [(2,1),(0,3),(2,8)] = <3,9,8,7>

update

Arguments

:: (Vector v a, Vector v (Int, a))

=> v a
initial vector (of length m)

-> v (Int, a)
vector of index/value pairs (of length n)

-> v a

O(m+n) For each pair (i,a) from the vector of index/value pairs, replace the vector element at position i by a.

update <5,9,2,7> <(2,1),(0,3),(2,8)> = <3,9,8,7>

update_

Arguments

:: (Vector v a, Vector v Int)

=> v a
initial vector (of length m)

-> v Int
index vector (of length n1)

-> v a
value vector (of length n2)

-> v a

O(m+min(n1,n2)) For each index i from the index vector and the corresponding value a from the value vector, replace the element of the initial vector at position i by a.

update_ <5,9,2,7>  <2,0,2> <1,3,8> = <3,9,8,7>

This function is useful for instances of Vector that cannot store pairs. Otherwise, update is probably more convenient.

update_ xs is ys = update xs (zip is ys)

unsafeUpd :: Vector v a => v a -> [(Int, a)] -> v a

Same as (//), but without bounds checking.

unsafeUpdate :: (Vector v a, Vector v (Int, a)) => v a -> v (Int, a) -> v a

Same as update, but without bounds checking.

unsafeUpdate_ :: (Vector v a, Vector v Int) => v a -> v Int -> v a -> v a

Same as update_, but without bounds checking.

Accumulations

accum

Arguments

:: Vector v a

=> (a -> b -> a)
accumulating function f

-> v a
initial vector (of length m)

-> [(Int, b)]
list of index/value pairs (of length n)

-> v a

O(m+n) For each pair (i,b) from the list, replace the vector element a at position i by f a b.

Examples

Expand

>>> import qualified Data.Vector as V
>>> V.accum (+) (V.fromList [1000,2000,3000]) [(2,4),(1,6),(0,3),(1,10)]
[1003,2016,3004]

accumulate

Arguments

:: (Vector v a, Vector v (Int, b))

=> (a -> b -> a)
accumulating function f

-> v a
initial vector (of length m)

-> v (Int, b)
vector of index/value pairs (of length n)

-> v a

O(m+n) For each pair (i,b) from the vector of pairs, replace the vector element a at position i by f a b.

Examples

Expand

>>> import qualified Data.Vector as V
>>> V.accumulate (+) (V.fromList [1000,2000,3000]) (V.fromList [(2,4),(1,6),(0,3),(1,10)])
[1003,2016,3004]

accumulate_

Arguments

:: (Vector v a, Vector v Int, Vector v b)

=> (a -> b -> a)
accumulating function f

-> v a
initial vector (of length m)

-> v Int
index vector (of length n1)

-> v b
value vector (of length n2)

-> v a

O(m+min(n1,n2)) For each index i from the index vector and the corresponding value b from the the value vector, replace the element of the initial vector at position i by f a b.

accumulate_ (+) <5,9,2> <2,1,0,1> <4,6,3,7> = <5+3, 9+6+7, 2+4>

This function is useful for instances of Vector that cannot store pairs. Otherwise, accumulate is probably more convenient:

accumulate_ f as is bs = accumulate f as (zip is bs)

unsafeAccum :: Vector v a => (a -> b -> a) -> v a -> [(Int, b)] -> v a

Same as accum, but without bounds checking.

unsafeAccumulate :: (Vector v a, Vector v (Int, b)) => (a -> b -> a) -> v a -> v (Int, b) -> v a

Same as accumulate, but without bounds checking.

unsafeAccumulate_ :: (Vector v a, Vector v Int, Vector v b) => (a -> b -> a) -> v a -> v Int -> v b -> v a

Same as accumulate_, but without bounds checking.

Permutations

reverse :: Vector v a => v a -> v a

O(n) Reverse a vector.

backpermute

Arguments

:: forall v a. (HasCallStack, Vector v a, Vector v Int)

=> v a
xs value vector

-> v Int
is index vector (of length n)

-> v a

O(n) Yield the vector obtained by replacing each element i of the index vector by xs!i. This is equivalent to map (xs!) is, but is often much more efficient.

backpermute <a,b,c,d> <0,3,2,3,1,0> = <a,d,c,d,b,a>

unsafeBackpermute :: (Vector v a, Vector v Int) => v a -> v Int -> v a

Same as backpermute, but without bounds checking.

Safe destructive updates

modify :: Vector v a => (forall s. Mutable v s a -> ST s ()) -> v a -> v a

Apply a destructive operation to a vector. The operation may be performed in place if it is safe to do so and will modify a copy of the vector otherwise (see New for details).

Examples

Expand

>>> import qualified Data.Vector as V
>>> import qualified Data.Vector.Mutable as MV
>>> V.modify (\v -> MV.write v 0 'x') $ V.replicate 4 'a'
"xaaa"

Elementwise operations

Indexing

indexed :: (Vector v a, Vector v (Int, a)) => v a -> v (Int, a)

O(n) Pair each element in a vector with its index.

Mapping

map :: (Vector v a, Vector v b) => (a -> b) -> v a -> v b

O(n) Map a function over a vector.

imap :: (Vector v a, Vector v b) => (Int -> a -> b) -> v a -> v b

O(n) Apply a function to every element of a vector and its index.

concatMap :: (Vector v a, Vector v b) => (a -> v b) -> v a -> v b

Map a function over a vector and concatenate the results.

Monadic mapping

mapM :: (Monad m, Vector v a, Vector v b) => (a -> m b) -> v a -> m (v b)

O(n) Apply the monadic action to all elements of the vector, yielding a vector of results.

imapM :: (Monad m, Vector v a, Vector v b) => (Int -> a -> m b) -> v a -> m (v b)

O(n) Apply the monadic action to every element of a vector and its index, yielding a vector of results.

mapM_ :: (Monad m, Vector v a) => (a -> m b) -> v a -> m ()

O(n) Apply the monadic action to all elements of a vector and ignore the results.

imapM_ :: (Monad m, Vector v a) => (Int -> a -> m b) -> v a -> m ()

O(n) Apply the monadic action to every element of a vector and its index, ignoring the results.

forM :: (Monad m, Vector v a, Vector v b) => v a -> (a -> m b) -> m (v b)

O(n) Apply the monadic action to all elements of the vector, yielding a vector of results. Equivalent to flip mapM.

forM_ :: (Monad m, Vector v a) => v a -> (a -> m b) -> m ()

O(n) Apply the monadic action to all elements of a vector and ignore the results. Equivalent to flip mapM_.

iforM :: (Monad m, Vector v a, Vector v b) => v a -> (Int -> a -> m b) -> m (v b)

O(n) Apply the monadic action to all elements of the vector and their indices, yielding a vector of results. Equivalent to flip imapM.

Since: 0.12.2.0

iforM_ :: (Monad m, Vector v a) => v a -> (Int -> a -> m b) -> m ()

O(n) Apply the monadic action to all elements of the vector and their indices and ignore the results. Equivalent to flip imapM_.

Since: 0.12.2.0

Zipping

zipWith :: (Vector v a, Vector v b, Vector v c) => (a -> b -> c) -> v a -> v b -> v c

O(min(m,n)) Zip two vectors with the given function.

zipWith3 :: (Vector v a, Vector v b, Vector v c, Vector v d) => (a -> b -> c -> d) -> v a -> v b -> v c -> v d

Zip three vectors with the given function.

zipWith4 :: (Vector v a, Vector v b, Vector v c, Vector v d, Vector v e) => (a -> b -> c -> d -> e) -> v a -> v b -> v c -> v d -> v e

zipWith5 :: (Vector v a, Vector v b, Vector v c, Vector v d, Vector v e, Vector v f) => (a -> b -> c -> d -> e -> f) -> v a -> v b -> v c -> v d -> v e -> v f

zipWith6 :: (Vector v a, Vector v b, Vector v c, Vector v d, Vector v e, Vector v f, Vector v g) => (a -> b -> c -> d -> e -> f -> g) -> v a -> v b -> v c -> v d -> v e -> v f -> v g

izipWith :: (Vector v a, Vector v b, Vector v c) => (Int -> a -> b -> c) -> v a -> v b -> v c

O(min(m,n)) Zip two vectors with a function that also takes the elements' indices.

izipWith3 :: (Vector v a, Vector v b, Vector v c, Vector v d) => (Int -> a -> b -> c -> d) -> v a -> v b -> v c -> v d

Zip three vectors and their indices with the given function.

izipWith4 :: (Vector v a, Vector v b, Vector v c, Vector v d, Vector v e) => (Int -> a -> b -> c -> d -> e) -> v a -> v b -> v c -> v d -> v e

izipWith5 :: (Vector v a, Vector v b, Vector v c, Vector v d, Vector v e, Vector v f) => (Int -> a -> b -> c -> d -> e -> f) -> v a -> v b -> v c -> v d -> v e -> v f

izipWith6 :: (Vector v a, Vector v b, Vector v c, Vector v d, Vector v e, Vector v f, Vector v g) => (Int -> a -> b -> c -> d -> e -> f -> g) -> v a -> v b -> v c -> v d -> v e -> v f -> v g

zip :: (Vector v a, Vector v b, Vector v (a, b)) => v a -> v b -> v (a, b)

O(min(m,n)) Zip two vectors.

zip3 :: (Vector v a, Vector v b, Vector v c, Vector v (a, b, c)) => v a -> v b -> v c -> v (a, b, c)

Zip together three vectors into a vector of triples.

zip4 :: (Vector v a, Vector v b, Vector v c, Vector v d, Vector v (a, b, c, d)) => v a -> v b -> v c -> v d -> v (a, b, c, d)

zip5 :: (Vector v a, Vector v b, Vector v c, Vector v d, Vector v e, Vector v (a, b, c, d, e)) => v a -> v b -> v c -> v d -> v e -> v (a, b, c, d, e)

zip6 :: (Vector v a, Vector v b, Vector v c, Vector v d, Vector v e, Vector v f, Vector v (a, b, c, d, e, f)) => v a -> v b -> v c -> v d -> v e -> v f -> v (a, b, c, d, e, f)

Monadic zipping

zipWithM :: (Monad m, Vector v a, Vector v b, Vector v c) => (a -> b -> m c) -> v a -> v b -> m (v c)

O(min(m,n)) Zip the two vectors with the monadic action and yield a vector of results.

izipWithM :: (Monad m, Vector v a, Vector v b, Vector v c) => (Int -> a -> b -> m c) -> v a -> v b -> m (v c)

O(min(m,n)) Zip the two vectors with a monadic action that also takes the element index and yield a vector of results.

zipWithM_ :: (Monad m, Vector v a, Vector v b) => (a -> b -> m c) -> v a -> v b -> m ()

O(min(m,n)) Zip the two vectors with the monadic action and ignore the results.

izipWithM_ :: (Monad m, Vector v a, Vector v b) => (Int -> a -> b -> m c) -> v a -> v b -> m ()

O(min(m,n)) Zip the two vectors with a monadic action that also takes the element index and ignore the results.

Unzipping

unzip :: (Vector v a, Vector v b, Vector v (a, b)) => v (a, b) -> (v a, v b)

O(min(m,n)) Unzip a vector of pairs.

unzip3 :: (Vector v a, Vector v b, Vector v c, Vector v (a, b, c)) => v (a, b, c) -> (v a, v b, v c)

unzip4 :: (Vector v a, Vector v b, Vector v c, Vector v d, Vector v (a, b, c, d)) => v (a, b, c, d) -> (v a, v b, v c, v d)

unzip5 :: (Vector v a, Vector v b, Vector v c, Vector v d, Vector v e, Vector v (a, b, c, d, e)) => v (a, b, c, d, e) -> (v a, v b, v c, v d, v e)

unzip6 :: (Vector v a, Vector v b, Vector v c, Vector v d, Vector v e, Vector v f, Vector v (a, b, c, d, e, f)) => v (a, b, c, d, e, f) -> (v a, v b, v c, v d, v e, v f)

Working with predicates

Filtering

filter :: Vector v a => (a -> Bool) -> v a -> v a

O(n) Drop all elements that do not satisfy the predicate.

ifilter :: Vector v a => (Int -> a -> Bool) -> v a -> v a

O(n) Drop all elements that do not satisfy the predicate which is applied to the values and their indices.

filterM :: (Monad m, Vector v a) => (a -> m Bool) -> v a -> m (v a)

O(n) Drop all elements that do not satisfy the monadic predicate.

uniq :: (Vector v a, Eq a) => v a -> v a

O(n) Drop repeated adjacent elements. The first element in each group is returned.

Examples

Expand

>>> import qualified Data.Vector as V
>>> V.uniq $ V.fromList [1,3,3,200,3]
[1,3,200,3]
>>> import Data.Semigroup
>>> V.uniq $ V.fromList [ Arg 1 'a', Arg 1 'b', Arg 1 'c']
[Arg 1 'a']

mapMaybe :: (Vector v a, Vector v b) => (a -> Maybe b) -> v a -> v b

O(n) Map the values and collect the Just results.

imapMaybe :: (Vector v a, Vector v b) => (Int -> a -> Maybe b) -> v a -> v b

O(n) Map the indices/values and collect the Just results.

mapMaybeM :: (Monad m, Vector v a, Vector v b) => (a -> m (Maybe b)) -> v a -> m (v b)

O(n) Apply the monadic function to each element of the vector and discard elements returning Nothing.

Since: 0.12.2.0

imapMaybeM :: (Monad m, Vector v a, Vector v b) => (Int -> a -> m (Maybe b)) -> v a -> m (v b)

O(n) Apply the monadic function to each element of the vector and its index. Discard elements returning Nothing.

Since: 0.12.2.0

takeWhile :: Vector v a => (a -> Bool) -> v a -> v a

O(n) Yield the longest prefix of elements satisfying the predicate. The current implementation is not copy-free, unless the result vector is fused away.

dropWhile :: Vector v a => (a -> Bool) -> v a -> v a

O(n) Drop the longest prefix of elements that satisfy the predicate without copying.

Partitioning

partition :: Vector v a => (a -> Bool) -> v a -> (v a, v a)

O(n) Split the vector in two parts, the first one containing those elements that satisfy the predicate and the second one those that don't. The relative order of the elements is preserved at the cost of a sometimes reduced performance compared to unstablePartition.

partitionWith :: (Vector v a, Vector v b, Vector v c) => (a -> Either b c) -> v a -> (v b, v c)

O(n) Split the vector into two parts, the first one containing the Left elements and the second containing the Right elements. The relative order of the elements is preserved.

Since: 0.12.1.0

unstablePartition :: Vector v a => (a -> Bool) -> v a -> (v a, v a)

O(n) Split the vector in two parts, the first one containing those elements that satisfy the predicate and the second one those that don't. The order of the elements is not preserved, but the operation is often faster than partition.

span :: Vector v a => (a -> Bool) -> v a -> (v a, v a)

O(n) Split the vector into the longest prefix of elements that satisfy the predicate and the rest without copying.

break :: Vector v a => (a -> Bool) -> v a -> (v a, v a)

O(n) Split the vector into the longest prefix of elements that do not satisfy the predicate and the rest without copying.

groupBy :: Vector v a => (a -> a -> Bool) -> v a -> [v a]

O(n) Split a vector into a list of slices.

The concatenation of this list of slices is equal to the argument vector, and each slice contains only equal elements, as determined by the equality predicate function.

>>> import qualified Data.Vector as V
>>> import           Data.Char (isUpper)
>>> V.groupBy (\a b -> isUpper a == isUpper b) (V.fromList "Mississippi River")
["M","ississippi ","R","iver"]

See also groupBy.

Since: 0.13.0.1

group :: (Vector v a, Eq a) => v a -> [v a]

O(n) Split a vector into a list of slices.

The concatenation of this list of slices is equal to the argument vector, and each slice contains only equal elements.

This is the equivalent of 'groupBy (==)'.

>>> import qualified Data.Vector as V
>>> V.group (V.fromList "Mississippi")
["M","i","ss","i","ss","i","pp","i"]

See also group.

Since: 0.13.0.1

Searching

elem :: (Vector v a, Eq a) => a -> v a -> Bool infix 4

O(n) Check if the vector contains an element.

notElem :: (Vector v a, Eq a) => a -> v a -> Bool infix 4

O(n) Check if the vector does not contain an element (inverse of elem).

find :: Vector v a => (a -> Bool) -> v a -> Maybe a

O(n) Yield Just the first element matching the predicate or Nothing if no such element exists.

findIndex :: Vector v a => (a -> Bool) -> v a -> Maybe Int

O(n) Yield Just the index of the first element matching the predicate or Nothing if no such element exists.

findIndexR :: Vector v a => (a -> Bool) -> v a -> Maybe Int

O(n) Yield Just the index of the last element matching the predicate or Nothing if no such element exists.

Does not fuse.

Since: 0.12.2.0

findIndices :: (Vector v a, Vector v Int) => (a -> Bool) -> v a -> v Int

O(n) Yield the indices of elements satisfying the predicate in ascending order.

elemIndex :: (Vector v a, Eq a) => a -> v a -> Maybe Int

O(n) Yield Just the index of the first occurrence of the given element or Nothing if the vector does not contain the element. This is a specialised version of findIndex.

elemIndices :: (Vector v a, Vector v Int, Eq a) => a -> v a -> v Int

O(n) Yield the indices of all occurrences of the given element in ascending order. This is a specialised version of findIndices.

Folding

foldl :: Vector v b => (a -> b -> a) -> a -> v b -> a

O(n) Left fold.

foldl1 :: Vector v a => (a -> a -> a) -> v a -> a

O(n) Left fold on non-empty vectors.

foldl' :: Vector v b => (a -> b -> a) -> a -> v b -> a

O(n) Left fold with strict accumulator.

foldl1' :: Vector v a => (a -> a -> a) -> v a -> a

O(n) Left fold on non-empty vectors with strict accumulator.

foldr :: Vector v a => (a -> b -> b) -> b -> v a -> b

O(n) Right fold.

foldr1 :: Vector v a => (a -> a -> a) -> v a -> a

O(n) Right fold on non-empty vectors.

foldr' :: Vector v a => (a -> b -> b) -> b -> v a -> b

O(n) Right fold with a strict accumulator.

foldr1' :: Vector v a => (a -> a -> a) -> v a -> a

O(n) Right fold on non-empty vectors with strict accumulator.

ifoldl :: Vector v b => (a -> Int -> b -> a) -> a -> v b -> a

O(n) Left fold using a function applied to each element and its index.

ifoldl' :: Vector v b => (a -> Int -> b -> a) -> a -> v b -> a

O(n) Left fold with strict accumulator using a function applied to each element and its index.

ifoldr :: Vector v a => (Int -> a -> b -> b) -> b -> v a -> b

O(n) Right fold using a function applied to each element and its index.

ifoldr' :: Vector v a => (Int -> a -> b -> b) -> b -> v a -> b

O(n) Right fold with strict accumulator using a function applied to each element and its index.

foldMap :: (Monoid m, Vector v a) => (a -> m) -> v a -> m

O(n) Map each element of the structure to a monoid and combine the results. It uses the same implementation as the corresponding method of the Foldable type cless. Note that it's implemented in terms of foldr and won't fuse with functions that traverse the vector from left to right (map, generate, etc.).

Since: 0.12.2.0

foldMap' :: (Monoid m, Vector v a) => (a -> m) -> v a -> m

O(n) Like foldMap, but strict in the accumulator. It uses the same implementation as the corresponding method of the Foldable type class. Note that it's implemented in terms of foldl', so it fuses in most contexts.

Since: 0.12.2.0

Specialised folds

all :: Vector v a => (a -> Bool) -> v a -> Bool

O(n) Check if all elements satisfy the predicate.

Examples

Expand

>>> import qualified Data.Vector as V
>>> V.all even $ V.fromList [2, 4, 12]
True
>>> V.all even $ V.fromList [2, 4, 13]
False
>>> V.all even (V.empty :: V.Vector Int)
True

any :: Vector v a => (a -> Bool) -> v a -> Bool

O(n) Check if any element satisfies the predicate.

Examples

Expand

>>> import qualified Data.Vector as V
>>> V.any even $ V.fromList [1, 3, 7]
False
>>> V.any even $ V.fromList [3, 2, 13]
True
>>> V.any even (V.empty :: V.Vector Int)
False

and :: Vector v Bool => v Bool -> Bool

O(n) Check if all elements are True.

Examples

Expand

>>> import qualified Data.Vector as V
>>> V.and $ V.fromList [True, False]
False
>>> V.and V.empty
True

or :: Vector v Bool => v Bool -> Bool

O(n) Check if any element is True.

Examples

Expand

>>> import qualified Data.Vector as V
>>> V.or $ V.fromList [True, False]
True
>>> V.or V.empty
False

sum :: (Vector v a, Num a) => v a -> a

O(n) Compute the sum of the elements.

Examples

Expand

>>> import qualified Data.Vector as V
>>> V.sum $ V.fromList [300,20,1]
321
>>> V.sum (V.empty :: V.Vector Int)
0

product :: (Vector v a, Num a) => v a -> a

O(n) Compute the product of the elements.

Examples

Expand

>>> import qualified Data.Vector as V
>>> V.product $ V.fromList [1,2,3,4]
24
>>> V.product (V.empty :: V.Vector Int)
1

maximum :: (Vector v a, Ord a) => v a -> a

O(n) Yield the maximum element of the vector. The vector may not be empty. In case of a tie, the first occurrence wins.

Examples

Expand

>>> import qualified Data.Vector as V
>>> V.maximum $ V.fromList [2, 1]
2
>>> import Data.Semigroup
>>> V.maximum $ V.fromList [Arg 1 'a', Arg 2 'b']
Arg 2 'b'
>>> V.maximum $ V.fromList [Arg 1 'a', Arg 1 'b']
Arg 1 'a'

maximumBy :: Vector v a => (a -> a -> Ordering) -> v a -> a

O(n) Yield the maximum element of the vector according to the given comparison function. The vector may not be empty. In case of a tie, the first occurrence wins. This behavior is different from maximumBy which returns the last tie.

Examples

Expand

>>> import Data.Ord
>>> import qualified Data.Vector as V
>>> V.maximumBy (comparing fst) $ V.fromList [(2,'a'), (1,'b')]
(2,'a')
>>> V.maximumBy (comparing fst) $ V.fromList [(1,'a'), (1,'b')]
(1,'a')

maximumOn :: (Ord b, Vector v a) => (a -> b) -> v a -> a

O(n) Yield the maximum element of the vector by comparing the results of a key function on each element. In case of a tie, the first occurrence wins. The vector may not be empty.

Examples

Expand

>>> import qualified Data.Vector as V
>>> V.maximumOn fst $ V.fromList [(2,'a'), (1,'b')]
(2,'a')
>>> V.maximumOn fst $ V.fromList [(1,'a'), (1,'b')]
(1,'a')

Since: 0.13.0.0

minimum :: (Vector v a, Ord a) => v a -> a

O(n) Yield the minimum element of the vector. The vector may not be empty. In case of a tie, the first occurrence wins.

Examples

Expand

>>> import qualified Data.Vector as V
>>> V.minimum $ V.fromList [2, 1]
1
>>> import Data.Semigroup
>>> V.minimum $ V.fromList [Arg 2 'a', Arg 1 'b']
Arg 1 'b'
>>> V.minimum $ V.fromList [Arg 1 'a', Arg 1 'b']
Arg 1 'a'

minimumBy :: Vector v a => (a -> a -> Ordering) -> v a -> a

O(n) Yield the minimum element of the vector according to the given comparison function. The vector may not be empty. In case of a tie, the first occurrence wins.

Examples

Expand

>>> import Data.Ord
>>> import qualified Data.Vector as V
>>> V.minimumBy (comparing fst) $ V.fromList [(2,'a'), (1,'b')]
(1,'b')
>>> V.minimumBy (comparing fst) $ V.fromList [(1,'a'), (1,'b')]
(1,'a')

minimumOn :: (Ord b, Vector v a) => (a -> b) -> v a -> a

O(n) Yield the minimum element of the vector by comparing the results of a key function on each element. In case of a tie, the first occurrence wins. The vector may not be empty.

Examples

Expand

>>> import qualified Data.Vector as V
>>> V.minimumOn fst $ V.fromList [(2,'a'), (1,'b')]
(1,'b')
>>> V.minimumOn fst $ V.fromList [(1,'a'), (1,'b')]
(1,'a')

Since: 0.13.0.0

minIndex :: (Vector v a, Ord a) => v a -> Int

O(n) Yield the index of the minimum element of the vector. The vector may not be empty.

minIndexBy :: Vector v a => (a -> a -> Ordering) -> v a -> Int

O(n) Yield the index of the minimum element of the vector according to the given comparison function. The vector may not be empty.

Examples

Expand

>>> import Data.Ord
>>> import qualified Data.Vector as V
>>> V.minIndexBy (comparing fst) $ V.fromList [(2,'a'), (1,'b')]
1
>>> V.minIndexBy (comparing fst) $ V.fromList [(1,'a'), (1,'b')]
0

maxIndex :: (Vector v a, Ord a) => v a -> Int

O(n) Yield the index of the maximum element of the vector. The vector may not be empty.

maxIndexBy :: Vector v a => (a -> a -> Ordering) -> v a -> Int

O(n) Yield the index of the maximum element of the vector according to the given comparison function. The vector may not be empty. In case of a tie, the first occurrence wins.

Examples

Expand

>>> import Data.Ord
>>> import qualified Data.Vector as V
>>> V.maxIndexBy (comparing fst) $ V.fromList [(2,'a'), (1,'b')]
0
>>> V.maxIndexBy (comparing fst) $ V.fromList [(1,'a'), (1,'b')]
0

Monadic folds

foldM :: (Monad m, Vector v b) => (a -> b -> m a) -> a -> v b -> m a

O(n) Monadic fold.

ifoldM :: (Monad m, Vector v b) => (a -> Int -> b -> m a) -> a -> v b -> m a

O(n) Monadic fold using a function applied to each element and its index.

foldM' :: (Monad m, Vector v b) => (a -> b -> m a) -> a -> v b -> m a

O(n) Monadic fold with strict accumulator.

ifoldM' :: (Monad m, Vector v b) => (a -> Int -> b -> m a) -> a -> v b -> m a

O(n) Monadic fold with strict accumulator using a function applied to each element and its index.

fold1M :: (Monad m, Vector v a) => (a -> a -> m a) -> v a -> m a

O(n) Monadic fold over non-empty vectors.

fold1M' :: (Monad m, Vector v a) => (a -> a -> m a) -> v a -> m a

O(n) Monadic fold over non-empty vectors with strict accumulator.

foldM_ :: (Monad m, Vector v b) => (a -> b -> m a) -> a -> v b -> m ()

O(n) Monadic fold that discards the result.

ifoldM_ :: (Monad m, Vector v b) => (a -> Int -> b -> m a) -> a -> v b -> m ()

O(n) Monadic fold that discards the result using a function applied to each element and its index.

foldM'_ :: (Monad m, Vector v b) => (a -> b -> m a) -> a -> v b -> m ()

O(n) Monadic fold with strict accumulator that discards the result.

ifoldM'_ :: (Monad m, Vector v b) => (a -> Int -> b -> m a) -> a -> v b -> m ()

O(n) Monadic fold with strict accumulator that discards the result using a function applied to each element and its index.

fold1M_ :: (Monad m, Vector v a) => (a -> a -> m a) -> v a -> m ()

O(n) Monadic fold over non-empty vectors that discards the result.

fold1M'_ :: (Monad m, Vector v a) => (a -> a -> m a) -> v a -> m ()

O(n) Monad fold over non-empty vectors with strict accumulator that discards the result.

Monadic sequencing

sequence :: (Monad m, Vector v a, Vector v (m a)) => v (m a) -> m (v a)

Evaluate each action and collect the results.

sequence_ :: (Monad m, Vector v (m a)) => v (m a) -> m ()

Evaluate each action and discard the results.

Scans

prescanl :: (Vector v a, Vector v b) => (a -> b -> a) -> a -> v b -> v a

O(n) Left-to-right prescan.

prescanl f z = init . scanl f z

Examples

Expand

>>> import qualified Data.Vector as V
>>> V.prescanl (+) 0 (V.fromList [1,2,3,4])
[0,1,3,6]

prescanl' :: (Vector v a, Vector v b) => (a -> b -> a) -> a -> v b -> v a

O(n) Left-to-right prescan with strict accumulator.

postscanl :: (Vector v a, Vector v b) => (a -> b -> a) -> a -> v b -> v a

O(n) Left-to-right postscan.

postscanl f z = tail . scanl f z

Examples

Expand

>>> import qualified Data.Vector as V
>>> V.postscanl (+) 0 (V.fromList [1,2,3,4])
[1,3,6,10]

postscanl' :: (Vector v a, Vector v b) => (a -> b -> a) -> a -> v b -> v a

O(n) Left-to-right postscan with strict accumulator.

scanl :: (Vector v a, Vector v b) => (a -> b -> a) -> a -> v b -> v a

O(n) Left-to-right scan.

scanl f z <x1,...,xn> = <y1,...,y(n+1)>
  where y1 = z
        yi = f y(i-1) x(i-1)

Examples

Expand

>>> import qualified Data.Vector as V
>>> V.scanl (+) 0 (V.fromList [1,2,3,4])
[0,1,3,6,10]

scanl' :: (Vector v a, Vector v b) => (a -> b -> a) -> a -> v b -> v a

O(n) Left-to-right scan with strict accumulator.

scanl1 :: Vector v a => (a -> a -> a) -> v a -> v a

O(n) Initial-value free left-to-right scan over a vector.

scanl f <x1,...,xn> = <y1,...,yn>
  where y1 = x1
        yi = f y(i-1) xi

Note: Since 0.13, application of this to an empty vector no longer results in an error; instead it produces an empty vector.

Examples

Expand

>>> import qualified Data.Vector as V
>>> V.scanl1 min $ V.fromListN 5 [4,2,4,1,3]
[4,2,2,1,1]
>>> V.scanl1 max $ V.fromListN 5 [1,3,2,5,4]
[1,3,3,5,5]
>>> V.scanl1 min (V.empty :: V.Vector Int)
[]

scanl1' :: Vector v a => (a -> a -> a) -> v a -> v a

O(n) Initial-value free left-to-right scan over a vector with a strict accumulator.

Note: Since 0.13, application of this to an empty vector no longer results in an error; instead it produces an empty vector.

Examples

Expand

>>> import qualified Data.Vector as V
>>> V.scanl1' min $ V.fromListN 5 [4,2,4,1,3]
[4,2,2,1,1]
>>> V.scanl1' max $ V.fromListN 5 [1,3,2,5,4]
[1,3,3,5,5]
>>> V.scanl1' min (V.empty :: V.Vector Int)
[]

iscanl :: (Vector v a, Vector v b) => (Int -> a -> b -> a) -> a -> v b -> v a

O(n) Left-to-right scan over a vector with its index.

iscanl' :: (Vector v a, Vector v b) => (Int -> a -> b -> a) -> a -> v b -> v a

O(n) Left-to-right scan over a vector (strictly) with its index.

prescanr :: (Vector v a, Vector v b) => (a -> b -> b) -> b -> v a -> v b

O(n) Right-to-left prescan.

prescanr f z = reverse . prescanl (flip f) z . reverse

prescanr' :: (Vector v a, Vector v b) => (a -> b -> b) -> b -> v a -> v b

O(n) Right-to-left prescan with strict accumulator.

postscanr :: (Vector v a, Vector v b) => (a -> b -> b) -> b -> v a -> v b

O(n) Right-to-left postscan.

postscanr' :: (Vector v a, Vector v b) => (a -> b -> b) -> b -> v a -> v b

O(n) Right-to-left postscan with strict accumulator.

scanr :: (Vector v a, Vector v b) => (a -> b -> b) -> b -> v a -> v b

O(n) Right-to-left scan.

scanr' :: (Vector v a, Vector v b) => (a -> b -> b) -> b -> v a -> v b

O(n) Right-to-left scan with strict accumulator.

scanr1 :: Vector v a => (a -> a -> a) -> v a -> v a

O(n) Right-to-left, initial-value free scan over a vector.

Note: Since 0.13, application of this to an empty vector no longer results in an error; instead it produces an empty vector.

Examples

Expand

>>> import qualified Data.Vector as V
>>> V.scanr1 min $ V.fromListN 5 [3,1,4,2,4]
[1,1,2,2,4]
>>> V.scanr1 max $ V.fromListN 5 [4,5,2,3,1]
[5,5,3,3,1]
>>> V.scanr1 min (V.empty :: V.Vector Int)
[]

scanr1' :: Vector v a => (a -> a -> a) -> v a -> v a

O(n) Right-to-left, initial-value free scan over a vector with a strict accumulator.

Note: Since 0.13, application of this to an empty vector no longer results in an error; instead it produces an empty vector.

Examples

Expand

>>> import qualified Data.Vector as V
>>> V.scanr1' min $ V.fromListN 5 [3,1,4,2,4]
[1,1,2,2,4]
>>> V.scanr1' max $ V.fromListN 5 [4,5,2,3,1]
[5,5,3,3,1]
>>> V.scanr1' min (V.empty :: V.Vector Int)
[]

iscanr :: (Vector v a, Vector v b) => (Int -> a -> b -> b) -> b -> v a -> v b

O(n) Right-to-left scan over a vector with its index.

iscanr' :: (Vector v a, Vector v b) => (Int -> a -> b -> b) -> b -> v a -> v b

O(n) Right-to-left scan over a vector (strictly) with its index.

Conversions

Lists

toList :: Vector v a => v a -> [a]

O(n) Convert a vector to a list.

fromList :: Vector v a => [a] -> v a

O(n) Convert a list to a vector.

fromListN :: Vector v a => Int -> [a] -> v a

O(n) Convert the first n elements of a list to a vector. It's expected that the supplied list will be exactly n elements long. As an optimization, this function allocates a buffer for n elements, which could be used for DoS-attacks by exhausting the memory if an attacker controls that parameter.

fromListN n xs = fromList (take n xs)

Examples

Expand

>>> import qualified Data.Vector as V
>>> V.fromListN 3 [1,2,3,4,5]
[1,2,3]
>>> V.fromListN 3 [1]
[1]

Different vector types

convert :: (Vector v a, Vector w a) => v a -> w a

O(n) Convert between different vector types.

Mutable vectors

freeze :: (PrimMonad m, Vector v a) => Mutable v (PrimState m) a -> m (v a)

O(n) Yield an immutable copy of the mutable vector.

thaw :: (PrimMonad m, Vector v a) => v a -> m (Mutable v (PrimState m) a)

O(n) Yield a mutable copy of an immutable vector.

copy :: (HasCallStack, PrimMonad m, Vector v a) => Mutable v (PrimState m) a -> v a -> m ()

O(n) Copy an immutable vector into a mutable one. The two vectors must have the same length.

unsafeFreeze :: (PrimMonad m, Vector v a) => Mutable v (PrimState m) a -> m (v a)

O(1) Unsafely convert a mutable vector to an immutable one without copying. The mutable vector may not be used after this operation.

unsafeThaw :: (PrimMonad m, Vector v a) => v a -> m (Mutable v (PrimState m) a)

O(1) Unsafely convert an immutable vector to a mutable one without copying. Note that this is a very dangerous function and generally it's only safe to read from the resulting vector. In this case, the immutable vector could be used safely as well.

Problems with mutation happen because GHC has a lot of freedom to introduce sharing. As a result mutable vectors produced by unsafeThaw may or may not share the same underlying buffer. For example:

foo = do
  let vec = V.generate 10 id
  mvec <- V.unsafeThaw vec
  do_something mvec

Here GHC could lift vec outside of foo which means that all calls to do_something will use same buffer with possibly disastrous results. Whether such aliasing happens or not depends on the program in question, optimization levels, and GHC flags.

All in all, attempts to modify a vector produced by unsafeThaw fall out of domain of software engineering and into realm of black magic, dark rituals, and unspeakable horrors. The only advice that could be given is: "Don't attempt to mutate a vector produced by unsafeThaw unless you know how to prevent GHC from aliasing buffers accidentally. We don't."

unsafeCopy :: (PrimMonad m, Vector v a) => Mutable v (PrimState m) a -> v a -> m ()

O(n) Copy an immutable vector into a mutable one. The two vectors must have the same length. This is not checked.

Fusion support

Conversion to/from Bundles

stream :: Vector v a => v a -> Bundle v a

O(1) Convert a vector to a Bundle.

unstream :: Vector v a => Bundle v a -> v a

O(n) Construct a vector from a Bundle.

unstreamM :: (Monad m, Vector v a) => MBundle m u a -> m (v a)

Load a monadic stream bundle into a newly allocated vector. This function goes through a list, so prefer using unstream, unless you need to be in a monad.

Since: 0.12.2.0

streamR :: Vector v a => v a -> Bundle u a

O(1) Convert a vector to a Bundle, proceeding from right to left.

unstreamR :: Vector v a => Bundle v a -> v a

O(n) Construct a vector from a Bundle, proceeding from right to left.

Recycling support

new :: Vector v a => New v a -> v a

Construct a vector from a monadic initialiser.

clone :: Vector v a => v a -> New v a

Convert a vector to an initialiser which, when run, produces a copy of the vector.

Utilities

Comparisons

eq :: (Vector v a, Eq a) => v a -> v a -> Bool

O(n) Check if two vectors are equal. All Vector instances are also instances of Eq and it is usually more appropriate to use those. This function is primarily intended for implementing Eq instances for new vector types.

cmp :: (Vector v a, Ord a) => v a -> v a -> Ordering

O(n) Compare two vectors lexicographically. All Vector instances are also instances of Ord and it is usually more appropriate to use those. This function is primarily intended for implementing Ord instances for new vector types.

eqBy :: (Vector v a, Vector v b) => (a -> b -> Bool) -> v a -> v b -> Bool

O(n) Check if two vectors are equal using the supplied equality predicate.

cmpBy :: (Vector v a, Vector v b) => (a -> b -> Ordering) -> v a -> v b -> Ordering

O(n) Compare two vectors using the supplied comparison function for vector elements. Comparison works the same as for lists.

cmpBy compare == cmp

Show and Read

showsPrec :: (Vector v a, Show a) => Int -> v a -> ShowS

Generic definition of showsPrec.

readPrec :: (Vector v a, Read a) => ReadPrec (v a)

Generic definition of readPrec.

liftShowsPrec :: Vector v a => (Int -> a -> ShowS) -> ([a] -> ShowS) -> Int -> v a -> ShowS

liftReadsPrec :: Vector v a => (Int -> ReadS a) -> ReadS [a] -> Int -> ReadS (v a)

Note: uses ReadS.

Data and Typeable

gfoldl :: (Vector v a, Data a) => (forall d b. Data d => c (d -> b) -> d -> c b) -> (forall g. g -> c g) -> v a -> c (v a)

Generic definion of gfoldl that views a Vector as a list.

gunfold :: (Vector v a, Data a, HasCallStack) => (forall b r. Data b => c (b -> r) -> c r) -> (forall r. r -> c r) -> Constr -> c (v a)

dataCast :: (Vector v a, Data a, Typeable v, Typeable t) => (forall d. Data d => c (t d)) -> Maybe (c (v a))

mkVecType :: String -> DataType

mkVecConstr :: String -> Constr

mkType :: String -> DataType

Deprecated: Use Data.Data.mkNoRepType