Data.Vector.Generic.Mutable

Description

Generic interface to mutable vectors

Class of mutable vector types

class MVector v a where

Class of mutable vectors parametrised with a primitive state token.

Minimal complete definition

basicLength, basicUnsafeSlice, basicOverlaps, basicUnsafeNew, basicInitialize, basicUnsafeRead, basicUnsafeWrite

Methods

basicLength :: v s a -> Int

Length of the mutable vector. This method should not be called directly, use length instead.

basicUnsafeSlice

Arguments

:: Int
starting index

-> Int
length of the slice

-> v s a

-> v s a

Yield a part of the mutable vector without copying it. This method should not be called directly, use unsafeSlice instead.

basicOverlaps :: v s a -> v s a -> Bool

Check whether two vectors overlap. This method should not be called directly, use overlaps instead.

basicUnsafeNew :: PrimMonad m => Int -> m (v (PrimState m) a)

Create a mutable vector of the given length. This method should not be called directly, use unsafeNew instead.

basicInitialize :: PrimMonad m => v (PrimState m) a -> m ()

Initialize a vector to a standard value. This is intended to be called as part of the safe new operation (and similar operations), to properly blank the newly allocated memory if necessary.

Vectors that are necessarily initialized as part of creation may implement this as a no-op.

Since: 0.11.0.0

basicUnsafeReplicate :: PrimMonad m => Int -> a -> m (v (PrimState m) a)

Create a mutable vector of the given length and fill it with an initial value. This method should not be called directly, use replicate instead.

basicUnsafeRead :: PrimMonad m => v (PrimState m) a -> Int -> m a

Yield the element at the given position. This method should not be called directly, use unsafeRead instead.

basicUnsafeWrite :: PrimMonad m => v (PrimState m) a -> Int -> a -> m ()

Replace the element at the given position. This method should not be called directly, use unsafeWrite instead.

basicClear :: PrimMonad m => v (PrimState m) a -> m ()

Reset all elements of the vector to some undefined value, clearing all references to external objects. This is usually a noop for unboxed vectors. This method should not be called directly, use clear instead.

basicSet :: PrimMonad m => v (PrimState m) a -> a -> m ()

Set all elements of the vector to the given value. This method should not be called directly, use set instead.

basicUnsafeCopy

Arguments

:: PrimMonad m

=> v (PrimState m) a
target

-> v (PrimState m) a
source

-> m ()

Copy a vector. The two vectors may not overlap. This method should not be called directly, use unsafeCopy instead.

basicUnsafeMove

Arguments

:: PrimMonad m

=> v (PrimState m) a
target

-> v (PrimState m) a
source

-> m ()

Move the contents of a vector. The two vectors may overlap. This method should not be called directly, use unsafeMove instead.

basicUnsafeGrow :: PrimMonad m => v (PrimState m) a -> Int -> m (v (PrimState m) a)

Grow a vector by the given number of elements. Allocates a new vector and copies all of the elements over starting at 0 index. This method should not be called directly, use grow/unsafeGrow instead.

Accessors

Length information

length :: MVector v a => v s a -> Int

Length of the mutable vector.

null :: MVector v a => v s a -> Bool

Check whether the vector is empty

Extracting subvectors

slice

Arguments

:: MVector v a

=> Int
i starting index

-> Int
n length

-> v s a

-> v s a

Yield a part of the mutable vector without copying it. The vector must contain at least i+n elements.

init :: MVector v a => v s a -> v s a

tail :: MVector v a => v s a -> v s a

take :: MVector v a => Int -> v s a -> v s a

drop :: MVector v a => Int -> v s a -> v s a

splitAt :: MVector v a => Int -> v s a -> (v s a, v s a)

unsafeSlice

Arguments

:: MVector v a

=> Int
starting index

-> Int
length of the slice

-> v s a

-> v s a

Yield a part of the mutable vector without copying it. No bounds checks are performed.

unsafeInit :: MVector v a => v s a -> v s a

unsafeTail :: MVector v a => v s a -> v s a

unsafeTake :: MVector v a => Int -> v s a -> v s a

unsafeDrop :: MVector v a => Int -> v s a -> v s a

Overlapping

overlaps :: MVector v a => v s a -> v s a -> Bool

Check whether two vectors overlap.

Construction

Initialisation

new :: (PrimMonad m, MVector v a) => Int -> m (v (PrimState m) a)

Create a mutable vector of the given length.

unsafeNew :: (PrimMonad m, MVector v a) => Int -> m (v (PrimState m) a)

Create a mutable vector of the given length. The vector content should be presumed uninitialized. However exact semantics depends on vector implementation. For example unboxed and storable vectors will create vector filled with whatever underlying memory buffer happens to contain, while boxed vector's elements are initialized to bottoms which will throw exception when evaluated.

Since: 0.4

replicate :: (PrimMonad m, MVector v a) => Int -> a -> m (v (PrimState m) a)

Create a mutable vector of the given length (0 if the length is negative) and fill it with an initial value.

replicateM :: (PrimMonad m, MVector v a) => Int -> m a -> m (v (PrimState m) a)

Create a mutable vector of the given length (0 if the length is negative) and fill it with values produced by repeatedly executing the monadic action.

generate :: (PrimMonad m, MVector v a) => Int -> (Int -> a) -> m (v (PrimState m) a)

O(n) Create a mutable vector of the given length (0 if the length is negative) and fill it with the results of applying the function to each index.

Since: 0.12.3.0

generateM :: (PrimMonad m, MVector v a) => Int -> (Int -> m a) -> m (v (PrimState m) a)

O(n) Create a mutable vector of the given length (0 if the length is negative) and fill it with the results of applying the monadic function to each index. Iteration starts at index 0.

Since: 0.12.3.0

clone :: (PrimMonad m, MVector v a) => v (PrimState m) a -> m (v (PrimState m) a)

Create a copy of a mutable vector.

Growing

grow :: (PrimMonad m, MVector v a) => v (PrimState m) a -> Int -> m (v (PrimState m) a)

Grow a vector by the given number of elements. The number must not be negative otherwise error is thrown. Semantics of this function is exactly the same as unsafeGrow, except that it will initialize the newly allocated memory first.

It is important to note that mutating the returned vector will not affect the vector that was used as a source. In other words it does not, nor will it ever have the semantics of realloc from C.

grow mv 0 === clone mv

Since: 0.4.0

unsafeGrow

Arguments

:: (PrimMonad m, MVector v a)

=> v (PrimState m) a
A mutable vector to copy the data from.

-> Int
Number of elements to grow the vector by. It must be non-negative but this is not checked.

-> m (v (PrimState m) a)

Grow a vector by allocating a new mutable vector of the same size plus the the given number of elements and copying all the data over to the new vector starting at its beginning. The newly allocated memory is not initialized and the extra space at the end will likely contain garbage data or uninitialzed error. Use unsafeGrowFront to make the extra space available in the front of the new vector.

It is important to note that mutating the returned vector will not affect elements of the vector that was used as a source. In other words it does not, nor will it ever have the semantics of realloc from C. Keep in mind, however, that values themselves can be of a mutable type (eg. Ptr), in which case it would be possible to affect values stored in both vectors.

unsafeGrow mv 0 === clone mv

Since: 0.4.0

growFront :: (PrimMonad m, MVector v a) => v (PrimState m) a -> Int -> m (v (PrimState m) a)

Same as grow, except that it copies data towards the end of the newly allocated vector making extra space available at the beginning.

Since: 0.11.0.0

unsafeGrowFront :: (PrimMonad m, MVector v a) => v (PrimState m) a -> Int -> m (v (PrimState m) a)

Same as unsafeGrow, except that it copies data towards the end of the newly allocated vector making extra space available at the beginning.

Since: 0.11.0.0

Restricting memory usage

clear :: (PrimMonad m, MVector v a) => v (PrimState m) a -> m ()

Reset all elements of the vector to some undefined value, clearing all references to external objects. This is usually a noop for unboxed vectors.

Accessing individual elements

read :: (PrimMonad m, MVector v a) => v (PrimState m) a -> Int -> m a

Yield the element at the given position.

write :: (PrimMonad m, MVector v a) => v (PrimState m) a -> Int -> a -> m ()

Replace the element at the given position.

modify :: (PrimMonad m, MVector v a) => v (PrimState m) a -> (a -> a) -> Int -> m ()

Modify the element at the given position.

modifyM :: (PrimMonad m, MVector v a) => v (PrimState m) a -> (a -> m a) -> Int -> m ()

Modify the element at the given position using a monadic function.

Since: 0.12.3.0

swap :: (PrimMonad m, MVector v a) => v (PrimState m) a -> Int -> Int -> m ()

Swap the elements at the given positions.

exchange :: (PrimMonad m, MVector v a) => v (PrimState m) a -> Int -> a -> m a

Replace the element at the given position and return the old element.

unsafeRead :: (PrimMonad m, MVector v a) => v (PrimState m) a -> Int -> m a

Yield the element at the given position. No bounds checks are performed.

unsafeWrite :: (PrimMonad m, MVector v a) => v (PrimState m) a -> Int -> a -> m ()

Replace the element at the given position. No bounds checks are performed.

unsafeModify :: (PrimMonad m, MVector v a) => v (PrimState m) a -> (a -> a) -> Int -> m ()

Modify the element at the given position. No bounds checks are performed.

unsafeModifyM :: (PrimMonad m, MVector v a) => v (PrimState m) a -> (a -> m a) -> Int -> m ()

Modify the element at the given position using a monadic function. No bounds checks are performed.

Since: 0.12.3.0

unsafeSwap :: (PrimMonad m, MVector v a) => v (PrimState m) a -> Int -> Int -> m ()

Swap the elements at the given positions. No bounds checks are performed.

unsafeExchange :: (PrimMonad m, MVector v a) => v (PrimState m) a -> Int -> a -> m a

Replace the element at the given position and return the old element. No bounds checks are performed.

Folds

mapM_ :: (PrimMonad m, MVector v a) => (a -> m b) -> v (PrimState m) a -> m ()

O(n) Apply the monadic action to every element of the vector, discarding the results.

Since: 0.12.3.0

imapM_ :: (PrimMonad m, MVector v a) => (Int -> a -> m b) -> v (PrimState m) a -> m ()

O(n) Apply the monadic action to every element of the vector and its index, discarding the results.

Since: 0.12.3.0

forM_ :: (PrimMonad m, MVector v a) => v (PrimState m) a -> (a -> m b) -> m ()

O(n) Apply the monadic action to every element of the vector, discarding the results. It's same as the flip mapM_.

Since: 0.12.3.0

iforM_ :: (PrimMonad m, MVector v a) => v (PrimState m) a -> (Int -> a -> m b) -> m ()

O(n) Apply the monadic action to every element of the vector and its index, discarding the results. It's same as the flip imapM_.

Since: 0.12.3.0

foldl :: (PrimMonad m, MVector v a) => (b -> a -> b) -> b -> v (PrimState m) a -> m b

O(n) Pure left fold.

Since: 0.12.3.0

foldl' :: (PrimMonad m, MVector v a) => (b -> a -> b) -> b -> v (PrimState m) a -> m b

O(n) Pure left fold with strict accumulator.

Since: 0.12.3.0

foldM :: (PrimMonad m, MVector v a) => (b -> a -> m b) -> b -> v (PrimState m) a -> m b

O(n) Monadic fold.

Since: 0.12.3.0

foldM' :: (PrimMonad m, MVector v a) => (b -> a -> m b) -> b -> v (PrimState m) a -> m b

O(n) Monadic fold with strict accumulator.

Since: 0.12.3.0

foldr :: (PrimMonad m, MVector v a) => (a -> b -> b) -> b -> v (PrimState m) a -> m b

O(n) Pure right fold.

Since: 0.12.3.0

foldr' :: (PrimMonad m, MVector v a) => (a -> b -> b) -> b -> v (PrimState m) a -> m b

O(n) Pure right fold with strict accumulator.

Since: 0.12.3.0

foldrM :: (PrimMonad m, MVector v a) => (a -> b -> m b) -> b -> v (PrimState m) a -> m b

O(n) Monadic right fold.

Since: 0.12.3.0

foldrM' :: (PrimMonad m, MVector v a) => (a -> b -> m b) -> b -> v (PrimState m) a -> m b

O(n) Monadic right fold with strict accumulator.

Since: 0.12.3.0

ifoldl :: (PrimMonad m, MVector v a) => (b -> Int -> a -> b) -> b -> v (PrimState m) a -> m b

O(n) Pure left fold (function applied to each element and its index).

Since: 0.12.3.0

ifoldl' :: (PrimMonad m, MVector v a) => (b -> Int -> a -> b) -> b -> v (PrimState m) a -> m b

O(n) Pure left fold with strict accumulator (function applied to each element and its index).

Since: 0.12.3.0

ifoldM :: (PrimMonad m, MVector v a) => (b -> Int -> a -> m b) -> b -> v (PrimState m) a -> m b

O(n) Monadic fold (action applied to each element and its index).

Since: 0.12.3.0

ifoldM' :: (PrimMonad m, MVector v a) => (b -> Int -> a -> m b) -> b -> v (PrimState m) a -> m b

O(n) Monadic fold with strict accumulator (action applied to each element and its index).

Since: 0.12.3.0

ifoldr :: (PrimMonad m, MVector v a) => (Int -> a -> b -> b) -> b -> v (PrimState m) a -> m b

O(n) Pure right fold (function applied to each element and its index).

Since: 0.12.3.0

ifoldr' :: (PrimMonad m, MVector v a) => (Int -> a -> b -> b) -> b -> v (PrimState m) a -> m b

O(n) Pure right fold with strict accumulator (function applied to each element and its index).

Since: 0.12.3.0

ifoldrM :: (PrimMonad m, MVector v a) => (Int -> a -> b -> m b) -> b -> v (PrimState m) a -> m b

O(n) Monadic right fold (action applied to each element and its index).

Since: 0.12.3.0

ifoldrM' :: (PrimMonad m, MVector v a) => (Int -> a -> b -> m b) -> b -> v (PrimState m) a -> m b

O(n) Monadic right fold with strict accumulator (action applied to each element and its index).

Since: 0.12.3.0

Modifying vectors

nextPermutation :: (PrimMonad m, Ord e, MVector v e) => v (PrimState m) e -> m Bool

Compute the next (lexicographically) permutation of given vector in-place. Returns False when input is the last permutation

Filling and copying

set :: (PrimMonad m, MVector v a) => v (PrimState m) a -> a -> m ()

Set all elements of the vector to the given value.

copy

Arguments

:: (PrimMonad m, MVector v a)

=> v (PrimState m) a
target

-> v (PrimState m) a
source

-> m ()

Copy a vector. The two vectors must have the same length and may not overlap.

move

Arguments

:: (PrimMonad m, MVector v a)

=> v (PrimState m) a
target

-> v (PrimState m) a
source

-> m ()

Move the contents of a vector. The two vectors must have the same length.

If the vectors do not overlap, then this is equivalent to copy. Otherwise, the copying is performed as if the source vector were copied to a temporary vector and then the temporary vector was copied to the target vector.

unsafeCopy

Arguments

:: (PrimMonad m, MVector v a)

=> v (PrimState m) a
target

-> v (PrimState m) a
source

-> m ()

Copy a vector. The two vectors must have the same length and may not overlap. This is not checked.

unsafeMove

Arguments

:: (PrimMonad m, MVector v a)

=> v (PrimState m) a
target

-> v (PrimState m) a
source

-> m ()

Move the contents of a vector. The two vectors must have the same length, but this is not checked.

If the vectors do not overlap, then this is equivalent to unsafeCopy. Otherwise, the copying is performed as if the source vector were copied to a temporary vector and then the temporary vector was copied to the target vector.

Internal operations

mstream :: (PrimMonad m, MVector v a) => v (PrimState m) a -> Stream m a

mstreamR :: (PrimMonad m, MVector v a) => v (PrimState m) a -> Stream m a

unstream :: (PrimMonad m, MVector v a) => Bundle u a -> m (v (PrimState m) a)

Create a new mutable vector and fill it with elements from the Bundle. The vector will grow exponentially if the maximum size of the Bundle is unknown.

unstreamR :: (PrimMonad m, MVector v a) => Bundle u a -> m (v (PrimState m) a)

Create a new mutable vector and fill it with elements from the Bundle from right to left. The vector will grow exponentially if the maximum size of the Bundle is unknown.

vunstream :: (PrimMonad m, Vector v a) => Bundle v a -> m (Mutable v (PrimState m) a)

Create a new mutable vector and fill it with elements from the Bundle. The vector will grow exponentially if the maximum size of the Bundle is unknown.

munstream :: (PrimMonad m, MVector v a) => MBundle m u a -> m (v (PrimState m) a)

Create a new mutable vector and fill it with elements from the monadic stream. The vector will grow exponentially if the maximum size of the stream is unknown.

munstreamR :: (PrimMonad m, MVector v a) => MBundle m u a -> m (v (PrimState m) a)

Create a new mutable vector and fill it with elements from the monadic stream from right to left. The vector will grow exponentially if the maximum size of the stream is unknown.

transform :: (PrimMonad m, MVector v a) => (Stream m a -> Stream m a) -> v (PrimState m) a -> m (v (PrimState m) a)

transformR :: (PrimMonad m, MVector v a) => (Stream m a -> Stream m a) -> v (PrimState m) a -> m (v (PrimState m) a)

fill :: (PrimMonad m, MVector v a) => v (PrimState m) a -> Stream m a -> m (v (PrimState m) a)

fillR :: (PrimMonad m, MVector v a) => v (PrimState m) a -> Stream m a -> m (v (PrimState m) a)

unsafeAccum :: (PrimMonad m, MVector v a) => (a -> b -> a) -> v (PrimState m) a -> Bundle u (Int, b) -> m ()

accum :: (PrimMonad m, MVector v a) => (a -> b -> a) -> v (PrimState m) a -> Bundle u (Int, b) -> m ()

unsafeUpdate :: (PrimMonad m, MVector v a) => v (PrimState m) a -> Bundle u (Int, a) -> m ()

update :: (PrimMonad m, MVector v a) => v (PrimState m) a -> Bundle u (Int, a) -> m ()

reverse :: (PrimMonad m, MVector v a) => v (PrimState m) a -> m ()

unstablePartition :: forall m v a. (PrimMonad m, MVector v a) => (a -> Bool) -> v (PrimState m) a -> m Int

unstablePartitionBundle :: (PrimMonad m, MVector v a) => (a -> Bool) -> Bundle u a -> m (v (PrimState m) a, v (PrimState m) a)

partitionBundle :: (PrimMonad m, MVector v a) => (a -> Bool) -> Bundle u a -> m (v (PrimState m) a, v (PrimState m) a)

partitionWithBundle :: (PrimMonad m, MVector v a, MVector v b, MVector v c) => (a -> Either b c) -> Bundle u a -> m (v (PrimState m) b, v (PrimState m) c)