Data.Vector.Mutable

Description

Mutable boxed vectors.

Mutable boxed vectors

data MVector s a

Mutable boxed vectors keyed on the monad they live in (IO or ST s).

Constructors

MVector !Int !Int !(MutableArray s a)

type IOVector = MVector RealWorld

type STVector s = MVector s

Accessors

Length information

length :: MVector s a -> Int

Length of the mutable vector.

null :: MVector s a -> Bool

Check whether the vector is empty.

Extracting subvectors

slice

Arguments

:: Int
i starting index

-> Int
n length

-> MVector s a

-> MVector s a

Yield a part of the mutable vector without copying it. The vector must contain at least i+n elements.

init :: MVector s a -> MVector s a

Drop the last element of the mutable vector without making a copy. If the vector is empty, an exception is thrown.

tail :: MVector s a -> MVector s a

Drop the first element of the mutable vector without making a copy. If the vector is empty, an exception is thrown.

take :: Int -> MVector s a -> MVector s a

Take the n first elements of the mutable vector without making a copy. For negative n, the empty vector is returned. If n is larger than the vector's length, the vector is returned unchanged.

drop :: Int -> MVector s a -> MVector s a

Drop the n first element of the mutable vector without making a copy. For negative n, the vector is returned unchanged. If n is larger than the vector's length, the empty vector is returned.

splitAt :: Int -> MVector s a -> (MVector s a, MVector s a)

O(1) Split the mutable vector into the first n elements and the remainder, without copying.

Note that splitAt n v is equivalent to (take n v, drop n v), but slightly more efficient.

unsafeSlice

Arguments

:: Int
starting index

-> Int
length of the slice

-> MVector s a

-> MVector s a

Yield a part of the mutable vector without copying it. No bounds checks are performed.

unsafeInit :: MVector s a -> MVector s a

Same as init, but doesn't do range checks.

unsafeTail :: MVector s a -> MVector s a

Same as tail, but doesn't do range checks.

unsafeTake :: Int -> MVector s a -> MVector s a

Unsafe variant of take. If n is out of range, it will simply create an invalid slice that likely violate memory safety.

unsafeDrop :: Int -> MVector s a -> MVector s a

Unsafe variant of drop. If n is out of range, it will simply create an invalid slice that likely violate memory safety.

Overlapping

overlaps :: MVector s a -> MVector s a -> Bool

Check whether two vectors overlap.

Construction

Initialisation

new :: PrimMonad m => Int -> m (MVector (PrimState m) a)

Create a mutable vector of the given length.

unsafeNew :: PrimMonad m => Int -> m (MVector (PrimState m) a)

Create a mutable vector of the given length. The vector elements are set to bottom, so accessing them will cause an exception.

Since: 0.5

replicate :: PrimMonad m => Int -> a -> m (MVector (PrimState m) a)

Create a mutable vector of the given length (0 if the length is negative) and fill it with an initial value.

replicateM :: PrimMonad m => Int -> m a -> m (MVector (PrimState m) a)

Create a mutable vector of the given length (0 if the length is negative) and fill it with values produced by repeatedly executing the monadic action.

generate :: PrimMonad m => Int -> (Int -> a) -> m (MVector (PrimState m) a)

O(n) Create a mutable vector of the given length (0 if the length is negative) and fill it with the results of applying the function to each index. Iteration starts at index 0.

Since: 0.12.3.0

generateM :: PrimMonad m => Int -> (Int -> m a) -> m (MVector (PrimState m) a)

O(n) Create a mutable vector of the given length (0 if the length is negative) and fill it with the results of applying the monadic function to each index. Iteration starts at index 0.

Since: 0.12.3.0

clone :: PrimMonad m => MVector (PrimState m) a -> m (MVector (PrimState m) a)

Create a copy of a mutable vector.

Growing

grow :: PrimMonad m => MVector (PrimState m) a -> Int -> m (MVector (PrimState m) a)

Grow a boxed vector by the given number of elements. The number must be non-negative. This has the same semantics as grow for generic vectors. It differs from grow functions for unpacked vectors, however, in that only pointers to values are copied over, therefore the values themselves will be shared between the two vectors. This is an important distinction to know about during memory usage analysis and in case the values themselves are of a mutable type, e.g. IORef or another mutable vector.

Examples

Expand

>>> import qualified Data.Vector as V
>>> import qualified Data.Vector.Mutable as MV
>>> mv <- V.thaw $ V.fromList ([10, 20, 30] :: [Integer])
>>> mv' <- MV.grow mv 2

The two extra elements at the end of the newly allocated vector will be uninitialized and will result in an error if evaluated, so me must overwrite them with new values first:

>>> MV.write mv' 3 999
>>> MV.write mv' 4 777
>>> V.freeze mv'
[10,20,30,999,777]

It is important to note that the source mutable vector is not affected when the newly allocated one is mutated.

>>> MV.write mv' 2 888
>>> V.freeze mv'
[10,20,888,999,777]
>>> V.freeze mv
[10,20,30]

Since: 0.5

unsafeGrow :: PrimMonad m => MVector (PrimState m) a -> Int -> m (MVector (PrimState m) a)

Grow a vector by the given number of elements. The number must be non-negative, but this is not checked. This has the same semantics as unsafeGrow for generic vectors.

Since: 0.5

Restricting memory usage

clear :: PrimMonad m => MVector (PrimState m) a -> m ()

Reset all elements of the vector to some undefined value, clearing all references to external objects.

Accessing individual elements

read :: PrimMonad m => MVector (PrimState m) a -> Int -> m a

Yield the element at the given position. Will throw an exception if the index is out of range.

Examples

Expand

>>> import qualified Data.Vector.Mutable as MV
>>> v <- MV.generate 10 (\x -> x*x)
>>> MV.read v 3
9

readMaybe :: PrimMonad m => MVector (PrimState m) a -> Int -> m (Maybe a)

Yield the element at the given position. Returns Nothing if the index is out of range.

Examples

Expand

>>> import qualified Data.Vector.Mutable as MV
>>> v <- MV.generate 10 (\x -> x*x)
>>> MV.readMaybe v 3
Just 9
>>> MV.readMaybe v 13
Nothing

Since: 0.13

write :: PrimMonad m => MVector (PrimState m) a -> Int -> a -> m ()

Replace the element at the given position.

modify :: PrimMonad m => MVector (PrimState m) a -> (a -> a) -> Int -> m ()

Modify the element at the given position.

modifyM :: PrimMonad m => MVector (PrimState m) a -> (a -> m a) -> Int -> m ()

Modify the element at the given position using a monadic function.

Since: 0.12.3.0

swap :: PrimMonad m => MVector (PrimState m) a -> Int -> Int -> m ()

Swap the elements at the given positions.

exchange :: PrimMonad m => MVector (PrimState m) a -> Int -> a -> m a

Replace the element at the given position and return the old element.

unsafeRead :: PrimMonad m => MVector (PrimState m) a -> Int -> m a

Yield the element at the given position. No bounds checks are performed.

unsafeWrite :: PrimMonad m => MVector (PrimState m) a -> Int -> a -> m ()

Replace the element at the given position. No bounds checks are performed.

unsafeModify :: PrimMonad m => MVector (PrimState m) a -> (a -> a) -> Int -> m ()

Modify the element at the given position. No bounds checks are performed.

unsafeModifyM :: PrimMonad m => MVector (PrimState m) a -> (a -> m a) -> Int -> m ()

Modify the element at the given position using a monadic function. No bounds checks are performed.

Since: 0.12.3.0

unsafeSwap :: PrimMonad m => MVector (PrimState m) a -> Int -> Int -> m ()

Swap the elements at the given positions. No bounds checks are performed.

unsafeExchange :: PrimMonad m => MVector (PrimState m) a -> Int -> a -> m a

Replace the element at the given position and return the old element. No bounds checks are performed.

Folds

mapM_ :: PrimMonad m => (a -> m b) -> MVector (PrimState m) a -> m ()

O(n) Apply the monadic action to every element of the vector, discarding the results.

Since: 0.12.3.0

imapM_ :: PrimMonad m => (Int -> a -> m b) -> MVector (PrimState m) a -> m ()

O(n) Apply the monadic action to every element of the vector and its index, discarding the results.

Since: 0.12.3.0

forM_ :: PrimMonad m => MVector (PrimState m) a -> (a -> m b) -> m ()

O(n) Apply the monadic action to every element of the vector, discarding the results. It's the same as flip mapM_.

Since: 0.12.3.0

iforM_ :: PrimMonad m => MVector (PrimState m) a -> (Int -> a -> m b) -> m ()

O(n) Apply the monadic action to every element of the vector and its index, discarding the results. It's the same as flip imapM_.

Since: 0.12.3.0

foldl :: PrimMonad m => (b -> a -> b) -> b -> MVector (PrimState m) a -> m b

O(n) Pure left fold.

Since: 0.12.3.0

foldl' :: PrimMonad m => (b -> a -> b) -> b -> MVector (PrimState m) a -> m b

O(n) Pure left fold with strict accumulator.

Since: 0.12.3.0

foldM :: PrimMonad m => (b -> a -> m b) -> b -> MVector (PrimState m) a -> m b

O(n) Monadic fold.

Since: 0.12.3.0

foldM' :: PrimMonad m => (b -> a -> m b) -> b -> MVector (PrimState m) a -> m b

O(n) Monadic fold with strict accumulator.

Since: 0.12.3.0

foldr :: PrimMonad m => (a -> b -> b) -> b -> MVector (PrimState m) a -> m b

O(n) Pure right fold.

Since: 0.12.3.0

foldr' :: PrimMonad m => (a -> b -> b) -> b -> MVector (PrimState m) a -> m b

O(n) Pure right fold with strict accumulator.

Since: 0.12.3.0

foldrM :: PrimMonad m => (a -> b -> m b) -> b -> MVector (PrimState m) a -> m b

O(n) Monadic right fold.

Since: 0.12.3.0

foldrM' :: PrimMonad m => (a -> b -> m b) -> b -> MVector (PrimState m) a -> m b

O(n) Monadic right fold with strict accumulator.

Since: 0.12.3.0

ifoldl :: PrimMonad m => (b -> Int -> a -> b) -> b -> MVector (PrimState m) a -> m b

O(n) Pure left fold using a function applied to each element and its index.

Since: 0.12.3.0

ifoldl' :: PrimMonad m => (b -> Int -> a -> b) -> b -> MVector (PrimState m) a -> m b

O(n) Pure left fold with strict accumulator using a function applied to each element and its index.

Since: 0.12.3.0

ifoldM :: PrimMonad m => (b -> Int -> a -> m b) -> b -> MVector (PrimState m) a -> m b

O(n) Monadic fold using a function applied to each element and its index.

Since: 0.12.3.0

ifoldM' :: PrimMonad m => (b -> Int -> a -> m b) -> b -> MVector (PrimState m) a -> m b

O(n) Monadic fold with strict accumulator using a function applied to each element and its index.

Since: 0.12.3.0

ifoldr :: PrimMonad m => (Int -> a -> b -> b) -> b -> MVector (PrimState m) a -> m b

O(n) Pure right fold using a function applied to each element and its index.

Since: 0.12.3.0

ifoldr' :: PrimMonad m => (Int -> a -> b -> b) -> b -> MVector (PrimState m) a -> m b

O(n) Pure right fold with strict accumulator using a function applied to each element and its index.

Since: 0.12.3.0

ifoldrM :: PrimMonad m => (Int -> a -> b -> m b) -> b -> MVector (PrimState m) a -> m b

O(n) Monadic right fold using a function applied to each element and its index.

Since: 0.12.3.0

ifoldrM' :: PrimMonad m => (Int -> a -> b -> m b) -> b -> MVector (PrimState m) a -> m b

O(n) Monadic right fold with strict accumulator using a function applied to each element and its index.

Since: 0.12.3.0

Modifying vectors

nextPermutation :: (PrimMonad m, Ord e) => MVector (PrimState m) e -> m Bool

Compute the (lexicographically) next permutation of the given vector in-place. Returns False when the input is the last permutation.

Filling and copying

set :: PrimMonad m => MVector (PrimState m) a -> a -> m ()

Set all elements of the vector to the given value.

copy

Arguments

:: PrimMonad m

=> MVector (PrimState m) a
target

-> MVector (PrimState m) a
source

-> m ()

Copy a vector. The two vectors must have the same length and may not overlap.

move

Arguments

:: PrimMonad m

=> MVector (PrimState m) a
target

-> MVector (PrimState m) a
source

-> m ()

Move the contents of a vector. The two vectors must have the same length.

If the vectors do not overlap, then this is equivalent to copy. Otherwise, the copying is performed as if the source vector were copied to a temporary vector and then the temporary vector was copied to the target vector.

unsafeCopy

Arguments

:: PrimMonad m

=> MVector (PrimState m) a
target

-> MVector (PrimState m) a
source

-> m ()

Copy a vector. The two vectors must have the same length and may not overlap, but this is not checked.

unsafeMove

Arguments

:: PrimMonad m

=> MVector (PrimState m) a
target

-> MVector (PrimState m) a
source

-> m ()

Move the contents of a vector. The two vectors must have the same length, but this is not checked.

If the vectors do not overlap, then this is equivalent to unsafeCopy. Otherwise, the copying is performed as if the source vector were copied to a temporary vector and then the temporary vector was copied to the target vector.

Arrays

fromMutableArray :: PrimMonad m => MutableArray (PrimState m) a -> m (MVector (PrimState m) a)

O(n) Make a copy of a mutable array to a new mutable vector.

Since: 0.12.2.0

toMutableArray :: PrimMonad m => MVector (PrimState m) a -> m (MutableArray (PrimState m) a)

O(n) Make a copy of a mutable vector into a new mutable array.

Since: 0.12.2.0

Re-exports

class Monad m => PrimMonad (m :: Type -> Type)

Class of monads which can perform primitive state-transformer actions.

Minimal complete definition

primitive

type family PrimState (m :: Type -> Type)

State token type.

data RealWorld

RealWorld is deeply magical. It is primitive, but it is not unlifted (hence ptrArg). We never manipulate values of type RealWorld; it's only used in the type system, to parameterise State#.