Data.Vector.Primitive.Mutable

Description

Mutable primitive vectors.

Mutable vectors of primitive types

data MVector s a

Mutable vectors of primitive types.

Constructors

MVector

Fields

!Int

offset

!Int

length

!(MutableByteArray s)

underlying mutable byte array

type IOVector = MVector RealWorld

type STVector s = MVector s

Accessors

Length information

length :: Prim a => MVector s a -> Int

Length of the mutable vector.

null :: Prim a => MVector s a -> Bool

Check whether the vector is empty.

Extracting subvectors

slice

Arguments

:: Prim a

=> Int
i starting index

-> Int
n length

-> MVector s a

-> MVector s a

Yield a part of the mutable vector without copying it. The vector must contain at least i+n elements.

init :: Prim a => MVector s a -> MVector s a

Drop the last element of the mutable vector without making a copy. If the vector is empty, an exception is thrown.

tail :: Prim a => MVector s a -> MVector s a

Drop the first element of the mutable vector without making a copy. If the vector is empty, an exception is thrown.

take :: Prim a => Int -> MVector s a -> MVector s a

Take the n first elements of the mutable vector without making a copy. For negative n, the empty vector is returned. If n is larger than the vector's length, the vector is returned unchanged.

drop :: Prim a => Int -> MVector s a -> MVector s a

Drop the n first element of the mutable vector without making a copy. For negative n, the vector is returned unchanged. If n is larger than the vector's length, the empty vector is returned.

splitAt :: Prim a => Int -> MVector s a -> (MVector s a, MVector s a)

O(1) Split the mutable vector into the first n elements and the remainder, without copying.

Note that splitAt n v is equivalent to (take n v, drop n v), but slightly more efficient.

unsafeSlice

Arguments

:: Prim a

=> Int
starting index

-> Int
length of the slice

-> MVector s a

-> MVector s a

Yield a part of the mutable vector without copying it. No bounds checks are performed.

unsafeInit :: Prim a => MVector s a -> MVector s a

Same as init, but doesn't do range checks.

unsafeTail :: Prim a => MVector s a -> MVector s a

Same as tail, but doesn't do range checks.

unsafeTake :: Prim a => Int -> MVector s a -> MVector s a

Unsafe variant of take. If n is out of range, it will simply create an invalid slice that likely violate memory safety.

unsafeDrop :: Prim a => Int -> MVector s a -> MVector s a

Unsafe variant of drop. If n is out of range, it will simply create an invalid slice that likely violate memory safety.

Overlapping

overlaps :: Prim a => MVector s a -> MVector s a -> Bool

Check whether two vectors overlap.

Construction

Initialisation

new :: (PrimMonad m, Prim a) => Int -> m (MVector (PrimState m) a)

Create a mutable vector of the given length.

unsafeNew :: (PrimMonad m, Prim a) => Int -> m (MVector (PrimState m) a)

Create a mutable vector of the given length. The vector content is uninitialized, which means it is filled with whatever the underlying memory buffer happens to contain.

Since: 0.5

replicate :: (PrimMonad m, Prim a) => Int -> a -> m (MVector (PrimState m) a)

Create a mutable vector of the given length (0 if the length is negative) and fill it with an initial value.

replicateM :: (PrimMonad m, Prim a) => Int -> m a -> m (MVector (PrimState m) a)

Create a mutable vector of the given length (0 if the length is negative) and fill it with values produced by repeatedly executing the monadic action.

generate :: (PrimMonad m, Prim a) => Int -> (Int -> a) -> m (MVector (PrimState m) a)

O(n) Create a mutable vector of the given length (0 if the length is negative) and fill it with the results of applying the function to each index. Iteration starts at index 0.

Since: 0.12.3.0

generateM :: (PrimMonad m, Prim a) => Int -> (Int -> m a) -> m (MVector (PrimState m) a)

O(n) Create a mutable vector of the given length (0 if the length is negative) and fill it with the results of applying the monadic function to each index. Iteration starts at index 0.

Since: 0.12.3.0

clone :: (PrimMonad m, Prim a) => MVector (PrimState m) a -> m (MVector (PrimState m) a)

Create a copy of a mutable vector.

Growing

grow :: (PrimMonad m, Prim a) => MVector (PrimState m) a -> Int -> m (MVector (PrimState m) a)

Grow a primitive vector by the given number of elements. The number must be non-negative. This has the same semantics as grow for generic vectors.

Examples

Expand

>>> import qualified Data.Vector.Primitive as VP
>>> import qualified Data.Vector.Primitive.Mutable as MVP
>>> mv <- VP.thaw $ VP.fromList ([10, 20, 30] :: [Int])
>>> mv' <- MVP.grow mv 2

Extra memory at the end of the newly allocated vector is initialized to 0 bytes, which for Prim instances will usually correspond to some default value for a particular type, e.g. 0 for Int, NUL for Char, etc. However, if unsafeGrow was used instead, this would not have been guaranteed and some garbage would be there instead.

>>> VP.freeze mv'
[10,20,30,0,0]

Having the extra space we can write new values in there:

>>> MVP.write mv' 3 999
>>> VP.freeze mv'
[10,20,30,999,0]

It is important to note that the source mutable vector is not affected when the newly allocated one is mutated.

>>> MVP.write mv' 2 888
>>> VP.freeze mv'
[10,20,888,999,0]
>>> VP.freeze mv
[10,20,30]

Since: 0.5

unsafeGrow :: (PrimMonad m, Prim a) => MVector (PrimState m) a -> Int -> m (MVector (PrimState m) a)

Grow a vector by the given number of elements. The number must be non-negative, but this is not checked. This has the same semantics as unsafeGrow for generic vectors.

Since: 0.5

Restricting memory usage

clear :: (PrimMonad m, Prim a) => MVector (PrimState m) a -> m ()

Reset all elements of the vector to some undefined value, clearing all references to external objects. This is a noop.

Accessing individual elements

read :: (PrimMonad m, Prim a) => MVector (PrimState m) a -> Int -> m a

Yield the element at the given position. Will throw an exception if the index is out of range.

Examples

Expand

>>> import qualified Data.Vector.Primitive.Mutable as MVP
>>> v <- MVP.generate 10 (\x -> x*x)
>>> MVP.read v 3
9

readMaybe :: (PrimMonad m, Prim a) => MVector (PrimState m) a -> Int -> m (Maybe a)

Yield the element at the given position. Returns Nothing if the index is out of range.

Examples

Expand

>>> import qualified Data.Vector.Primitive.Mutable as MVP
>>> v <- MVP.generate 10 (\x -> x*x)
>>> MVP.readMaybe v 3
Just 9
>>> MVP.readMaybe v 13
Nothing

Since: 0.13

write :: (PrimMonad m, Prim a) => MVector (PrimState m) a -> Int -> a -> m ()

Replace the element at the given position.

modify :: (PrimMonad m, Prim a) => MVector (PrimState m) a -> (a -> a) -> Int -> m ()

Modify the element at the given position.

modifyM :: (PrimMonad m, Prim a) => MVector (PrimState m) a -> (a -> m a) -> Int -> m ()

Modify the element at the given position using a monadic function.

Since: 0.12.3.0

swap :: (PrimMonad m, Prim a) => MVector (PrimState m) a -> Int -> Int -> m ()

Swap the elements at the given positions.

exchange :: (PrimMonad m, Prim a) => MVector (PrimState m) a -> Int -> a -> m a

Replace the element at the given position and return the old element.

unsafeRead :: (PrimMonad m, Prim a) => MVector (PrimState m) a -> Int -> m a

Yield the element at the given position. No bounds checks are performed.

unsafeWrite :: (PrimMonad m, Prim a) => MVector (PrimState m) a -> Int -> a -> m ()

Replace the element at the given position. No bounds checks are performed.

unsafeModify :: (PrimMonad m, Prim a) => MVector (PrimState m) a -> (a -> a) -> Int -> m ()

Modify the element at the given position. No bounds checks are performed.

unsafeModifyM :: (PrimMonad m, Prim a) => MVector (PrimState m) a -> (a -> m a) -> Int -> m ()

Modify the element at the given position using a monadic function. No bounds checks are performed.

Since: 0.12.3.0

unsafeSwap :: (PrimMonad m, Prim a) => MVector (PrimState m) a -> Int -> Int -> m ()

Swap the elements at the given positions. No bounds checks are performed.

unsafeExchange :: (PrimMonad m, Prim a) => MVector (PrimState m) a -> Int -> a -> m a

Replace the element at the given position and return the old element. No bounds checks are performed.

Folds

mapM_ :: (PrimMonad m, Prim a) => (a -> m b) -> MVector (PrimState m) a -> m ()

O(n) Apply the monadic action to every element of the vector, discarding the results.

Since: 0.12.3.0

imapM_ :: (PrimMonad m, Prim a) => (Int -> a -> m b) -> MVector (PrimState m) a -> m ()

O(n) Apply the monadic action to every element of the vector and its index, discarding the results.

Since: 0.12.3.0

forM_ :: (PrimMonad m, Prim a) => MVector (PrimState m) a -> (a -> m b) -> m ()

O(n) Apply the monadic action to every element of the vector, discarding the results. It's the same as flip mapM_.

Since: 0.12.3.0

iforM_ :: (PrimMonad m, Prim a) => MVector (PrimState m) a -> (Int -> a -> m b) -> m ()

O(n) Apply the monadic action to every element of the vector and its index, discarding the results. It's the same as flip imapM_.

Since: 0.12.3.0

foldl :: (PrimMonad m, Prim a) => (b -> a -> b) -> b -> MVector (PrimState m) a -> m b

O(n) Pure left fold.

Since: 0.12.3.0

foldl' :: (PrimMonad m, Prim a) => (b -> a -> b) -> b -> MVector (PrimState m) a -> m b

O(n) Pure left fold with strict accumulator.

Since: 0.12.3.0

foldM :: (PrimMonad m, Prim a) => (b -> a -> m b) -> b -> MVector (PrimState m) a -> m b

O(n) Monadic fold.

Since: 0.12.3.0

foldM' :: (PrimMonad m, Prim a) => (b -> a -> m b) -> b -> MVector (PrimState m) a -> m b

O(n) Monadic fold with strict accumulator.

Since: 0.12.3.0

foldr :: (PrimMonad m, Prim a) => (a -> b -> b) -> b -> MVector (PrimState m) a -> m b

O(n) Pure right fold.

Since: 0.12.3.0

foldr' :: (PrimMonad m, Prim a) => (a -> b -> b) -> b -> MVector (PrimState m) a -> m b

O(n) Pure right fold with strict accumulator.

Since: 0.12.3.0

foldrM :: (PrimMonad m, Prim a) => (a -> b -> m b) -> b -> MVector (PrimState m) a -> m b

O(n) Monadic right fold.

Since: 0.12.3.0

foldrM' :: (PrimMonad m, Prim a) => (a -> b -> m b) -> b -> MVector (PrimState m) a -> m b

O(n) Monadic right fold with strict accumulator.

Since: 0.12.3.0

ifoldl :: (PrimMonad m, Prim a) => (b -> Int -> a -> b) -> b -> MVector (PrimState m) a -> m b

O(n) Pure left fold using a function applied to each element and its index.

Since: 0.12.3.0

ifoldl' :: (PrimMonad m, Prim a) => (b -> Int -> a -> b) -> b -> MVector (PrimState m) a -> m b

O(n) Pure left fold with strict accumulator using a function applied to each element and its index.

Since: 0.12.3.0

ifoldM :: (PrimMonad m, Prim a) => (b -> Int -> a -> m b) -> b -> MVector (PrimState m) a -> m b

O(n) Monadic fold using a function applied to each element and its index.

Since: 0.12.3.0

ifoldM' :: (PrimMonad m, Prim a) => (b -> Int -> a -> m b) -> b -> MVector (PrimState m) a -> m b

O(n) Monadic fold with strict accumulator using a function applied to each element and its index.

Since: 0.12.3.0

ifoldr :: (PrimMonad m, Prim a) => (Int -> a -> b -> b) -> b -> MVector (PrimState m) a -> m b

O(n) Pure right fold using a function applied to each element and its index.

Since: 0.12.3.0

ifoldr' :: (PrimMonad m, Prim a) => (Int -> a -> b -> b) -> b -> MVector (PrimState m) a -> m b

O(n) Pure right fold with strict accumulator using a function applied to each element and its index.

Since: 0.12.3.0

ifoldrM :: (PrimMonad m, Prim a) => (Int -> a -> b -> m b) -> b -> MVector (PrimState m) a -> m b

O(n) Monadic right fold using a function applied to each element and its index.

Since: 0.12.3.0

ifoldrM' :: (PrimMonad m, Prim a) => (Int -> a -> b -> m b) -> b -> MVector (PrimState m) a -> m b

O(n) Monadic right fold with strict accumulator using a function applied to each element and its index.

Since: 0.12.3.0

Modifying vectors

nextPermutation :: (PrimMonad m, Ord e, Prim e) => MVector (PrimState m) e -> m Bool

Compute the (lexicographically) next permutation of the given vector in-place. Returns False when the input is the last permutation.

Filling and copying

set :: (PrimMonad m, Prim a) => MVector (PrimState m) a -> a -> m ()

Set all elements of the vector to the given value.

copy

Arguments

:: (PrimMonad m, Prim a)

=> MVector (PrimState m) a
target

-> MVector (PrimState m) a
source

-> m ()

Copy a vector. The two vectors must have the same length and may not overlap.

move

Arguments

:: (PrimMonad m, Prim a)

=> MVector (PrimState m) a
target

-> MVector (PrimState m) a
source

-> m ()

Move the contents of a vector. The two vectors must have the same length.

If the vectors do not overlap, then this is equivalent to copy. Otherwise, the copying is performed as if the source vector were copied to a temporary vector and then the temporary vector was copied to the target vector.

unsafeCopy

Arguments

:: (PrimMonad m, Prim a)

=> MVector (PrimState m) a
target

-> MVector (PrimState m) a
source

-> m ()

Copy a vector. The two vectors must have the same length and may not overlap, but this is not checked.

unsafeMove

Arguments

:: (PrimMonad m, Prim a)

=> MVector (PrimState m) a
target

-> MVector (PrimState m) a
source

-> m ()

Move the contents of a vector. The two vectors must have the same length, but this is not checked.

If the vectors do not overlap, then this is equivalent to unsafeCopy. Otherwise, the copying is performed as if the source vector were copied to a temporary vector and then the temporary vector was copied to the target vector.

Unsafe conversions

unsafeCoerceMVector :: Coercible a b => MVector s a -> MVector s b

O(1) Unsafely coerce a mutable vector from one element type to another, representationally equal type. The operation just changes the type of the underlying pointer and does not modify the elements.

Note that this function is unsafe. The Coercible constraint guarantees that the element types are representationally equal. It however cannot guarantee that their respective Prim instances are compatible.

unsafeCast :: forall a b s. (HasCallStack, Prim a, Prim b) => MVector s a -> MVector s b

O(1) Unsafely cast a vector from one element type to another. This operation just changes the type of the vector and does not modify the elements.

This function will throw an error if elements are of mismatching sizes.

| @since 0.13.0.0

Re-exports

class Prim a

Class of types supporting primitive array operations. This includes interfacing with GC-managed memory (functions suffixed with ByteArray#) and interfacing with unmanaged memory (functions suffixed with Addr#). Endianness is platform-dependent.

Minimal complete definition

sizeOf#, alignment#, indexByteArray#, readByteArray#, writeByteArray#, setByteArray#, indexOffAddr#, readOffAddr#, writeOffAddr#, setOffAddr#

class Monad m => PrimMonad (m :: Type -> Type)

Class of monads which can perform primitive state-transformer actions.

Minimal complete definition

primitive

type family PrimState (m :: Type -> Type)

State token type.

data RealWorld

RealWorld is deeply magical. It is primitive, but it is not unlifted (hence ptrArg). We never manipulate values of type RealWorld; it's only used in the type system, to parameterise State#.