Data.Vector.Storable

Description

Storable-based vectors.

Storable vectors

data Vector a

Storable-based vectors.

data MVector s a

Mutable Storable-based vectors.

Constructors

MVector !Int !(ForeignPtr a)

Accessors

Length information

length :: Storable a => Vector a -> Int

O(1) Yield the length of the vector.

null :: Storable a => Vector a -> Bool

O(1) Test whether a vector is empty.

Indexing

(!) :: Storable a => Vector a -> Int -> a

O(1) Indexing.

(!?) :: Storable a => Vector a -> Int -> Maybe a

O(1) Safe indexing.

head :: Storable a => Vector a -> a

O(1) First element.

last :: Storable a => Vector a -> a

O(1) Last element.

unsafeIndex :: Storable a => Vector a -> Int -> a

O(1) Unsafe indexing without bounds checking.

unsafeHead :: Storable a => Vector a -> a

O(1) First element, without checking if the vector is empty.

unsafeLast :: Storable a => Vector a -> a

O(1) Last element, without checking if the vector is empty.

Monadic indexing

indexM :: (Storable a, Monad m) => Vector a -> Int -> m a

O(1) Indexing in a monad.

The monad allows operations to be strict in the vector when necessary. Suppose vector copying is implemented like this:

copy mv v = ... write mv i (v ! i) ...

For lazy vectors, v ! i would not be evaluated which means that mv would unnecessarily retain a reference to v in each element written.

With indexM, copying can be implemented like this instead:

copy mv v = ... do
                  x <- indexM v i
                  write mv i x

Here, no references to v are retained because indexing (but not the element) is evaluated eagerly.

headM :: (Storable a, Monad m) => Vector a -> m a

O(1) First element of a vector in a monad. See indexM for an explanation of why this is useful.

lastM :: (Storable a, Monad m) => Vector a -> m a

O(1) Last element of a vector in a monad. See indexM for an explanation of why this is useful.

unsafeIndexM :: (Storable a, Monad m) => Vector a -> Int -> m a

O(1) Indexing in a monad, without bounds checks. See indexM for an explanation of why this is useful.

unsafeHeadM :: (Storable a, Monad m) => Vector a -> m a

O(1) First element in a monad, without checking for empty vectors. See indexM for an explanation of why this is useful.

unsafeLastM :: (Storable a, Monad m) => Vector a -> m a

O(1) Last element in a monad, without checking for empty vectors. See indexM for an explanation of why this is useful.

Extracting subvectors (slicing)

slice

Arguments

:: Storable a

=> Int
i starting index

-> Int
n length

-> Vector a

-> Vector a

O(1) Yield a slice of the vector without copying it. The vector must contain at least i+n elements.

init :: Storable a => Vector a -> Vector a

O(1) Yield all but the last element without copying. The vector may not be empty.

tail :: Storable a => Vector a -> Vector a

O(1) Yield all but the first element without copying. The vector may not be empty.

take :: Storable a => Int -> Vector a -> Vector a

O(1) Yield at the first n elements without copying. The vector may contain less than n elements, in which case it is returned unchanged.

drop :: Storable a => Int -> Vector a -> Vector a

O(1) Yield all but the first n elements without copying. The vector may contain less than n elements, in which case an empty vector is returned.

splitAt :: Storable a => Int -> Vector a -> (Vector a, Vector a)

O(1) Yield the first n elements paired with the remainder, without copying.

Note that splitAt n v is equivalent to (take n v, drop n v), but slightly more efficient.

Since: 0.7.1

uncons :: Storable a => Vector a -> Maybe (a, Vector a)

O(1) Yield the head and tail of the vector, or Nothing if the vector is empty.

Since: 0.12.2.0

unsnoc :: Storable a => Vector a -> Maybe (Vector a, a)

O(1) Yield the last and init of the vector, or Nothing if the vector is empty.

Since: 0.12.2.0

unsafeSlice

Arguments

:: Storable a

=> Int
i starting index

-> Int
n length

-> Vector a

-> Vector a

O(1) Yield a slice of the vector without copying. The vector must contain at least i+n elements, but this is not checked.

unsafeInit :: Storable a => Vector a -> Vector a

O(1) Yield all but the last element without copying. The vector may not be empty, but this is not checked.

unsafeTail :: Storable a => Vector a -> Vector a

O(1) Yield all but the first element without copying. The vector may not be empty, but this is not checked.

unsafeTake :: Storable a => Int -> Vector a -> Vector a

O(1) Yield the first n elements without copying. The vector must contain at least n elements, but this is not checked.

unsafeDrop :: Storable a => Int -> Vector a -> Vector a

O(1) Yield all but the first n elements without copying. The vector must contain at least n elements, but this is not checked.

Construction

Initialisation

empty :: Storable a => Vector a

O(1) The empty vector.

singleton :: Storable a => a -> Vector a

O(1) A vector with exactly one element.

replicate :: Storable a => Int -> a -> Vector a

O(n) A vector of the given length with the same value in each position.

generate :: Storable a => Int -> (Int -> a) -> Vector a

O(n) Construct a vector of the given length by applying the function to each index.

iterateN :: Storable a => Int -> (a -> a) -> a -> Vector a

O(n) Apply the function \(\max(n - 1, 0)\) times to an initial value, producing a vector of length \(\max(n, 0)\). The 0th element will contain the initial value, which is why there is one less function application than the number of elements in the produced vector.

\( \underbrace{x, f (x), f (f (x)), \ldots}_{\max(0,n)\rm{~elements}} \)

Examples

Expand

>>> import qualified Data.Vector.Storable as VS
>>> VS.iterateN 0 undefined undefined :: VS.Vector Int
[]
>>> VS.iterateN 26 succ 'a'
"abcdefghijklmnopqrstuvwxyz"

Since: 0.7.1

Monadic initialisation

replicateM :: (Monad m, Storable a) => Int -> m a -> m (Vector a)

O(n) Execute the monadic action the given number of times and store the results in a vector.

generateM :: (Monad m, Storable a) => Int -> (Int -> m a) -> m (Vector a)

O(n) Construct a vector of the given length by applying the monadic action to each index.

iterateNM :: (Monad m, Storable a) => Int -> (a -> m a) -> a -> m (Vector a)

O(n) Apply the monadic function \(\max(n - 1, 0)\) times to an initial value, producing a vector of length \(\max(n, 0)\). The 0th element will contain the initial value, which is why there is one less function application than the number of elements in the produced vector.

For a non-monadic version, see iterateN.

Since: 0.12.0.0

create :: Storable a => (forall s. ST s (MVector s a)) -> Vector a

Execute the monadic action and freeze the resulting vector.

create (do { v <- new 2; write v 0 'a'; write v 1 'b'; return v }) = <a,b>

createT :: (Traversable f, Storable a) => (forall s. ST s (f (MVector s a))) -> f (Vector a)

Execute the monadic action and freeze the resulting vectors.

Unfolding

unfoldr :: Storable a => (b -> Maybe (a, b)) -> b -> Vector a

O(n) Construct a vector by repeatedly applying the generator function to a seed. The generator function yields Just the next element and the new seed or Nothing if there are no more elements.

unfoldr (\n -> if n == 0 then Nothing else Just (n,n-1)) 10
 = <10,9,8,7,6,5,4,3,2,1>

unfoldrN :: Storable a => Int -> (b -> Maybe (a, b)) -> b -> Vector a

O(n) Construct a vector with at most n elements by repeatedly applying the generator function to a seed. The generator function yields Just the next element and the new seed or Nothing if there are no more elements.

unfoldrN 3 (\n -> Just (n,n-1)) 10 = <10,9,8>

unfoldrExactN :: Storable a => Int -> (b -> (a, b)) -> b -> Vector a

O(n) Construct a vector with exactly n elements by repeatedly applying the generator function to a seed. The generator function yields the next element and the new seed.

unfoldrExactN 3 (\n -> (n,n-1)) 10 = <10,9,8>

Since: 0.12.2.0

unfoldrM :: (Monad m, Storable a) => (b -> m (Maybe (a, b))) -> b -> m (Vector a)

O(n) Construct a vector by repeatedly applying the monadic generator function to a seed. The generator function yields Just the next element and the new seed or Nothing if there are no more elements.

unfoldrNM :: (Monad m, Storable a) => Int -> (b -> m (Maybe (a, b))) -> b -> m (Vector a)

O(n) Construct a vector by repeatedly applying the monadic generator function to a seed. The generator function yields Just the next element and the new seed or Nothing if there are no more elements.

unfoldrExactNM :: (Monad m, Storable a) => Int -> (b -> m (a, b)) -> b -> m (Vector a)

O(n) Construct a vector with exactly n elements by repeatedly applying the monadic generator function to a seed. The generator function yields the next element and the new seed.

Since: 0.12.2.0

constructN :: Storable a => Int -> (Vector a -> a) -> Vector a

O(n) Construct a vector with n elements by repeatedly applying the generator function to the already constructed part of the vector.

constructN 3 f = let a = f <> ; b = f <a> ; c = f <a,b> in <a,b,c>

constructrN :: Storable a => Int -> (Vector a -> a) -> Vector a

O(n) Construct a vector with n elements from right to left by repeatedly applying the generator function to the already constructed part of the vector.

constructrN 3 f = let a = f <> ; b = f<a> ; c = f <b,a> in <c,b,a>

Enumeration

enumFromN :: (Storable a, Num a) => a -> Int -> Vector a

O(n) Yield a vector of the given length, containing the values x, x+1 etc. This operation is usually more efficient than enumFromTo.

enumFromN 5 3 = <5,6,7>

enumFromStepN :: (Storable a, Num a) => a -> a -> Int -> Vector a

O(n) Yield a vector of the given length, containing the values x, x+y, x+y+y etc. This operations is usually more efficient than enumFromThenTo.

enumFromStepN 1 2 5 = <1,3,5,7,9>

enumFromTo :: (Storable a, Enum a) => a -> a -> Vector a

O(n) Enumerate values from x to y.

WARNING: This operation can be very inefficient. If possible, use enumFromN instead.

enumFromThenTo :: (Storable a, Enum a) => a -> a -> a -> Vector a

O(n) Enumerate values from x to y with a specific step z.

WARNING: This operation can be very inefficient. If possible, use enumFromStepN instead.

Concatenation

cons :: Storable a => a -> Vector a -> Vector a

O(n) Prepend an element.

snoc :: Storable a => Vector a -> a -> Vector a

O(n) Append an element.

(++) :: Storable a => Vector a -> Vector a -> Vector a infixr 5

O(m+n) Concatenate two vectors.

concat :: Storable a => [Vector a] -> Vector a

O(n) Concatenate all vectors in the list.

Restricting memory usage

force :: Storable a => Vector a -> Vector a

O(n) Yield the argument, but force it not to retain any extra memory, possibly by copying it.

This is especially useful when dealing with slices. For example:

force (slice 0 2 <huge vector>)

Here, the slice retains a reference to the huge vector. Forcing it creates a copy of just the elements that belong to the slice and allows the huge vector to be garbage collected.

Modifying vectors

Bulk updates

(//)

Arguments

:: Storable a

=> Vector a
initial vector (of length m)

-> [(Int, a)]
list of index/value pairs (of length n)

-> Vector a

O(m+n) For each pair (i,a) from the list of index/value pairs, replace the vector element at position i by a.

<5,9,2,7> // [(2,1),(0,3),(2,8)] = <3,9,8,7>

update_

Arguments

:: Storable a

=> Vector a
initial vector (of length m)

-> Vector Int
index vector (of length n1)

-> Vector a
value vector (of length n2)

-> Vector a

O(m+min(n1,n2)) For each index i from the index vector and the corresponding value a from the value vector, replace the element of the initial vector at position i by a.

update_ <5,9,2,7>  <2,0,2> <1,3,8> = <3,9,8,7>

unsafeUpd :: Storable a => Vector a -> [(Int, a)] -> Vector a

Same as (//), but without bounds checking.

unsafeUpdate_ :: Storable a => Vector a -> Vector Int -> Vector a -> Vector a

Same as update_, but without bounds checking.

Accumulations

accum

Arguments

:: Storable a

=> (a -> b -> a)
accumulating function f

-> Vector a
initial vector (of length m)

-> [(Int, b)]
list of index/value pairs (of length n)

-> Vector a

O(m+n) For each pair (i,b) from the list, replace the vector element a at position i by f a b.

Examples

Expand

>>> import qualified Data.Vector.Storable as VS
>>> VS.accum (+) (VS.fromList [1000,2000,3000 :: Int]) [(2,4),(1,6),(0,3),(1,10)]
[1003,2016,3004]

accumulate_

Arguments

:: (Storable a, Storable b)

=> (a -> b -> a)
accumulating function f

-> Vector a
initial vector (of length m)

-> Vector Int
index vector (of length n1)

-> Vector b
value vector (of length n2)

-> Vector a

O(m+min(n1,n2)) For each index i from the index vector and the corresponding value b from the the value vector, replace the element of the initial vector at position i by f a b.

accumulate_ (+) <5,9,2> <2,1,0,1> <4,6,3,7> = <5+3, 9+6+7, 2+4>

unsafeAccum :: Storable a => (a -> b -> a) -> Vector a -> [(Int, b)] -> Vector a

Same as accum, but without bounds checking.

unsafeAccumulate_ :: (Storable a, Storable b) => (a -> b -> a) -> Vector a -> Vector Int -> Vector b -> Vector a

Same as accumulate_, but without bounds checking.

Permutations

reverse :: Storable a => Vector a -> Vector a

O(n) Reverse a vector.

backpermute :: Storable a => Vector a -> Vector Int -> Vector a

O(n) Yield the vector obtained by replacing each element i of the index vector by xs!i. This is equivalent to map (xs!) is, but is often much more efficient.

backpermute <a,b,c,d> <0,3,2,3,1,0> = <a,d,c,d,b,a>

unsafeBackpermute :: Storable a => Vector a -> Vector Int -> Vector a

Same as backpermute, but without bounds checking.

Safe destructive updates

modify :: Storable a => (forall s. MVector s a -> ST s ()) -> Vector a -> Vector a

Apply a destructive operation to a vector. The operation may be performed in place if it is safe to do so and will modify a copy of the vector otherwise (see New for details).

Examples

Expand

>>> import qualified Data.Vector.Storable as VS
>>> import qualified Data.Vector.Storable.Mutable as MVS
>>> VS.modify (\v -> MVS.write v 0 'x') $ VS.replicate 4 'a'
"xaaa"

Elementwise operations

Mapping

map :: (Storable a, Storable b) => (a -> b) -> Vector a -> Vector b

O(n) Map a function over a vector.

imap :: (Storable a, Storable b) => (Int -> a -> b) -> Vector a -> Vector b

O(n) Apply a function to every element of a vector and its index.

concatMap :: (Storable a, Storable b) => (a -> Vector b) -> Vector a -> Vector b

Map a function over a vector and concatenate the results.

Monadic mapping

mapM :: (Monad m, Storable a, Storable b) => (a -> m b) -> Vector a -> m (Vector b)

O(n) Apply the monadic action to all elements of the vector, yielding a vector of results.

imapM :: (Monad m, Storable a, Storable b) => (Int -> a -> m b) -> Vector a -> m (Vector b)

O(n) Apply the monadic action to every element of a vector and its index, yielding a vector of results.

Since: 0.12.2.0

mapM_ :: (Monad m, Storable a) => (a -> m b) -> Vector a -> m ()

O(n) Apply the monadic action to all elements of a vector and ignore the results.

imapM_ :: (Monad m, Storable a) => (Int -> a -> m b) -> Vector a -> m ()

O(n) Apply the monadic action to every element of a vector and its index, ignoring the results.

Since: 0.12.2.0

forM :: (Monad m, Storable a, Storable b) => Vector a -> (a -> m b) -> m (Vector b)

O(n) Apply the monadic action to all elements of the vector, yielding a vector of results. Equivalent to flip mapM.

forM_ :: (Monad m, Storable a) => Vector a -> (a -> m b) -> m ()

O(n) Apply the monadic action to all elements of a vector and ignore the results. Equivalent to flip mapM_.

iforM :: (Monad m, Storable a, Storable b) => Vector a -> (Int -> a -> m b) -> m (Vector b)

O(n) Apply the monadic action to all elements of the vector and their indices, yielding a vector of results. Equivalent to flip imapM.

Since: 0.12.2.0

iforM_ :: (Monad m, Storable a) => Vector a -> (Int -> a -> m b) -> m ()

O(n) Apply the monadic action to all elements of the vector and their indices and ignore the results. Equivalent to flip imapM_.

Since: 0.12.2.0

Zipping

zipWith :: (Storable a, Storable b, Storable c) => (a -> b -> c) -> Vector a -> Vector b -> Vector c

O(min(m,n)) Zip two vectors with the given function.

zipWith3 :: (Storable a, Storable b, Storable c, Storable d) => (a -> b -> c -> d) -> Vector a -> Vector b -> Vector c -> Vector d

Zip three vectors with the given function.

zipWith4 :: (Storable a, Storable b, Storable c, Storable d, Storable e) => (a -> b -> c -> d -> e) -> Vector a -> Vector b -> Vector c -> Vector d -> Vector e

zipWith5 :: (Storable a, Storable b, Storable c, Storable d, Storable e, Storable f) => (a -> b -> c -> d -> e -> f) -> Vector a -> Vector b -> Vector c -> Vector d -> Vector e -> Vector f

zipWith6 :: (Storable a, Storable b, Storable c, Storable d, Storable e, Storable f, Storable g) => (a -> b -> c -> d -> e -> f -> g) -> Vector a -> Vector b -> Vector c -> Vector d -> Vector e -> Vector f -> Vector g

izipWith :: (Storable a, Storable b, Storable c) => (Int -> a -> b -> c) -> Vector a -> Vector b -> Vector c

O(min(m,n)) Zip two vectors with a function that also takes the elements' indices.

izipWith3 :: (Storable a, Storable b, Storable c, Storable d) => (Int -> a -> b -> c -> d) -> Vector a -> Vector b -> Vector c -> Vector d

Zip three vectors and their indices with the given function.

izipWith4 :: (Storable a, Storable b, Storable c, Storable d, Storable e) => (Int -> a -> b -> c -> d -> e) -> Vector a -> Vector b -> Vector c -> Vector d -> Vector e

izipWith5 :: (Storable a, Storable b, Storable c, Storable d, Storable e, Storable f) => (Int -> a -> b -> c -> d -> e -> f) -> Vector a -> Vector b -> Vector c -> Vector d -> Vector e -> Vector f

izipWith6 :: (Storable a, Storable b, Storable c, Storable d, Storable e, Storable f, Storable g) => (Int -> a -> b -> c -> d -> e -> f -> g) -> Vector a -> Vector b -> Vector c -> Vector d -> Vector e -> Vector f -> Vector g

Monadic zipping

zipWithM :: (Monad m, Storable a, Storable b, Storable c) => (a -> b -> m c) -> Vector a -> Vector b -> m (Vector c)

O(min(m,n)) Zip the two vectors with the monadic action and yield a vector of results.

izipWithM :: (Monad m, Storable a, Storable b, Storable c) => (Int -> a -> b -> m c) -> Vector a -> Vector b -> m (Vector c)

O(min(m,n)) Zip the two vectors with a monadic action that also takes the element index and yield a vector of results.

Since: 0.12.2.0

zipWithM_ :: (Monad m, Storable a, Storable b) => (a -> b -> m c) -> Vector a -> Vector b -> m ()

O(min(m,n)) Zip the two vectors with the monadic action and ignore the results.

izipWithM_ :: (Monad m, Storable a, Storable b) => (Int -> a -> b -> m c) -> Vector a -> Vector b -> m ()

O(min(m,n)) Zip the two vectors with a monadic action that also takes the element index and ignore the results.

Since: 0.12.2.0

Working with predicates

Filtering

filter :: Storable a => (a -> Bool) -> Vector a -> Vector a

O(n) Drop all elements that do not satisfy the predicate.

ifilter :: Storable a => (Int -> a -> Bool) -> Vector a -> Vector a

O(n) Drop all elements that do not satisfy the predicate which is applied to the values and their indices.

filterM :: (Monad m, Storable a) => (a -> m Bool) -> Vector a -> m (Vector a)

O(n) Drop all elements that do not satisfy the monadic predicate.

uniq :: (Storable a, Eq a) => Vector a -> Vector a

O(n) Drop repeated adjacent elements. The first element in each group is returned.

Examples

Expand

>>> import qualified Data.Vector.Storable as VS
>>> VS.uniq $ VS.fromList [1,3,3,200,3 :: Int]
[1,3,200,3]

mapMaybe :: (Storable a, Storable b) => (a -> Maybe b) -> Vector a -> Vector b

O(n) Map the values and collect the Just results.

imapMaybe :: (Storable a, Storable b) => (Int -> a -> Maybe b) -> Vector a -> Vector b

O(n) Map the indices/values and collect the Just results.

mapMaybeM :: (Monad m, Storable a, Storable b) => (a -> m (Maybe b)) -> Vector a -> m (Vector b)

O(n) Apply the monadic function to each element of the vector and discard elements returning Nothing.

Since: 0.12.2.0

imapMaybeM :: (Monad m, Storable a, Storable b) => (Int -> a -> m (Maybe b)) -> Vector a -> m (Vector b)

O(n) Apply the monadic function to each element of the vector and its index. Discard elements returning Nothing.

Since: 0.12.2.0

takeWhile :: Storable a => (a -> Bool) -> Vector a -> Vector a

O(n) Yield the longest prefix of elements satisfying the predicate. The current implementation is not copy-free, unless the result vector is fused away.

dropWhile :: Storable a => (a -> Bool) -> Vector a -> Vector a

O(n) Drop the longest prefix of elements that satisfy the predicate without copying.

Partitioning

partition :: Storable a => (a -> Bool) -> Vector a -> (Vector a, Vector a)

O(n) Split the vector in two parts, the first one containing those elements that satisfy the predicate and the second one those that don't. The relative order of the elements is preserved at the cost of a sometimes reduced performance compared to unstablePartition.

unstablePartition :: Storable a => (a -> Bool) -> Vector a -> (Vector a, Vector a)

O(n) Split the vector in two parts, the first one containing those elements that satisfy the predicate and the second one those that don't. The order of the elements is not preserved, but the operation is often faster than partition.

partitionWith :: (Storable a, Storable b, Storable c) => (a -> Either b c) -> Vector a -> (Vector b, Vector c)

O(n) Split the vector into two parts, the first one containing the Left elements and the second containing the Right elements. The relative order of the elements is preserved.

Since: 0.12.1.0

span :: Storable a => (a -> Bool) -> Vector a -> (Vector a, Vector a)

O(n) Split the vector into the longest prefix of elements that satisfy the predicate and the rest without copying.

break :: Storable a => (a -> Bool) -> Vector a -> (Vector a, Vector a)

O(n) Split the vector into the longest prefix of elements that do not satisfy the predicate and the rest without copying.

groupBy :: Storable a => (a -> a -> Bool) -> Vector a -> [Vector a]

O(n) Split a vector into a list of slices, using a predicate function.

The concatenation of this list of slices is equal to the argument vector, and each slice contains only equal elements, as determined by the equality predicate function.

Does not fuse.

>>> import qualified Data.Vector.Storable as VS
>>> import           Data.Char (isUpper)
>>> VS.groupBy (\a b -> isUpper a == isUpper b) (VS.fromList "Mississippi River")
["M","ississippi ","R","iver"]

See also groupBy, group.

Since: 0.13.0.1

group :: (Storable a, Eq a) => Vector a -> [Vector a]

O(n) Split a vector into a list of slices of the input vector.

The concatenation of this list of slices is equal to the argument vector, and each slice contains only equal elements.

Does not fuse.

This is the equivalent of 'groupBy (==)'.

>>> import qualified Data.Vector.Storable as VS
>>> VS.group (VS.fromList "Mississippi")
["M","i","ss","i","ss","i","pp","i"]

See also group.

Since: 0.13.0.1

Searching

elem :: (Storable a, Eq a) => a -> Vector a -> Bool infix 4

O(n) Check if the vector contains an element.

notElem :: (Storable a, Eq a) => a -> Vector a -> Bool infix 4

O(n) Check if the vector does not contain an element (inverse of elem).

find :: Storable a => (a -> Bool) -> Vector a -> Maybe a

O(n) Yield Just the first element matching the predicate or Nothing if no such element exists.

findIndex :: Storable a => (a -> Bool) -> Vector a -> Maybe Int

O(n) Yield Just the index of the first element matching the predicate or Nothing if no such element exists.

findIndexR :: Storable a => (a -> Bool) -> Vector a -> Maybe Int

O(n) Yield Just the index of the last element matching the predicate or Nothing if no such element exists.

Does not fuse.

findIndices :: Storable a => (a -> Bool) -> Vector a -> Vector Int

O(n) Yield the indices of elements satisfying the predicate in ascending order.

elemIndex :: (Storable a, Eq a) => a -> Vector a -> Maybe Int

O(n) Yield Just the index of the first occurrence of the given element or Nothing if the vector does not contain the element. This is a specialised version of findIndex.

elemIndices :: (Storable a, Eq a) => a -> Vector a -> Vector Int

O(n) Yield the indices of all occurrences of the given element in ascending order. This is a specialised version of findIndices.

Folding

foldl :: Storable b => (a -> b -> a) -> a -> Vector b -> a

O(n) Left fold.

foldl1 :: Storable a => (a -> a -> a) -> Vector a -> a

O(n) Left fold on non-empty vectors.

foldl' :: Storable b => (a -> b -> a) -> a -> Vector b -> a

O(n) Left fold with strict accumulator.

foldl1' :: Storable a => (a -> a -> a) -> Vector a -> a

O(n) Left fold on non-empty vectors with strict accumulator.

foldr :: Storable a => (a -> b -> b) -> b -> Vector a -> b

O(n) Right fold.

foldr1 :: Storable a => (a -> a -> a) -> Vector a -> a

O(n) Right fold on non-empty vectors.

foldr' :: Storable a => (a -> b -> b) -> b -> Vector a -> b

O(n) Right fold with a strict accumulator.

foldr1' :: Storable a => (a -> a -> a) -> Vector a -> a

O(n) Right fold on non-empty vectors with strict accumulator.

ifoldl :: Storable b => (a -> Int -> b -> a) -> a -> Vector b -> a

O(n) Left fold using a function applied to each element and its index.

ifoldl' :: Storable b => (a -> Int -> b -> a) -> a -> Vector b -> a

O(n) Left fold with strict accumulator using a function applied to each element and its index.

ifoldr :: Storable a => (Int -> a -> b -> b) -> b -> Vector a -> b

O(n) Right fold using a function applied to each element and its index.

ifoldr' :: Storable a => (Int -> a -> b -> b) -> b -> Vector a -> b

O(n) Right fold with strict accumulator using a function applied to each element and its index.

foldMap :: (Monoid m, Storable a) => (a -> m) -> Vector a -> m

O(n) Map each element of the structure to a monoid and combine the results. It uses the same implementation as the corresponding method of the Foldable type class. Note that it's implemented in terms of foldr and won't fuse with functions that traverse the vector from left to right (map, generate, etc.).

Since: 0.12.2.0

foldMap' :: (Monoid m, Storable a) => (a -> m) -> Vector a -> m

O(n) Like foldMap, but strict in the accumulator. It uses the same implementation as the corresponding method of the Foldable type class. Note that it's implemented in terms of foldl', so it fuses in most contexts.

Since: 0.12.2.0

Specialised folds

all :: Storable a => (a -> Bool) -> Vector a -> Bool

O(n) Check if all elements satisfy the predicate.

Examples

Expand

>>> import qualified Data.Vector.Storable as VS
>>> VS.all even $ VS.fromList [2, 4, 12 :: Int]
True
>>> VS.all even $ VS.fromList [2, 4, 13 :: Int]
False
>>> VS.all even (VS.empty :: VS.Vector Int)
True

any :: Storable a => (a -> Bool) -> Vector a -> Bool

O(n) Check if any element satisfies the predicate.

Examples

Expand

>>> import qualified Data.Vector.Storable as VS
>>> VS.any even $ VS.fromList [1, 3, 7 :: Int]
False
>>> VS.any even $ VS.fromList [3, 2, 13 :: Int]
True
>>> VS.any even (VS.empty :: VS.Vector Int)
False

and :: Vector Bool -> Bool

O(n) Check if all elements are True.

Examples

Expand

>>> import qualified Data.Vector.Storable as VS
>>> VS.and $ VS.fromList [True, False]
False
>>> VS.and VS.empty
True

or :: Vector Bool -> Bool

O(n) Check if any element is True.

Examples

Expand

>>> import qualified Data.Vector.Storable as VS
>>> VS.or $ VS.fromList [True, False]
True
>>> VS.or VS.empty
False

sum :: (Storable a, Num a) => Vector a -> a

O(n) Compute the sum of the elements.

Examples

Expand

>>> import qualified Data.Vector.Storable as VS
>>> VS.sum $ VS.fromList [300,20,1 :: Int]
321
>>> VS.sum (VS.empty :: VS.Vector Int)
0

product :: (Storable a, Num a) => Vector a -> a

O(n) Compute the product of the elements.

Examples

Expand

>>> import qualified Data.Vector.Storable as VS
>>> VS.product $ VS.fromList [1,2,3,4 :: Int]
24
>>> VS.product (VS.empty :: VS.Vector Int)
1

maximum :: (Storable a, Ord a) => Vector a -> a

O(n) Yield the maximum element of the vector. The vector may not be empty. In case of a tie, the first occurrence wins.

Examples

Expand

>>> import qualified Data.Vector.Storable as VS
>>> VS.maximum $ VS.fromList [2, 1 :: Int]
2

maximumBy :: Storable a => (a -> a -> Ordering) -> Vector a -> a

O(n) Yield the maximum element of the vector according to the given comparison function. The vector may not be empty. In case of a tie, the first occurrence wins. This behavior is different from maximumBy which returns the last tie.

maximumOn :: (Ord b, Storable a) => (a -> b) -> Vector a -> a

O(n) Yield the maximum element of the vector by comparing the results of a key function on each element. In case of a tie, the first occurrence wins. The vector may not be empty.

Since: 0.13.0.0

minimum :: (Storable a, Ord a) => Vector a -> a

O(n) Yield the minimum element of the vector. The vector may not be empty. In case of a tie, the first occurrence wins.

Examples

Expand

>>> import qualified Data.Vector.Storable as VS
>>> VS.minimum $ VS.fromList [2, 1 :: Int]
1

minimumBy :: Storable a => (a -> a -> Ordering) -> Vector a -> a

O(n) Yield the minimum element of the vector according to the given comparison function. The vector may not be empty. In case of a tie, the first occurrence wins.

minimumOn :: (Ord b, Storable a) => (a -> b) -> Vector a -> a

O(n) Yield the minimum element of the vector by comparing the results of a key function on each element. In case of a tie, the first occurrence wins. The vector may not be empty.

Since: 0.13.0.0

minIndex :: (Storable a, Ord a) => Vector a -> Int

O(n) Yield the index of the minimum element of the vector. The vector may not be empty.

minIndexBy :: Storable a => (a -> a -> Ordering) -> Vector a -> Int

O(n) Yield the index of the minimum element of the vector according to the given comparison function. The vector may not be empty.

maxIndex :: (Storable a, Ord a) => Vector a -> Int

O(n) Yield the index of the maximum element of the vector. The vector may not be empty.

maxIndexBy :: Storable a => (a -> a -> Ordering) -> Vector a -> Int

O(n) Yield the index of the maximum element of the vector according to the given comparison function. The vector may not be empty. In case of a tie, the first occurrence wins.

Monadic folds

foldM :: (Monad m, Storable b) => (a -> b -> m a) -> a -> Vector b -> m a

O(n) Monadic fold.

ifoldM :: (Monad m, Storable b) => (a -> Int -> b -> m a) -> a -> Vector b -> m a

O(n) Monadic fold using a function applied to each element and its index.

Since: 0.12.2.0

foldM' :: (Monad m, Storable b) => (a -> b -> m a) -> a -> Vector b -> m a

O(n) Monadic fold with strict accumulator.

ifoldM' :: (Monad m, Storable b) => (a -> Int -> b -> m a) -> a -> Vector b -> m a

O(n) Monadic fold with strict accumulator using a function applied to each element and its index.

Since: 0.12.2.0

fold1M :: (Monad m, Storable a) => (a -> a -> m a) -> Vector a -> m a

O(n) Monadic fold over non-empty vectors.

fold1M' :: (Monad m, Storable a) => (a -> a -> m a) -> Vector a -> m a

O(n) Monadic fold over non-empty vectors with strict accumulator.

foldM_ :: (Monad m, Storable b) => (a -> b -> m a) -> a -> Vector b -> m ()

O(n) Monadic fold that discards the result.

ifoldM_ :: (Monad m, Storable b) => (a -> Int -> b -> m a) -> a -> Vector b -> m ()

O(n) Monadic fold that discards the result using a function applied to each element and its index.

Since: 0.12.2.0

foldM'_ :: (Monad m, Storable b) => (a -> b -> m a) -> a -> Vector b -> m ()

O(n) Monadic fold with strict accumulator that discards the result.

ifoldM'_ :: (Monad m, Storable b) => (a -> Int -> b -> m a) -> a -> Vector b -> m ()

O(n) Monadic fold with strict accumulator that discards the result using a function applied to each element and its index.

Since: 0.12.2.0

fold1M_ :: (Monad m, Storable a) => (a -> a -> m a) -> Vector a -> m ()

O(n) Monadic fold over non-empty vectors that discards the result.

fold1M'_ :: (Monad m, Storable a) => (a -> a -> m a) -> Vector a -> m ()

O(n) Monadic fold over non-empty vectors with strict accumulator that discards the result.

Scans

prescanl :: (Storable a, Storable b) => (a -> b -> a) -> a -> Vector b -> Vector a

O(n) Left-to-right prescan.

prescanl f z = init . scanl f z

Examples

Expand

>>> import qualified Data.Vector.Storable as VS
>>> VS.prescanl (+) 0 (VS.fromList [1,2,3,4 :: Int])
[0,1,3,6]

prescanl' :: (Storable a, Storable b) => (a -> b -> a) -> a -> Vector b -> Vector a

O(n) Left-to-right prescan with strict accumulator.

postscanl :: (Storable a, Storable b) => (a -> b -> a) -> a -> Vector b -> Vector a

O(n) Left-to-right postscan.

postscanl f z = tail . scanl f z

Examples

Expand

>>> import qualified Data.Vector.Storable as VS
>>> VS.postscanl (+) 0 (VS.fromList [1,2,3,4 :: Int])
[1,3,6,10]

postscanl' :: (Storable a, Storable b) => (a -> b -> a) -> a -> Vector b -> Vector a

O(n) Left-to-right postscan with strict accumulator.

scanl :: (Storable a, Storable b) => (a -> b -> a) -> a -> Vector b -> Vector a

O(n) Left-to-right scan.

scanl f z <x1,...,xn> = <y1,...,y(n+1)>
  where y1 = z
        yi = f y(i-1) x(i-1)

Examples

Expand

>>> import qualified Data.Vector.Storable as VS
>>> VS.scanl (+) 0 (VS.fromList [1,2,3,4 :: Int])
[0,1,3,6,10]

scanl' :: (Storable a, Storable b) => (a -> b -> a) -> a -> Vector b -> Vector a

O(n) Left-to-right scan with strict accumulator.

scanl1 :: Storable a => (a -> a -> a) -> Vector a -> Vector a

O(n) Initial-value free left-to-right scan over a vector.

scanl f <x1,...,xn> = <y1,...,yn>
  where y1 = x1
        yi = f y(i-1) xi

Note: Since 0.13, application of this to an empty vector no longer results in an error; instead it produces an empty vector.

Examples

Expand

>>> import qualified Data.Vector.Storable as VS
>>> VS.scanl1 min $ VS.fromListN 5 [4,2,4,1,3 :: Int]
[4,2,2,1,1]
>>> VS.scanl1 max $ VS.fromListN 5 [1,3,2,5,4 :: Int]
[1,3,3,5,5]
>>> VS.scanl1 min (VS.empty :: VS.Vector Int)
[]

scanl1' :: Storable a => (a -> a -> a) -> Vector a -> Vector a

O(n) Initial-value free left-to-right scan over a vector with a strict accumulator.

Note: Since 0.13, application of this to an empty vector no longer results in an error; instead it produces an empty vector.

Examples

Expand

>>> import qualified Data.Vector.Storable as VS
>>> VS.scanl1' min $ VS.fromListN 5 [4,2,4,1,3 :: Int]
[4,2,2,1,1]
>>> VS.scanl1' max $ VS.fromListN 5 [1,3,2,5,4 :: Int]
[1,3,3,5,5]
>>> VS.scanl1' min (VS.empty :: VS.Vector Int)
[]

iscanl :: (Storable a, Storable b) => (Int -> a -> b -> a) -> a -> Vector b -> Vector a

O(n) Left-to-right scan over a vector with its index.

Since: 0.12.2.0

iscanl' :: (Storable a, Storable b) => (Int -> a -> b -> a) -> a -> Vector b -> Vector a

O(n) Left-to-right scan over a vector (strictly) with its index.

Since: 0.12.2.0

prescanr :: (Storable a, Storable b) => (a -> b -> b) -> b -> Vector a -> Vector b

O(n) Right-to-left prescan.

prescanr f z = reverse . prescanl (flip f) z . reverse

prescanr' :: (Storable a, Storable b) => (a -> b -> b) -> b -> Vector a -> Vector b

O(n) Right-to-left prescan with strict accumulator.

postscanr :: (Storable a, Storable b) => (a -> b -> b) -> b -> Vector a -> Vector b

O(n) Right-to-left postscan.

postscanr' :: (Storable a, Storable b) => (a -> b -> b) -> b -> Vector a -> Vector b

O(n) Right-to-left postscan with strict accumulator.

scanr :: (Storable a, Storable b) => (a -> b -> b) -> b -> Vector a -> Vector b

O(n) Right-to-left scan.

scanr' :: (Storable a, Storable b) => (a -> b -> b) -> b -> Vector a -> Vector b

O(n) Right-to-left scan with strict accumulator.

scanr1 :: Storable a => (a -> a -> a) -> Vector a -> Vector a

O(n) Right-to-left, initial-value free scan over a vector.

Note: Since 0.13, application of this to an empty vector no longer results in an error; instead it produces an empty vector.

Examples

Expand

>>> import qualified Data.Vector.Storable as VS
>>> VS.scanr1 min $ VS.fromListN 5 [3,1,4,2,4 :: Int]
[1,1,2,2,4]
>>> VS.scanr1 max $ VS.fromListN 5 [4,5,2,3,1 :: Int]
[5,5,3,3,1]
>>> VS.scanr1 min (VS.empty :: VS.Vector Int)
[]

scanr1' :: Storable a => (a -> a -> a) -> Vector a -> Vector a

O(n) Right-to-left, initial-value free scan over a vector with a strict accumulator.

Note: Since 0.13, application of this to an empty vector no longer results in an error; instead it produces an empty vector.

Examples

Expand

>>> import qualified Data.Vector.Storable as VS
>>> VS.scanr1' min $ VS.fromListN 5 [3,1,4,2,4 :: Int]
[1,1,2,2,4]
>>> VS.scanr1' max $ VS.fromListN 5 [4,5,2,3,1 :: Int]
[5,5,3,3,1]
>>> VS.scanr1' min (VS.empty :: VS.Vector Int)
[]

iscanr :: (Storable a, Storable b) => (Int -> a -> b -> b) -> b -> Vector a -> Vector b

O(n) Right-to-left scan over a vector with its index.

Since: 0.12.2.0

iscanr' :: (Storable a, Storable b) => (Int -> a -> b -> b) -> b -> Vector a -> Vector b

O(n) Right-to-left scan over a vector (strictly) with its index.

Since: 0.12.2.0

Comparisons

eqBy :: (Storable a, Storable b) => (a -> b -> Bool) -> Vector a -> Vector b -> Bool

O(n) Check if two vectors are equal using the supplied equality predicate.

Since: 0.12.2.0

cmpBy :: (Storable a, Storable b) => (a -> b -> Ordering) -> Vector a -> Vector b -> Ordering

O(n) Compare two vectors using supplied the comparison function for vector elements. Comparison works the same as for lists.

cmpBy compare == compare

Since: 0.12.2.0

Utilities

Comparisons

isSameVector :: Storable a => Vector a -> Vector a -> Bool

Checks whether two values are the same vector: they have same length and share the same buffer.

>>> import qualified Data.Vector.Storable as VS
>>> let xs = VS.fromList [0/0::Double] in VS.isSameVector xs xs
True

Conversions

Lists

toList :: Storable a => Vector a -> [a]

O(n) Convert a vector to a list.

fromList :: Storable a => [a] -> Vector a

O(n) Convert a list to a vector.

fromListN :: Storable a => Int -> [a] -> Vector a

O(n) Convert the first n elements of a list to a vector. It's expected that the supplied list will be exactly n elements long. As an optimization, this function allocates a buffer for n elements, which could be used for DoS-attacks by exhausting the memory if an attacker controls that parameter.

fromListN n xs = fromList (take n xs)

Examples

Expand

>>> import qualified Data.Vector.Storable as VS
>>> VS.fromListN 3 [1,2,3,4,5 :: Int]
[1,2,3]
>>> VS.fromListN 3 [1 :: Int]
[1]

Other vector types

convert :: (Vector v a, Vector w a) => v a -> w a

O(n) Convert between different vector types.

unsafeCast :: forall a b. (Storable a, Storable b) => Vector a -> Vector b

O(1) Unsafely cast a vector from one element type to another. This operation just changes the type of the underlying pointer and does not modify the elements.

The resulting vector contains as many elements as can fit into the underlying memory block.

unsafeCoerceVector :: Coercible a b => Vector a -> Vector b

O(1) Unsafely coerce a mutable vector from one element type to another, representationally equal type. The operation just changes the type of the underlying pointer and does not modify the elements.

This is marginally safer than unsafeCast, since this function imposes an extra Coercible constraint. This function is still not safe, however, since it cannot guarantee that the two types have memory-compatible Storable instances.

Mutable vectors

freeze :: (Storable a, PrimMonad m) => MVector (PrimState m) a -> m (Vector a)

O(n) Yield an immutable copy of the mutable vector.

thaw :: (Storable a, PrimMonad m) => Vector a -> m (MVector (PrimState m) a)

O(n) Yield a mutable copy of an immutable vector.

copy :: (Storable a, PrimMonad m) => MVector (PrimState m) a -> Vector a -> m ()

O(n) Copy an immutable vector into a mutable one. The two vectors must have the same length.

unsafeFreeze :: (Storable a, PrimMonad m) => MVector (PrimState m) a -> m (Vector a)

O(1) Unsafely convert a mutable vector to an immutable one without copying. The mutable vector may not be used after this operation.

unsafeThaw :: (Storable a, PrimMonad m) => Vector a -> m (MVector (PrimState m) a)

O(1) Unsafely convert an immutable vector to a mutable one without copying. Note that this is a very dangerous function and generally it's only safe to read from the resulting vector. In this case, the immutable vector could be used safely as well.

Problems with mutation happen because GHC has a lot of freedom to introduce sharing. As a result mutable vectors produced by unsafeThaw may or may not share the same underlying buffer. For example:

foo = do
  let vec = V.generate 10 id
  mvec <- V.unsafeThaw vec
  do_something mvec

Here GHC could lift vec outside of foo which means that all calls to do_something will use same buffer with possibly disastrous results. Whether such aliasing happens or not depends on the program in question, optimization levels, and GHC flags.

All in all, attempts to modify a vector produced by unsafeThaw fall out of domain of software engineering and into realm of black magic, dark rituals, and unspeakable horrors. The only advice that could be given is: "Don't attempt to mutate a vector produced by unsafeThaw unless you know how to prevent GHC from aliasing buffers accidentally. We don't."

unsafeCopy :: (Storable a, PrimMonad m) => MVector (PrimState m) a -> Vector a -> m ()

O(n) Copy an immutable vector into a mutable one. The two vectors must have the same length. This is not checked.

Raw pointers

unsafeFromForeignPtr

Arguments

:: Storable a

=> ForeignPtr a
pointer

-> Int
offset

-> Int
length

-> Vector a

O(1) Create a vector from a ForeignPtr with an offset and a length.

The data may not be modified through the pointer afterwards.

If your offset is 0 it is more efficient to use unsafeFromForeignPtr0.

unsafeFromForeignPtr0

Arguments

:: ForeignPtr a
pointer

-> Int
length

-> Vector a

O(1) Create a vector from a ForeignPtr and a length.

It is assumed the pointer points directly to the data (no offset). Use unsafeFromForeignPtr if you need to specify an offset.

The data may not be modified through the pointer afterwards.

unsafeToForeignPtr :: Vector a -> (ForeignPtr a, Int, Int)

O(1) Yield the underlying ForeignPtr together with the offset to the data and its length. The data may not be modified through the ForeignPtr.

unsafeToForeignPtr0 :: Vector a -> (ForeignPtr a, Int)

O(1) Yield the underlying ForeignPtr together with its length.

You can assume that the pointer points directly to the data (no offset).

The data may not be modified through the ForeignPtr.

unsafeWith :: Storable a => Vector a -> (Ptr a -> IO b) -> IO b

Pass a pointer to the vector's data to the IO action. The data may not be modified through the 'Ptr.

Re-exports

class Storable a

The member functions of this class facilitate writing values of primitive types to raw memory (which may have been allocated with the above mentioned routines) and reading values from blocks of raw memory. The class, furthermore, includes support for computing the storage requirements and alignment restrictions of storable types.

Memory addresses are represented as values of type Ptr a, for some a which is an instance of class Storable. The type argument to Ptr helps provide some valuable type safety in FFI code (you can't mix pointers of different types without an explicit cast), while helping the Haskell type system figure out which marshalling method is needed for a given pointer.

All marshalling between Haskell and a foreign language ultimately boils down to translating Haskell data structures into the binary representation of a corresponding data structure of the foreign language and vice versa. To code this marshalling in Haskell, it is necessary to manipulate primitive data types stored in unstructured memory blocks. The class Storable facilitates this manipulation on all types for which it is instantiated, which are the standard basic types of Haskell, the fixed size Int types (Int8, Int16, Int32, Int64), the fixed size Word types (Word8, Word16, Word32, Word64), StablePtr, all types from Foreign.C.Types, as well as Ptr.

Minimal complete definition

sizeOf, alignment, (peek | peekElemOff | peekByteOff), (poke | pokeElemOff | pokeByteOff)