Data.Vector.Storable.Mutable

Description

Mutable vectors based on Storable.

Mutable vectors of Storable types

data MVector s a

Mutable Storable-based vectors.

Constructors

MVector !Int !(ForeignPtr a)

type IOVector = MVector RealWorld

type STVector s = MVector s

Accessors

Length information

length :: Storable a => MVector s a -> Int

Length of the mutable vector.

null :: Storable a => MVector s a -> Bool

Check whether the vector is empty.

Extracting subvectors

slice

Arguments

:: Storable a

=> Int
i starting index

-> Int
n length

-> MVector s a

-> MVector s a

Yield a part of the mutable vector without copying it. The vector must contain at least i+n elements.

init :: Storable a => MVector s a -> MVector s a

Drop the last element of the mutable vector without making a copy. If the vector is empty, an exception is thrown.

tail :: Storable a => MVector s a -> MVector s a

Drop the first element of the mutable vector without making a copy. If the vector is empty, an exception is thrown.

take :: Storable a => Int -> MVector s a -> MVector s a

Take the n first elements of the mutable vector without making a copy. For negative n, the empty vector is returned. If n is larger than the vector's length, the vector is returned unchanged.

drop :: Storable a => Int -> MVector s a -> MVector s a

Drop the n first element of the mutable vector without making a copy. For negative n, the vector is returned unchanged. If n is larger than the vector's length, the empty vector is returned.

splitAt :: Storable a => Int -> MVector s a -> (MVector s a, MVector s a)

O(1) Split the mutable vector into the first n elements and the remainder, without copying.

Note that splitAt n v is equivalent to (take n v, drop n v), but slightly more efficient.

unsafeSlice

Arguments

:: Storable a

=> Int
starting index

-> Int
length of the slice

-> MVector s a

-> MVector s a

Yield a part of the mutable vector without copying it. No bounds checks are performed.

unsafeInit :: Storable a => MVector s a -> MVector s a

Same as init, but doesn't do range checks.

unsafeTail :: Storable a => MVector s a -> MVector s a

Same as tail, but doesn't do range checks.

unsafeTake :: Storable a => Int -> MVector s a -> MVector s a

Unsafe variant of take. If n is out of range, it will simply create an invalid slice that likely violate memory safety.

unsafeDrop :: Storable a => Int -> MVector s a -> MVector s a

Unsafe variant of drop. If n is out of range, it will simply create an invalid slice that likely violate memory safety.

Overlapping

overlaps :: Storable a => MVector s a -> MVector s a -> Bool

Check whether two vectors overlap.

Construction

Initialisation

new :: (PrimMonad m, Storable a) => Int -> m (MVector (PrimState m) a)

Create a mutable vector of the given length.

unsafeNew :: (PrimMonad m, Storable a) => Int -> m (MVector (PrimState m) a)

Create a mutable vector of the given length. The vector content is uninitialized, which means it is filled with whatever the underlying memory buffer happens to contain.

Since: 0.5

replicate :: (PrimMonad m, Storable a) => Int -> a -> m (MVector (PrimState m) a)

Create a mutable vector of the given length (0 if the length is negative) and fill it with an initial value.

replicateM :: (PrimMonad m, Storable a) => Int -> m a -> m (MVector (PrimState m) a)

Create a mutable vector of the given length (0 if the length is negative) and fill it with values produced by repeatedly executing the monadic action.

generate :: (PrimMonad m, Storable a) => Int -> (Int -> a) -> m (MVector (PrimState m) a)

O(n) Create a mutable vector of the given length (0 if the length is negative) and fill it with the results of applying the function to each index. Iteration starts at index 0.

Since: 0.12.3.0

generateM :: (PrimMonad m, Storable a) => Int -> (Int -> m a) -> m (MVector (PrimState m) a)

O(n) Create a mutable vector of the given length (0 if the length is negative) and fill it with the results of applying the monadic function to each index. Iteration starts at index 0.

Since: 0.12.3.0

clone :: (PrimMonad m, Storable a) => MVector (PrimState m) a -> m (MVector (PrimState m) a)

Create a copy of a mutable vector.

Growing

grow :: (PrimMonad m, Storable a) => MVector (PrimState m) a -> Int -> m (MVector (PrimState m) a)

Grow a storable vector by the given number of elements. The number must be non-negative. This has the same semantics as grow for generic vectors.

Examples

Expand

>>> import qualified Data.Vector.Storable as VS
>>> import qualified Data.Vector.Storable.Mutable as MVS
>>> mv <- VS.thaw $ VS.fromList ([10, 20, 30] :: [Int])
>>> mv' <- MVS.grow mv 2

Extra memory at the end of the newly allocated vector is initialized to 0 bytes, which for Storable instances will usually correspond to some default value for a particular type, e.g. 0 for Int, False for Bool, etc. However, if unsafeGrow was used instead, this would not have been guaranteed and some garbage would be there instead.

>>> VS.freeze mv'
[10,20,30,0,0]

Having the extra space we can write new values in there:

>>> MVS.write mv' 3 999
>>> VS.freeze mv'
[10,20,30,999,0]

It is important to note that the source mutable vector is not affected when the newly allocated one is mutated.

>>> MVS.write mv' 2 888
>>> VS.freeze mv'
[10,20,888,999,0]
>>> VS.freeze mv
[10,20,30]

Since: 0.5

unsafeGrow :: (PrimMonad m, Storable a) => MVector (PrimState m) a -> Int -> m (MVector (PrimState m) a)

Grow a vector by the given number of elements. The number must be non-negative, but this is not checked. This has the same semantics as unsafeGrow for generic vectors.

Since: 0.5

Restricting memory usage

clear :: (PrimMonad m, Storable a) => MVector (PrimState m) a -> m ()

Reset all elements of the vector to some undefined value, clearing all references to external objects. This is a noop.

Accessing individual elements

read :: (PrimMonad m, Storable a) => MVector (PrimState m) a -> Int -> m a

Yield the element at the given position. Will throw an exception if the index is out of range.

Examples

Expand

>>> import qualified Data.Vector.Storable.Mutable as MVS
>>> v <- MVS.generate 10 (\x -> x*x)
>>> MVS.read v 3
9

readMaybe :: (PrimMonad m, Storable a) => MVector (PrimState m) a -> Int -> m (Maybe a)

Yield the element at the given position. Returns Nothing if the index is out of range.

Examples

Expand

>>> import qualified Data.Vector.Storable.Mutable as MVS
>>> v <- MVS.generate 10 (\x -> x*x)
>>> MVS.readMaybe v 3
Just 9
>>> MVS.readMaybe v 13
Nothing

Since: 0.13

write :: (PrimMonad m, Storable a) => MVector (PrimState m) a -> Int -> a -> m ()

Replace the element at the given position.

modify :: (PrimMonad m, Storable a) => MVector (PrimState m) a -> (a -> a) -> Int -> m ()

Modify the element at the given position.

modifyM :: (PrimMonad m, Storable a) => MVector (PrimState m) a -> (a -> m a) -> Int -> m ()

Modify the element at the given position using a monadic function.

Since: 0.12.3.0

swap :: (PrimMonad m, Storable a) => MVector (PrimState m) a -> Int -> Int -> m ()

Swap the elements at the given positions.

exchange :: (PrimMonad m, Storable a) => MVector (PrimState m) a -> Int -> a -> m a

Replace the element at the given position and return the old element.

unsafeRead :: (PrimMonad m, Storable a) => MVector (PrimState m) a -> Int -> m a

Yield the element at the given position. No bounds checks are performed.

unsafeWrite :: (PrimMonad m, Storable a) => MVector (PrimState m) a -> Int -> a -> m ()

Replace the element at the given position. No bounds checks are performed.

unsafeModify :: (PrimMonad m, Storable a) => MVector (PrimState m) a -> (a -> a) -> Int -> m ()

Modify the element at the given position. No bounds checks are performed.

unsafeModifyM :: (PrimMonad m, Storable a) => MVector (PrimState m) a -> (a -> m a) -> Int -> m ()

Modify the element at the given position using a monadic function. No bounds checks are performed.

Since: 0.12.3.0

unsafeSwap :: (PrimMonad m, Storable a) => MVector (PrimState m) a -> Int -> Int -> m ()

Swap the elements at the given positions. No bounds checks are performed.

unsafeExchange :: (PrimMonad m, Storable a) => MVector (PrimState m) a -> Int -> a -> m a

Replace the element at the given position and return the old element. No bounds checks are performed.

Folds

mapM_ :: (PrimMonad m, Storable a) => (a -> m b) -> MVector (PrimState m) a -> m ()

O(n) Apply the monadic action to every element of the vector, discarding the results.

Since: 0.12.3.0

imapM_ :: (PrimMonad m, Storable a) => (Int -> a -> m b) -> MVector (PrimState m) a -> m ()

O(n) Apply the monadic action to every element of the vector and its index, discarding the results.

Since: 0.12.3.0

forM_ :: (PrimMonad m, Storable a) => MVector (PrimState m) a -> (a -> m b) -> m ()

O(n) Apply the monadic action to every element of the vector, discarding the results. It's the same as flip mapM_.

Since: 0.12.3.0

iforM_ :: (PrimMonad m, Storable a) => MVector (PrimState m) a -> (Int -> a -> m b) -> m ()

O(n) Apply the monadic action to every element of the vector and its index, discarding the results. It's the same as flip imapM_.

Since: 0.12.3.0

foldl :: (PrimMonad m, Storable a) => (b -> a -> b) -> b -> MVector (PrimState m) a -> m b

O(n) Pure left fold.

Since: 0.12.3.0

foldl' :: (PrimMonad m, Storable a) => (b -> a -> b) -> b -> MVector (PrimState m) a -> m b

O(n) Pure left fold with strict accumulator.

Since: 0.12.3.0

foldM :: (PrimMonad m, Storable a) => (b -> a -> m b) -> b -> MVector (PrimState m) a -> m b

O(n) Monadic fold.

Since: 0.12.3.0

foldM' :: (PrimMonad m, Storable a) => (b -> a -> m b) -> b -> MVector (PrimState m) a -> m b

O(n) Monadic fold with strict accumulator.

Since: 0.12.3.0

foldr :: (PrimMonad m, Storable a) => (a -> b -> b) -> b -> MVector (PrimState m) a -> m b

O(n) Pure right fold.

Since: 0.12.3.0

foldr' :: (PrimMonad m, Storable a) => (a -> b -> b) -> b -> MVector (PrimState m) a -> m b

O(n) Pure right fold with strict accumulator.

Since: 0.12.3.0

foldrM :: (PrimMonad m, Storable a) => (a -> b -> m b) -> b -> MVector (PrimState m) a -> m b

O(n) Monadic right fold.

Since: 0.12.3.0

foldrM' :: (PrimMonad m, Storable a) => (a -> b -> m b) -> b -> MVector (PrimState m) a -> m b

O(n) Monadic right fold with strict accumulator.

Since: 0.12.3.0

ifoldl :: (PrimMonad m, Storable a) => (b -> Int -> a -> b) -> b -> MVector (PrimState m) a -> m b

O(n) Pure left fold using a function applied to each element and its index.

Since: 0.12.3.0

ifoldl' :: (PrimMonad m, Storable a) => (b -> Int -> a -> b) -> b -> MVector (PrimState m) a -> m b

O(n) Pure left fold with strict accumulator using a function applied to each element and its index.

Since: 0.12.3.0

ifoldM :: (PrimMonad m, Storable a) => (b -> Int -> a -> m b) -> b -> MVector (PrimState m) a -> m b

O(n) Monadic fold using a function applied to each element and its index.

Since: 0.12.3.0

ifoldM' :: (PrimMonad m, Storable a) => (b -> Int -> a -> m b) -> b -> MVector (PrimState m) a -> m b

O(n) Monadic fold with strict accumulator using a function applied to each element and its index.

Since: 0.12.3.0

ifoldr :: (PrimMonad m, Storable a) => (Int -> a -> b -> b) -> b -> MVector (PrimState m) a -> m b

O(n) Pure right fold using a function applied to each element and its index.

Since: 0.12.3.0

ifoldr' :: (PrimMonad m, Storable a) => (Int -> a -> b -> b) -> b -> MVector (PrimState m) a -> m b

O(n) Pure right fold with strict accumulator using a function applied to each element and its index.

Since: 0.12.3.0

ifoldrM :: (PrimMonad m, Storable a) => (Int -> a -> b -> m b) -> b -> MVector (PrimState m) a -> m b

O(n) Monadic right fold using a function applied to each element and its index.

Since: 0.12.3.0

ifoldrM' :: (PrimMonad m, Storable a) => (Int -> a -> b -> m b) -> b -> MVector (PrimState m) a -> m b

O(n) Monadic right fold with strict accumulator using a function applied to each element and its index.

Since: 0.12.3.0

Modifying vectors

nextPermutation :: (PrimMonad m, Storable e, Ord e) => MVector (PrimState m) e -> m Bool

Compute the (lexicographically) next permutation of the given vector in-place. Returns False when the input is the last permutation.

Filling and copying

set :: (PrimMonad m, Storable a) => MVector (PrimState m) a -> a -> m ()

Set all elements of the vector to the given value.

copy

Arguments

:: (PrimMonad m, Storable a)

=> MVector (PrimState m) a
target

-> MVector (PrimState m) a
source

-> m ()

Copy a vector. The two vectors must have the same length and may not overlap.

move

Arguments

:: (PrimMonad m, Storable a)

=> MVector (PrimState m) a
target

-> MVector (PrimState m) a
source

-> m ()

Move the contents of a vector. The two vectors must have the same length.

If the vectors do not overlap, then this is equivalent to copy. Otherwise, the copying is performed as if the source vector were copied to a temporary vector and then the temporary vector was copied to the target vector.

unsafeCopy

Arguments

:: (PrimMonad m, Storable a)

=> MVector (PrimState m) a
target

-> MVector (PrimState m) a
source

-> m ()

Copy a vector. The two vectors must have the same length and may not overlap, but this is not checked.

unsafeMove

Arguments

:: (PrimMonad m, Storable a)

=> MVector (PrimState m) a
target

-> MVector (PrimState m) a
source

-> m ()

Move the contents of a vector. The two vectors must have the same length, but this is not checked.

If the vectors do not overlap, then this is equivalent to unsafeCopy. Otherwise, the copying is performed as if the source vector were copied to a temporary vector and then the temporary vector was copied to the target vector.

Unsafe conversions

unsafeCast :: forall a b s. (Storable a, Storable b) => MVector s a -> MVector s b

O(1) Unsafely cast a mutable vector from one element type to another. The operation just changes the type of the underlying pointer and does not modify the elements.

The resulting vector contains as many elements as can fit into the underlying memory block.

unsafeCoerceMVector :: Coercible a b => MVector s a -> MVector s b

O(1) Unsafely coerce a mutable vector from one element type to another, representationally equal type. The operation just changes the type of the underlying pointer and does not modify the elements.

This is marginally safer than unsafeCast, since this function imposes an extra Coercible constraint. This function is still not safe, however, since it cannot guarantee that the two types have memory-compatible Storable instances.

Raw pointers

unsafeFromForeignPtr

Arguments

:: Storable a

=> ForeignPtr a
pointer

-> Int
offset

-> Int
length

-> MVector s a

O(1) Create a mutable vector from a ForeignPtr with an offset and a length.

Modifying data through the ForeignPtr afterwards is unsafe if the vector could have been frozen before the modification.

If your offset is 0, it is more efficient to use unsafeFromForeignPtr0.

unsafeFromForeignPtr0

Arguments

:: ForeignPtr a
pointer

-> Int
length

-> MVector s a

O(1) Create a mutable vector from a ForeignPtr and a length.

It is assumed that the pointer points directly to the data (no offset). Use unsafeFromForeignPtr if you need to specify an offset.

Modifying data through the ForeignPtr afterwards is unsafe if the vector could have been frozen before the modification.

unsafeToForeignPtr :: MVector s a -> (ForeignPtr a, Int, Int)

O(1) Yield the underlying ForeignPtr together with the offset to the data and its length. Modifying the data through the ForeignPtr is unsafe if the vector could have been frozen before the modification.

unsafeToForeignPtr0 :: MVector s a -> (ForeignPtr a, Int)

O(1) Yield the underlying ForeignPtr together with its length.

You can assume that the pointer points directly to the data (no offset).

Modifying the data through the ForeignPtr is unsafe if the vector could have been frozen before the modification.

unsafeWith :: Storable a => IOVector a -> (Ptr a -> IO b) -> IO b

Pass a pointer to the vector's data to the IO action. Modifying data through the pointer is unsafe if the vector could have been frozen before the modification.

Re-exports

class Storable a

The member functions of this class facilitate writing values of primitive types to raw memory (which may have been allocated with the above mentioned routines) and reading values from blocks of raw memory. The class, furthermore, includes support for computing the storage requirements and alignment restrictions of storable types.

Memory addresses are represented as values of type Ptr a, for some a which is an instance of class Storable. The type argument to Ptr helps provide some valuable type safety in FFI code (you can't mix pointers of different types without an explicit cast), while helping the Haskell type system figure out which marshalling method is needed for a given pointer.

All marshalling between Haskell and a foreign language ultimately boils down to translating Haskell data structures into the binary representation of a corresponding data structure of the foreign language and vice versa. To code this marshalling in Haskell, it is necessary to manipulate primitive data types stored in unstructured memory blocks. The class Storable facilitates this manipulation on all types for which it is instantiated, which are the standard basic types of Haskell, the fixed size Int types (Int8, Int16, Int32, Int64), the fixed size Word types (Word8, Word16, Word32, Word64), StablePtr, all types from Foreign.C.Types, as well as Ptr.

Minimal complete definition

sizeOf, alignment, (peek | peekElemOff | peekByteOff), (poke | pokeElemOff | pokeByteOff)

class Monad m => PrimMonad (m :: Type -> Type)

Class of monads which can perform primitive state-transformer actions.

Minimal complete definition

primitive

type family PrimState (m :: Type -> Type)

State token type.

data RealWorld

RealWorld is deeply magical. It is primitive, but it is not unlifted (hence ptrArg). We never manipulate values of type RealWorld; it's only used in the type system, to parameterise State#.