Data.Vector.Unboxed

Description

Adaptive unboxed vectors. The implementation is based on type families and picks an efficient, specialised representation for every element type. For example, unboxed vectors of pairs are represented as pairs of unboxed vectors.

Implementing unboxed vectors for new data types can be very easy. Here is how the library does this for Complex by simply wrapping vectors of pairs.

newtype instance MVector s (Complex a) = MV_Complex (MVector s (a,a))
newtype instance Vector    (Complex a) = V_Complex  (Vector    (a,a))

instance (RealFloat a, Unbox a) => MVector MVector (Complex a) where
  {-# INLINE basicLength #-}
  basicLength (MV_Complex v) = basicLength v
  ...

instance (RealFloat a, Unbox a) => Data.Vector.Generic.Vector Vector (Complex a) where
  {-# INLINE basicLength #-}
  basicLength (V_Complex v) = Data.Vector.Generic.basicLength v
  ...

instance (RealFloat a, Unbox a) => Unbox (Complex a)

For newtypes, defining instances is easier since one could use GeneralizedNewtypeDeriving in order to derive instances for Vector and MVector, since they're very cumbersome to write by hand:

>>> :set -XTypeFamilies -XStandaloneDeriving -XMultiParamTypeClasses -XGeneralizedNewtypeDeriving
>>> 
>>> import qualified Data.Vector.Generic         as VG
>>> import qualified Data.Vector.Generic.Mutable as VGM
>>> import qualified Data.Vector.Unboxed         as VU
>>> 
>>> newtype Foo = Foo Int
>>> 
>>> newtype instance VU.MVector s Foo = MV_Int (VU.MVector s Int)
>>> newtype instance VU.Vector    Foo = V_Int  (VU.Vector    Int)
>>> deriving instance VGM.MVector VU.MVector Foo
>>> deriving instance VG.Vector   VU.Vector  Foo
>>> instance VU.Unbox Foo

Unboxed vectors

data family Vector a

data family MVector s a

class (Vector Vector a, MVector MVector a) => Unbox a

Accessors

Length information

length :: Unbox a => Vector a -> Int

O(1) Yield the length of the vector.

null :: Unbox a => Vector a -> Bool

O(1) Test whether a vector is empty.

Indexing

(!) :: Unbox a => Vector a -> Int -> a

O(1) Indexing.

(!?) :: Unbox a => Vector a -> Int -> Maybe a

O(1) Safe indexing.

head :: Unbox a => Vector a -> a

O(1) First element.

last :: Unbox a => Vector a -> a

O(1) Last element.

unsafeIndex :: Unbox a => Vector a -> Int -> a

O(1) Unsafe indexing without bounds checking.

unsafeHead :: Unbox a => Vector a -> a

O(1) First element, without checking if the vector is empty.

unsafeLast :: Unbox a => Vector a -> a

O(1) Last element, without checking if the vector is empty.

Monadic indexing

indexM :: (Unbox a, Monad m) => Vector a -> Int -> m a

O(1) Indexing in a monad.

The monad allows operations to be strict in the vector when necessary. Suppose vector copying is implemented like this:

copy mv v = ... write mv i (v ! i) ...

For lazy vectors, v ! i would not be evaluated which means that mv would unnecessarily retain a reference to v in each element written.

With indexM, copying can be implemented like this instead:

copy mv v = ... do
                  x <- indexM v i
                  write mv i x

Here, no references to v are retained because indexing (but not the element) is evaluated eagerly.

headM :: (Unbox a, Monad m) => Vector a -> m a

O(1) First element of a vector in a monad. See indexM for an explanation of why this is useful.

lastM :: (Unbox a, Monad m) => Vector a -> m a

O(1) Last element of a vector in a monad. See indexM for an explanation of why this is useful.

unsafeIndexM :: (Unbox a, Monad m) => Vector a -> Int -> m a

O(1) Indexing in a monad, without bounds checks. See indexM for an explanation of why this is useful.

unsafeHeadM :: (Unbox a, Monad m) => Vector a -> m a

O(1) First element in a monad, without checking for empty vectors. See indexM for an explanation of why this is useful.

unsafeLastM :: (Unbox a, Monad m) => Vector a -> m a

O(1) Last element in a monad, without checking for empty vectors. See indexM for an explanation of why this is useful.

Extracting subvectors (slicing)

slice

Arguments

:: Unbox a

=> Int
i starting index

-> Int
n length

-> Vector a

-> Vector a

O(1) Yield a slice of the vector without copying it. The vector must contain at least i+n elements.

init :: Unbox a => Vector a -> Vector a

O(1) Yield all but the last element without copying. The vector may not be empty.

tail :: Unbox a => Vector a -> Vector a

O(1) Yield all but the first element without copying. The vector may not be empty.

take :: Unbox a => Int -> Vector a -> Vector a

O(1) Yield at the first n elements without copying. The vector may contain less than n elements, in which case it is returned unchanged.

drop :: Unbox a => Int -> Vector a -> Vector a

O(1) Yield all but the first n elements without copying. The vector may contain less than n elements, in which case an empty vector is returned.

splitAt :: Unbox a => Int -> Vector a -> (Vector a, Vector a)

O(1) Yield the first n elements paired with the remainder, without copying.

Note that splitAt n v is equivalent to (take n v, drop n v), but slightly more efficient.

Since: 0.7.1

uncons :: Unbox a => Vector a -> Maybe (a, Vector a)

O(1) Yield the head and tail of the vector, or Nothing if the vector is empty.

Since: 0.12.2.0

unsnoc :: Unbox a => Vector a -> Maybe (Vector a, a)

O(1) Yield the last and init of the vector, or Nothing if the vector is empty.

Since: 0.12.2.0

unsafeSlice

Arguments

:: Unbox a

=> Int
i starting index

-> Int
n length

-> Vector a

-> Vector a

O(1) Yield a slice of the vector without copying. The vector must contain at least i+n elements, but this is not checked.

unsafeInit :: Unbox a => Vector a -> Vector a

O(1) Yield all but the last element without copying. The vector may not be empty, but this is not checked.

unsafeTail :: Unbox a => Vector a -> Vector a

O(1) Yield all but the first element without copying. The vector may not be empty, but this is not checked.

unsafeTake :: Unbox a => Int -> Vector a -> Vector a

O(1) Yield the first n elements without copying. The vector must contain at least n elements, but this is not checked.

unsafeDrop :: Unbox a => Int -> Vector a -> Vector a

O(1) Yield all but the first n elements without copying. The vector must contain at least n elements, but this is not checked.

Construction

Initialisation

empty :: Unbox a => Vector a

O(1) The empty vector.

singleton :: Unbox a => a -> Vector a

O(1) A vector with exactly one element.

replicate :: Unbox a => Int -> a -> Vector a

O(n) A vector of the given length with the same value in each position.

generate :: Unbox a => Int -> (Int -> a) -> Vector a

O(n) Construct a vector of the given length by applying the function to each index.

iterateN :: Unbox a => Int -> (a -> a) -> a -> Vector a

O(n) Apply the function \(\max(n - 1, 0)\) times to an initial value, producing a vector of length \(\max(n, 0)\). The 0th element will contain the initial value, which is why there is one less function application than the number of elements in the produced vector.

\( \underbrace{x, f (x), f (f (x)), \ldots}_{\max(0,n)\rm{~elements}} \)

Examples

Expand

>>> import qualified Data.Vector.Unboxed as VU
>>> VU.iterateN 0 undefined undefined :: VU.Vector Int
[]
>>> VU.iterateN 3 (\(i, c) -> (pred i, succ c)) (0 :: Int, 'a')
[(0,'a'),(-1,'b'),(-2,'c')]

Since: 0.7.1

Monadic initialisation

replicateM :: (Monad m, Unbox a) => Int -> m a -> m (Vector a)

O(n) Execute the monadic action the given number of times and store the results in a vector.

generateM :: (Monad m, Unbox a) => Int -> (Int -> m a) -> m (Vector a)

O(n) Construct a vector of the given length by applying the monadic action to each index.

iterateNM :: (Monad m, Unbox a) => Int -> (a -> m a) -> a -> m (Vector a)

O(n) Apply the monadic function \(\max(n - 1, 0)\) times to an initial value, producing a vector of length \(\max(n, 0)\). The 0th element will contain the initial value, which is why there is one less function application than the number of elements in the produced vector.

For a non-monadic version, see iterateN.

Since: 0.12.0.0

create :: Unbox a => (forall s. ST s (MVector s a)) -> Vector a

Execute the monadic action and freeze the resulting vector.

create (do { v <- new 2; write v 0 'a'; write v 1 'b'; return v }) = <a,b>

createT :: (Traversable f, Unbox a) => (forall s. ST s (f (MVector s a))) -> f (Vector a)

Execute the monadic action and freeze the resulting vectors.

Unfolding

unfoldr :: Unbox a => (b -> Maybe (a, b)) -> b -> Vector a

O(n) Construct a vector by repeatedly applying the generator function to a seed. The generator function yields Just the next element and the new seed or Nothing if there are no more elements.

unfoldr (\n -> if n == 0 then Nothing else Just (n,n-1)) 10
 = <10,9,8,7,6,5,4,3,2,1>

unfoldrN :: Unbox a => Int -> (b -> Maybe (a, b)) -> b -> Vector a

O(n) Construct a vector with at most n elements by repeatedly applying the generator function to a seed. The generator function yields Just the next element and the new seed or Nothing if there are no more elements.

unfoldrN 3 (\n -> Just (n,n-1)) 10 = <10,9,8>

unfoldrExactN :: Unbox a => Int -> (b -> (a, b)) -> b -> Vector a

O(n) Construct a vector with exactly n elements by repeatedly applying the generator function to a seed. The generator function yields the next element and the new seed.

unfoldrExactN 3 (\n -> (n,n-1)) 10 = <10,9,8>

Since: 0.12.2.0

unfoldrM :: (Monad m, Unbox a) => (b -> m (Maybe (a, b))) -> b -> m (Vector a)

O(n) Construct a vector by repeatedly applying the monadic generator function to a seed. The generator function yields Just the next element and the new seed or Nothing if there are no more elements.

unfoldrNM :: (Monad m, Unbox a) => Int -> (b -> m (Maybe (a, b))) -> b -> m (Vector a)

O(n) Construct a vector by repeatedly applying the monadic generator function to a seed. The generator function yields Just the next element and the new seed or Nothing if there are no more elements.

unfoldrExactNM :: (Monad m, Unbox a) => Int -> (b -> m (a, b)) -> b -> m (Vector a)

O(n) Construct a vector with exactly n elements by repeatedly applying the monadic generator function to a seed. The generator function yields the next element and the new seed.

Since: 0.12.2.0

constructN :: Unbox a => Int -> (Vector a -> a) -> Vector a

O(n) Construct a vector with n elements by repeatedly applying the generator function to the already constructed part of the vector.

constructN 3 f = let a = f <> ; b = f <a> ; c = f <a,b> in <a,b,c>

constructrN :: Unbox a => Int -> (Vector a -> a) -> Vector a

O(n) Construct a vector with n elements from right to left by repeatedly applying the generator function to the already constructed part of the vector.

constructrN 3 f = let a = f <> ; b = f<a> ; c = f <b,a> in <c,b,a>

Enumeration

enumFromN :: (Unbox a, Num a) => a -> Int -> Vector a

O(n) Yield a vector of the given length, containing the values x, x+1 etc. This operation is usually more efficient than enumFromTo.

enumFromN 5 3 = <5,6,7>

enumFromStepN :: (Unbox a, Num a) => a -> a -> Int -> Vector a

O(n) Yield a vector of the given length, containing the values x, x+y, x+y+y etc. This operations is usually more efficient than enumFromThenTo.

enumFromStepN 1 2 5 = <1,3,5,7,9>

enumFromTo :: (Unbox a, Enum a) => a -> a -> Vector a

O(n) Enumerate values from x to y.

WARNING: This operation can be very inefficient. If possible, use enumFromN instead.

enumFromThenTo :: (Unbox a, Enum a) => a -> a -> a -> Vector a

O(n) Enumerate values from x to y with a specific step z.

WARNING: This operation can be very inefficient. If possible, use enumFromStepN instead.

Concatenation

cons :: Unbox a => a -> Vector a -> Vector a

O(n) Prepend an element.

snoc :: Unbox a => Vector a -> a -> Vector a

O(n) Append an element.

(++) :: Unbox a => Vector a -> Vector a -> Vector a infixr 5

O(m+n) Concatenate two vectors.

concat :: Unbox a => [Vector a] -> Vector a

O(n) Concatenate all vectors in the list.

Restricting memory usage

force :: Unbox a => Vector a -> Vector a

O(n) Yield the argument, but force it not to retain any extra memory, possibly by copying it.

This is especially useful when dealing with slices. For example:

force (slice 0 2 <huge vector>)

Here, the slice retains a reference to the huge vector. Forcing it creates a copy of just the elements that belong to the slice and allows the huge vector to be garbage collected.

Modifying vectors

Bulk updates

(//)

Arguments

:: Unbox a

=> Vector a
initial vector (of length m)

-> [(Int, a)]
list of index/value pairs (of length n)

-> Vector a

O(m+n) For each pair (i,a) from the list of idnex/value pairs, replace the vector element at position i by a.

<5,9,2,7> // [(2,1),(0,3),(2,8)] = <3,9,8,7>

update

Arguments

:: Unbox a

=> Vector a
initial vector (of length m)

-> Vector (Int, a)
vector of index/value pairs (of length n)

-> Vector a

O(m+n) For each pair (i,a) from the vector of index/value pairs, replace the vector element at position i by a.

update <5,9,2,7> <(2,1),(0,3),(2,8)> = <3,9,8,7>

update_

Arguments

:: Unbox a

=> Vector a
initial vector (of length m)

-> Vector Int
index vector (of length n1)

-> Vector a
value vector (of length n2)

-> Vector a

O(m+min(n1,n2)) For each index i from the index vector and the corresponding value a from the value vector, replace the element of the initial vector at position i by a.

update_ <5,9,2,7>  <2,0,2> <1,3,8> = <3,9,8,7>

The function update provides the same functionality and is usually more convenient.

update_ xs is ys = update xs (zip is ys)

unsafeUpd :: Unbox a => Vector a -> [(Int, a)] -> Vector a

Same as (//), but without bounds checking.

unsafeUpdate :: Unbox a => Vector a -> Vector (Int, a) -> Vector a

Same as update, but without bounds checking.

unsafeUpdate_ :: Unbox a => Vector a -> Vector Int -> Vector a -> Vector a

Same as update_, but without bounds checking.

Accumulations

accum

Arguments

:: Unbox a

=> (a -> b -> a)
accumulating function f

-> Vector a
initial vector (of length m)

-> [(Int, b)]
list of index/value pairs (of length n)

-> Vector a

O(m+n) For each pair (i,b) from the list, replace the vector element a at position i by f a b.

Examples

Expand

>>> import qualified Data.Vector.Unboxed as VU
>>> VU.accum (+) (VU.fromList [1000,2000,3000 :: Int]) [(2,4),(1,6),(0,3),(1,10)]
[1003,2016,3004]

accumulate

Arguments

:: (Unbox a, Unbox b)

=> (a -> b -> a)
accumulating function f

-> Vector a
initial vector (of length m)

-> Vector (Int, b)
vector of index/value pairs (of length n)

-> Vector a

O(m+n) For each pair (i,b) from the vector of pairs, replace the vector element a at position i by f a b.

Examples

Expand

>>> import qualified Data.Vector.Unboxed as VU
>>> VU.accumulate (+) (VU.fromList [1000,2000,3000 :: Int]) (VU.fromList [(2,4),(1,6),(0,3),(1,10)])
[1003,2016,3004]

accumulate_

Arguments

:: (Unbox a, Unbox b)

=> (a -> b -> a)
accumulating function f

-> Vector a
initial vector (of length m)

-> Vector Int
index vector (of length n1)

-> Vector b
value vector (of length n2)

-> Vector a

O(m+min(n1,n2)) For each index i from the index vector and the corresponding value b from the the value vector, replace the element of the initial vector at position i by f a b.

accumulate_ (+) <5,9,2> <2,1,0,1> <4,6,3,7> = <5+3, 9+6+7, 2+4>

The function accumulate provides the same functionality and is usually more convenient.

accumulate_ f as is bs = accumulate f as (zip is bs)

unsafeAccum :: Unbox a => (a -> b -> a) -> Vector a -> [(Int, b)] -> Vector a

Same as accum, but without bounds checking.

unsafeAccumulate :: (Unbox a, Unbox b) => (a -> b -> a) -> Vector a -> Vector (Int, b) -> Vector a

Same as accumulate, but without bounds checking.

unsafeAccumulate_ :: (Unbox a, Unbox b) => (a -> b -> a) -> Vector a -> Vector Int -> Vector b -> Vector a

Same as accumulate_, but without bounds checking.

Permutations

reverse :: Unbox a => Vector a -> Vector a

O(n) Reverse a vector.

backpermute :: Unbox a => Vector a -> Vector Int -> Vector a

O(n) Yield the vector obtained by replacing each element i of the index vector by xs!i. This is equivalent to map (xs!) is, but is often much more efficient.

backpermute <a,b,c,d> <0,3,2,3,1,0> = <a,d,c,d,b,a>

unsafeBackpermute :: Unbox a => Vector a -> Vector Int -> Vector a

Same as backpermute, but without bounds checking.

Safe destructive updates

modify :: Unbox a => (forall s. MVector s a -> ST s ()) -> Vector a -> Vector a

Apply a destructive operation to a vector. The operation may be performed in place if it is safe to do so and will modify a copy of the vector otherwise (see New for details).

Examples

Expand

>>> import qualified Data.Vector.Unboxed as VU
>>> import qualified Data.Vector.Unboxed.Mutable as MVU
>>> VU.modify (\v -> MVU.write v 0 'x') $ VU.replicate 4 'a'
"xaaa"

Elementwise operations

Indexing

indexed :: Unbox a => Vector a -> Vector (Int, a)

O(n) Pair each element in a vector with its index.

Mapping

map :: (Unbox a, Unbox b) => (a -> b) -> Vector a -> Vector b

O(n) Map a function over a vector.

imap :: (Unbox a, Unbox b) => (Int -> a -> b) -> Vector a -> Vector b

O(n) Apply a function to every element of a vector and its index.

concatMap :: (Unbox a, Unbox b) => (a -> Vector b) -> Vector a -> Vector b

Map a function over a vector and concatenate the results.

Monadic mapping

mapM :: (Monad m, Unbox a, Unbox b) => (a -> m b) -> Vector a -> m (Vector b)

O(n) Apply the monadic action to all elements of the vector, yielding a vector of results.

imapM :: (Monad m, Unbox a, Unbox b) => (Int -> a -> m b) -> Vector a -> m (Vector b)

O(n) Apply the monadic action to every element of a vector and its index, yielding a vector of results.

mapM_ :: (Monad m, Unbox a) => (a -> m b) -> Vector a -> m ()

O(n) Apply the monadic action to all elements of a vector and ignore the results.

imapM_ :: (Monad m, Unbox a) => (Int -> a -> m b) -> Vector a -> m ()

O(n) Apply the monadic action to every element of a vector and its index, ignoring the results.

forM :: (Monad m, Unbox a, Unbox b) => Vector a -> (a -> m b) -> m (Vector b)

O(n) Apply the monadic action to all elements of the vector, yielding a vector of results. Equivalent to flip mapM.

forM_ :: (Monad m, Unbox a) => Vector a -> (a -> m b) -> m ()

O(n) Apply the monadic action to all elements of a vector and ignore the results. Equivalent to flip mapM_.

iforM :: (Monad m, Unbox a, Unbox b) => Vector a -> (Int -> a -> m b) -> m (Vector b)

O(n) Apply the monadic action to all elements of the vector and their indices, yielding a vector of results. Equivalent to flip imapM.

Since: 0.12.2.0

iforM_ :: (Monad m, Unbox a) => Vector a -> (Int -> a -> m b) -> m ()

O(n) Apply the monadic action to all elements of the vector and their indices and ignore the results. Equivalent to flip imapM_.

Since: 0.12.2.0

Zipping

zipWith :: (Unbox a, Unbox b, Unbox c) => (a -> b -> c) -> Vector a -> Vector b -> Vector c

O(min(m,n)) Zip two vectors with the given function.

zipWith3 :: (Unbox a, Unbox b, Unbox c, Unbox d) => (a -> b -> c -> d) -> Vector a -> Vector b -> Vector c -> Vector d

Zip three vectors with the given function.

zipWith4 :: (Unbox a, Unbox b, Unbox c, Unbox d, Unbox e) => (a -> b -> c -> d -> e) -> Vector a -> Vector b -> Vector c -> Vector d -> Vector e

zipWith5 :: (Unbox a, Unbox b, Unbox c, Unbox d, Unbox e, Unbox f) => (a -> b -> c -> d -> e -> f) -> Vector a -> Vector b -> Vector c -> Vector d -> Vector e -> Vector f

zipWith6 :: (Unbox a, Unbox b, Unbox c, Unbox d, Unbox e, Unbox f, Unbox g) => (a -> b -> c -> d -> e -> f -> g) -> Vector a -> Vector b -> Vector c -> Vector d -> Vector e -> Vector f -> Vector g

izipWith :: (Unbox a, Unbox b, Unbox c) => (Int -> a -> b -> c) -> Vector a -> Vector b -> Vector c

O(min(m,n)) Zip two vectors with a function that also takes the elements' indices.

izipWith3 :: (Unbox a, Unbox b, Unbox c, Unbox d) => (Int -> a -> b -> c -> d) -> Vector a -> Vector b -> Vector c -> Vector d

Zip three vectors and their indices with the given function.

izipWith4 :: (Unbox a, Unbox b, Unbox c, Unbox d, Unbox e) => (Int -> a -> b -> c -> d -> e) -> Vector a -> Vector b -> Vector c -> Vector d -> Vector e

izipWith5 :: (Unbox a, Unbox b, Unbox c, Unbox d, Unbox e, Unbox f) => (Int -> a -> b -> c -> d -> e -> f) -> Vector a -> Vector b -> Vector c -> Vector d -> Vector e -> Vector f

izipWith6 :: (Unbox a, Unbox b, Unbox c, Unbox d, Unbox e, Unbox f, Unbox g) => (Int -> a -> b -> c -> d -> e -> f -> g) -> Vector a -> Vector b -> Vector c -> Vector d -> Vector e -> Vector f -> Vector g

zip :: (Unbox a, Unbox b) => Vector a -> Vector b -> Vector (a, b)

O(1) Zip 2 vectors.

zip3 :: (Unbox a, Unbox b, Unbox c) => Vector a -> Vector b -> Vector c -> Vector (a, b, c)

O(1) Zip 3 vectors.

zip4 :: (Unbox a, Unbox b, Unbox c, Unbox d) => Vector a -> Vector b -> Vector c -> Vector d -> Vector (a, b, c, d)

O(1) Zip 4 vectors.

zip5 :: (Unbox a, Unbox b, Unbox c, Unbox d, Unbox e) => Vector a -> Vector b -> Vector c -> Vector d -> Vector e -> Vector (a, b, c, d, e)

O(1) Zip 5 vectors.

zip6 :: (Unbox a, Unbox b, Unbox c, Unbox d, Unbox e, Unbox f) => Vector a -> Vector b -> Vector c -> Vector d -> Vector e -> Vector f -> Vector (a, b, c, d, e, f)

O(1) Zip 6 vectors.

Monadic zipping

zipWithM :: (Monad m, Unbox a, Unbox b, Unbox c) => (a -> b -> m c) -> Vector a -> Vector b -> m (Vector c)

O(min(m,n)) Zip the two vectors with the monadic action and yield a vector of results.

izipWithM :: (Monad m, Unbox a, Unbox b, Unbox c) => (Int -> a -> b -> m c) -> Vector a -> Vector b -> m (Vector c)

O(min(m,n)) Zip the two vectors with a monadic action that also takes the element index and yield a vector of results.

zipWithM_ :: (Monad m, Unbox a, Unbox b) => (a -> b -> m c) -> Vector a -> Vector b -> m ()

O(min(m,n)) Zip the two vectors with the monadic action and ignore the results.

izipWithM_ :: (Monad m, Unbox a, Unbox b) => (Int -> a -> b -> m c) -> Vector a -> Vector b -> m ()

O(min(m,n)) Zip the two vectors with a monadic action that also takes the element index and ignore the results.

Unzipping

unzip :: (Unbox a, Unbox b) => Vector (a, b) -> (Vector a, Vector b)

O(1) Unzip 2 vectors.

unzip3 :: (Unbox a, Unbox b, Unbox c) => Vector (a, b, c) -> (Vector a, Vector b, Vector c)

O(1) Unzip 3 vectors.

unzip4 :: (Unbox a, Unbox b, Unbox c, Unbox d) => Vector (a, b, c, d) -> (Vector a, Vector b, Vector c, Vector d)

O(1) Unzip 4 vectors.

unzip5 :: (Unbox a, Unbox b, Unbox c, Unbox d, Unbox e) => Vector (a, b, c, d, e) -> (Vector a, Vector b, Vector c, Vector d, Vector e)

O(1) Unzip 5 vectors.

unzip6 :: (Unbox a, Unbox b, Unbox c, Unbox d, Unbox e, Unbox f) => Vector (a, b, c, d, e, f) -> (Vector a, Vector b, Vector c, Vector d, Vector e, Vector f)

O(1) Unzip 6 vectors.

Working with predicates

Filtering

filter :: Unbox a => (a -> Bool) -> Vector a -> Vector a

O(n) Drop all elements that do not satisfy the predicate.

ifilter :: Unbox a => (Int -> a -> Bool) -> Vector a -> Vector a

O(n) Drop all elements that do not satisfy the predicate which is applied to the values and their indices.

filterM :: (Monad m, Unbox a) => (a -> m Bool) -> Vector a -> m (Vector a)

O(n) Drop all elements that do not satisfy the monadic predicate.

uniq :: (Unbox a, Eq a) => Vector a -> Vector a

O(n) Drop repeated adjacent elements. The first element in each group is returned.

Examples

Expand

>>> import qualified Data.Vector.Unboxed as VU
>>> VU.uniq $ VU.fromList [1,3,3,200,3 :: Int]
[1,3,200,3]
>>> import Data.Semigroup
>>> VU.uniq $ VU.fromList [ Arg 1 'a', Arg 1 'b', Arg (1 :: Int) 'c']
[Arg 1 'a']

mapMaybe :: (Unbox a, Unbox b) => (a -> Maybe b) -> Vector a -> Vector b

O(n) Map the values and collect the Just results.

imapMaybe :: (Unbox a, Unbox b) => (Int -> a -> Maybe b) -> Vector a -> Vector b

O(n) Map the indices/values and collect the Just results.

mapMaybeM :: (Monad m, Unbox a, Unbox b) => (a -> m (Maybe b)) -> Vector a -> m (Vector b)

O(n) Apply the monadic function to each element of the vector and discard elements returning Nothing.

Since: 0.12.2.0

imapMaybeM :: (Monad m, Unbox a, Unbox b) => (Int -> a -> m (Maybe b)) -> Vector a -> m (Vector b)

O(n) Apply the monadic function to each element of the vector and its index. Discard elements returning Nothing.

Since: 0.12.2.0

takeWhile :: Unbox a => (a -> Bool) -> Vector a -> Vector a

O(n) Yield the longest prefix of elements satisfying the predicate. The current implementation is not copy-free, unless the result vector is fused away.

dropWhile :: Unbox a => (a -> Bool) -> Vector a -> Vector a

O(n) Drop the longest prefix of elements that satisfy the predicate without copying.

Partitioning

partition :: Unbox a => (a -> Bool) -> Vector a -> (Vector a, Vector a)

O(n) Split the vector in two parts, the first one containing those elements that satisfy the predicate and the second one those that don't. The relative order of the elements is preserved at the cost of a sometimes reduced performance compared to unstablePartition.

unstablePartition :: Unbox a => (a -> Bool) -> Vector a -> (Vector a, Vector a)

O(n) Split the vector in two parts, the first one containing those elements that satisfy the predicate and the second one those that don't. The order of the elements is not preserved, but the operation is often faster than partition.

partitionWith :: (Unbox a, Unbox b, Unbox c) => (a -> Either b c) -> Vector a -> (Vector b, Vector c)

O(n) Split the vector into two parts, the first one containing the Left elements and the second containing the Right elements. The relative order of the elements is preserved.

Since: 0.12.1.0

span :: Unbox a => (a -> Bool) -> Vector a -> (Vector a, Vector a)

O(n) Split the vector into the longest prefix of elements that satisfy the predicate and the rest without copying.

break :: Unbox a => (a -> Bool) -> Vector a -> (Vector a, Vector a)

O(n) Split the vector into the longest prefix of elements that do not satisfy the predicate and the rest without copying.

groupBy :: Unbox a => (a -> a -> Bool) -> Vector a -> [Vector a]

O(n) Split a vector into a list of slices, using a predicate function.

The concatenation of this list of slices is equal to the argument vector, and each slice contains only equal elements, as determined by the equality predicate function.

Does not fuse.

>>> import qualified Data.Vector.Unboxed as VU
>>> import           Data.Char (isUpper)
>>> VU.groupBy (\a b -> isUpper a == isUpper b) (VU.fromList "Mississippi River")
["M","ississippi ","R","iver"]

See also groupBy, group.

Since: 0.13.0.1

group :: (Unbox a, Eq a) => Vector a -> [Vector a]

O(n) Split a vector into a list of slices of the input vector.

The concatenation of this list of slices is equal to the argument vector, and each slice contains only equal elements.

Does not fuse.

This is the equivalent of 'groupBy (==)'.

>>> import qualified Data.Vector.Unboxed as VU
>>> VU.group (VU.fromList "Mississippi")
["M","i","ss","i","ss","i","pp","i"]

See also group.

Since: 0.13.0.1

Searching

elem :: (Unbox a, Eq a) => a -> Vector a -> Bool infix 4

O(n) Check if the vector contains an element.

notElem :: (Unbox a, Eq a) => a -> Vector a -> Bool infix 4

O(n) Check if the vector does not contain an element (inverse of elem).

find :: Unbox a => (a -> Bool) -> Vector a -> Maybe a

O(n) Yield Just the first element matching the predicate or Nothing if no such element exists.

findIndex :: Unbox a => (a -> Bool) -> Vector a -> Maybe Int

O(n) Yield Just the index of the first element matching the predicate or Nothing if no such element exists.

findIndexR :: Unbox a => (a -> Bool) -> Vector a -> Maybe Int

O(n) Yield Just the index of the last element matching the predicate or Nothing if no such element exists.

Does not fuse.

findIndices :: Unbox a => (a -> Bool) -> Vector a -> Vector Int

O(n) Yield the indices of elements satisfying the predicate in ascending order.

elemIndex :: (Unbox a, Eq a) => a -> Vector a -> Maybe Int

O(n) Yield Just the index of the first occurrence of the given element or Nothing if the vector does not contain the element. This is a specialised version of findIndex.

elemIndices :: (Unbox a, Eq a) => a -> Vector a -> Vector Int

O(n) Yield the indices of all occurrences of the given element in ascending order. This is a specialised version of findIndices.

Folding

foldl :: Unbox b => (a -> b -> a) -> a -> Vector b -> a

O(n) Left fold.

foldl1 :: Unbox a => (a -> a -> a) -> Vector a -> a

O(n) Left fold on non-empty vectors.

foldl' :: Unbox b => (a -> b -> a) -> a -> Vector b -> a

O(n) Left fold with strict accumulator.

foldl1' :: Unbox a => (a -> a -> a) -> Vector a -> a

O(n) Left fold on non-empty vectors with strict accumulator.

foldr :: Unbox a => (a -> b -> b) -> b -> Vector a -> b

O(n) Right fold.

foldr1 :: Unbox a => (a -> a -> a) -> Vector a -> a

O(n) Right fold on non-empty vectors.

foldr' :: Unbox a => (a -> b -> b) -> b -> Vector a -> b

O(n) Right fold with a strict accumulator.

foldr1' :: Unbox a => (a -> a -> a) -> Vector a -> a

O(n) Right fold on non-empty vectors with strict accumulator.

ifoldl :: Unbox b => (a -> Int -> b -> a) -> a -> Vector b -> a

O(n) Left fold using a function applied to each element and its index.

ifoldl' :: Unbox b => (a -> Int -> b -> a) -> a -> Vector b -> a

O(n) Left fold with strict accumulator using a function applied to each element and its index.

ifoldr :: Unbox a => (Int -> a -> b -> b) -> b -> Vector a -> b

O(n) Right fold using a function applied to each element and its index.

ifoldr' :: Unbox a => (Int -> a -> b -> b) -> b -> Vector a -> b

O(n) Right fold with strict accumulator using a function applied to each element and its index.

foldMap :: (Monoid m, Unbox a) => (a -> m) -> Vector a -> m

O(n) Map each element of the structure to a monoid and combine the results. It uses the same implementation as the corresponding method of the Foldable type cless. Note that it's implemented in terms of foldr and won't fuse with functions that traverse the vector from left to right (map, generate, etc.).

Since: 0.12.2.0

foldMap' :: (Monoid m, Unbox a) => (a -> m) -> Vector a -> m

O(n) Like foldMap, but strict in the accumulator. It uses the same implementation as the corresponding method of the Foldable type class. Note that it's implemented in terms of foldl', so it fuses in most contexts.

Since: 0.12.2.0

Specialised folds

all :: Unbox a => (a -> Bool) -> Vector a -> Bool

O(n) Check if all elements satisfy the predicate.

Examples

Expand

>>> import qualified Data.Vector.Unboxed as VU
>>> VU.all even $ VU.fromList [2, 4, 12 :: Int]
True
>>> VU.all even $ VU.fromList [2, 4, 13 :: Int]
False
>>> VU.all even (VU.empty :: VU.Vector Int)
True

any :: Unbox a => (a -> Bool) -> Vector a -> Bool

O(n) Check if any element satisfies the predicate.

Examples

Expand

>>> import qualified Data.Vector.Unboxed as VU
>>> VU.any even $ VU.fromList [1, 3, 7 :: Int]
False
>>> VU.any even $ VU.fromList [3, 2, 13 :: Int]
True
>>> VU.any even (VU.empty :: VU.Vector Int)
False

and :: Vector Bool -> Bool

O(n) Check if all elements are True.

Examples

Expand

>>> import qualified Data.Vector.Unboxed as VU
>>> VU.and $ VU.fromList [True, False]
False
>>> VU.and VU.empty
True

or :: Vector Bool -> Bool

O(n) Check if any element is True.

Examples

Expand

>>> import qualified Data.Vector.Unboxed as VU
>>> VU.or $ VU.fromList [True, False]
True
>>> VU.or VU.empty
False

sum :: (Unbox a, Num a) => Vector a -> a

O(n) Compute the sum of the elements.

Examples

Expand

>>> import qualified Data.Vector.Unboxed as VU
>>> VU.sum $ VU.fromList [300,20,1 :: Int]
321
>>> VU.sum (VU.empty :: VU.Vector Int)
0

product :: (Unbox a, Num a) => Vector a -> a

O(n) Compute the product of the elements.

Examples

Expand

>>> import qualified Data.Vector.Unboxed as VU
>>> VU.product $ VU.fromList [1,2,3,4 :: Int]
24
>>> VU.product (VU.empty :: VU.Vector Int)
1

maximum :: (Unbox a, Ord a) => Vector a -> a

O(n) Yield the maximum element of the vector. The vector may not be empty. In case of a tie, the first occurrence wins.

Examples

Expand

>>> import qualified Data.Vector.Unboxed as VU
>>> VU.maximum $ VU.fromList [2, 1 :: Int]
2
>>> import Data.Semigroup
>>> VU.maximum $ VU.fromList [Arg 1 'a', Arg (2 :: Int) 'b']
Arg 2 'b'
>>> VU.maximum $ VU.fromList [Arg 1 'a', Arg (1 :: Int) 'b']
Arg 1 'a'

maximumBy :: Unbox a => (a -> a -> Ordering) -> Vector a -> a

O(n) Yield the maximum element of the vector according to the given comparison function. The vector may not be empty. In case of a tie, the first occurrence wins. This behavior is different from maximumBy which returns the last tie.

Examples

Expand

>>> import Data.Ord
>>> import qualified Data.Vector.Unboxed as VU
>>> VU.maximumBy (comparing fst) $ VU.fromList [(2,'a'), (1 :: Int,'b')]
(2,'a')
>>> VU.maximumBy (comparing fst) $ VU.fromList [(1,'a'), (1 :: Int,'b')]
(1,'a')

maximumOn :: (Ord b, Unbox a) => (a -> b) -> Vector a -> a

O(n) Yield the maximum element of the vector by comparing the results of a key function on each element. In case of a tie, the first occurrence wins. The vector may not be empty.

Examples

Expand

>>> import qualified Data.Vector.Unboxed as VU
>>> VU.maximumOn fst $ VU.fromList [(2,'a'), (1 :: Int,'b')]
(2,'a')
>>> VU.maximumOn fst $ VU.fromList [(1,'a'), (1 :: Int,'b')]
(1,'a')

Since: 0.13.0.0

minimum :: (Unbox a, Ord a) => Vector a -> a

O(n) Yield the minimum element of the vector. The vector may not be empty. In case of a tie, the first occurrence wins.

Examples

Expand

>>> import qualified Data.Vector.Unboxed as VU
>>> VU.minimum $ VU.fromList [2, 1 :: Int]
1
>>> import Data.Semigroup
>>> VU.minimum $ VU.fromList [Arg 2 'a', Arg (1 :: Int) 'b']
Arg 1 'b'
>>> VU.minimum $ VU.fromList [Arg 1 'a', Arg (1 :: Int) 'b']
Arg 1 'a'

minimumBy :: Unbox a => (a -> a -> Ordering) -> Vector a -> a

O(n) Yield the minimum element of the vector according to the given comparison function. The vector may not be empty.

O(n) Yield the minimum element of the vector according to the given comparison function. The vector may not be empty. In case of a tie, the first occurrence wins.

Examples

Expand

>>> import Data.Ord
>>> import qualified Data.Vector.Unboxed as VU
>>> VU.minimumBy (comparing fst) $ VU.fromList [(2,'a'), (1 :: Int,'b')]
(1,'b')
>>> VU.minimumBy (comparing fst) $ VU.fromList [(1,'a'), (1 :: Int,'b')]
(1,'a')

minimumOn :: (Ord b, Unbox a) => (a -> b) -> Vector a -> a

O(n) Yield the minimum element of the vector by comparing the results of a key function on each element. In case of a tie, the first occurrence wins. The vector may not be empty.

Examples

Expand

>>> import qualified Data.Vector.Unboxed as VU
>>> VU.minimumOn fst $ VU.fromList [(2,'a'), (1 :: Int,'b')]
(1,'b')
>>> VU.minimumOn fst $ VU.fromList [(1,'a'), (1 :: Int,'b')]
(1,'a')

Since: 0.13.0.0

minIndex :: (Unbox a, Ord a) => Vector a -> Int

O(n) Yield the index of the minimum element of the vector. The vector may not be empty.

minIndexBy :: Unbox a => (a -> a -> Ordering) -> Vector a -> Int

O(n) Yield the index of the minimum element of the vector according to the given comparison function. The vector may not be empty.

Examples

Expand

>>> import Data.Ord
>>> import qualified Data.Vector.Unboxed as VU
>>> VU.minIndexBy (comparing fst) $ VU.fromList [(2,'a'), (1,'b')]
1
>>> VU.minIndexBy (comparing fst) $ VU.fromList [(1,'a'), (1,'b')]
0

maxIndex :: (Unbox a, Ord a) => Vector a -> Int

O(n) Yield the index of the maximum element of the vector. The vector may not be empty.

maxIndexBy :: Unbox a => (a -> a -> Ordering) -> Vector a -> Int

O(n) Yield the index of the maximum element of the vector according to the given comparison function. The vector may not be empty. In case of a tie, the first occurrence wins.

Examples

Expand

>>> import Data.Ord
>>> import qualified Data.Vector.Unboxed as VU
>>> VU.maxIndexBy (comparing fst) $ VU.fromList [(2,'a'), (1,'b')]
0
>>> VU.maxIndexBy (comparing fst) $ VU.fromList [(1,'a'), (1,'b')]
0

Monadic folds

foldM :: (Monad m, Unbox b) => (a -> b -> m a) -> a -> Vector b -> m a

O(n) Monadic fold.

ifoldM :: (Monad m, Unbox b) => (a -> Int -> b -> m a) -> a -> Vector b -> m a

O(n) Monadic fold using a function applied to each element and its index.

foldM' :: (Monad m, Unbox b) => (a -> b -> m a) -> a -> Vector b -> m a

O(n) Monadic fold with strict accumulator.

ifoldM' :: (Monad m, Unbox b) => (a -> Int -> b -> m a) -> a -> Vector b -> m a

O(n) Monadic fold with strict accumulator using a function applied to each element and its index.

fold1M :: (Monad m, Unbox a) => (a -> a -> m a) -> Vector a -> m a

O(n) Monadic fold over non-empty vectors.

fold1M' :: (Monad m, Unbox a) => (a -> a -> m a) -> Vector a -> m a

O(n) Monadic fold over non-empty vectors with strict accumulator.

foldM_ :: (Monad m, Unbox b) => (a -> b -> m a) -> a -> Vector b -> m ()

O(n) Monadic fold that discards the result.

ifoldM_ :: (Monad m, Unbox b) => (a -> Int -> b -> m a) -> a -> Vector b -> m ()

O(n) Monadic fold that discards the result using a function applied to each element and its index.

foldM'_ :: (Monad m, Unbox b) => (a -> b -> m a) -> a -> Vector b -> m ()

O(n) Monadic fold with strict accumulator that discards the result.

ifoldM'_ :: (Monad m, Unbox b) => (a -> Int -> b -> m a) -> a -> Vector b -> m ()

O(n) Monadic fold with strict accumulator that discards the result using a function applied to each element and its index.

fold1M_ :: (Monad m, Unbox a) => (a -> a -> m a) -> Vector a -> m ()

O(n) Monadic fold over non-empty vectors that discards the result.

fold1M'_ :: (Monad m, Unbox a) => (a -> a -> m a) -> Vector a -> m ()

O(n) Monadic fold over non-empty vectors with strict accumulator that discards the result.

Scans

prescanl :: (Unbox a, Unbox b) => (a -> b -> a) -> a -> Vector b -> Vector a

O(n) Left-to-right prescan.

prescanl f z = init . scanl f z

Examples

Expand

>>> import qualified Data.Vector as VU
>>> VU.prescanl (+) 0 (VU.fromList [1,2,3,4 :: Int])
[0,1,3,6]

prescanl' :: (Unbox a, Unbox b) => (a -> b -> a) -> a -> Vector b -> Vector a

O(n) Left-to-right prescan with strict accumulator.

postscanl :: (Unbox a, Unbox b) => (a -> b -> a) -> a -> Vector b -> Vector a

O(n) Left-to-right postscan.

postscanl f z = tail . scanl f z

Examples

Expand

>>> import qualified Data.Vector.Unboxed as VU
>>> VU.postscanl (+) 0 (VU.fromList [1,2,3,4 :: Int])
[1,3,6,10]

postscanl' :: (Unbox a, Unbox b) => (a -> b -> a) -> a -> Vector b -> Vector a

O(n) Left-to-right postscan with strict accumulator.

scanl :: (Unbox a, Unbox b) => (a -> b -> a) -> a -> Vector b -> Vector a

O(n) Left-to-right scan.

scanl f z <x1,...,xn> = <y1,...,y(n+1)>
  where y1 = z
        yi = f y(i-1) x(i-1)

Examples

Expand

>>> import qualified Data.Vector.Unboxed as VU
>>> VU.scanl (+) 0 (VU.fromList [1,2,3,4 :: Int])
[0,1,3,6,10]

scanl' :: (Unbox a, Unbox b) => (a -> b -> a) -> a -> Vector b -> Vector a

O(n) Left-to-right scan with strict accumulator.

scanl1 :: Unbox a => (a -> a -> a) -> Vector a -> Vector a

O(n) Initial-value free left-to-right scan over a vector.

scanl f <x1,...,xn> = <y1,...,yn>
  where y1 = x1
        yi = f y(i-1) xi

Note: Since 0.13, application of this to an empty vector no longer results in an error; instead it produces an empty vector.

Examples

Expand

>>> import qualified Data.Vector.Unboxed as VU
>>> VU.scanl1 min $ VU.fromListN 5 [4,2,4,1,3 :: Int]
[4,2,2,1,1]
>>> VU.scanl1 max $ VU.fromListN 5 [1,3,2,5,4 :: Int]
[1,3,3,5,5]
>>> VU.scanl1 min (VU.empty :: VU.Vector Int)
[]

scanl1' :: Unbox a => (a -> a -> a) -> Vector a -> Vector a

O(n) Initial-value free left-to-right scan over a vector with a strict accumulator.

Note: Since 0.13, application of this to an empty vector no longer results in an error; instead it produces an empty vector.

Examples

Expand

>>> import qualified Data.Vector.Unboxed as VU
>>> VU.scanl1' min $ VU.fromListN 5 [4,2,4,1,3 :: Int]
[4,2,2,1,1]
>>> VU.scanl1' max $ VU.fromListN 5 [1,3,2,5,4 :: Int]
[1,3,3,5,5]
>>> VU.scanl1' min (VU.empty :: VU.Vector Int)
[]

iscanl :: (Unbox a, Unbox b) => (Int -> a -> b -> a) -> a -> Vector b -> Vector a

O(n) Left-to-right scan over a vector with its index.

Since: 0.12.2.0

iscanl' :: (Unbox a, Unbox b) => (Int -> a -> b -> a) -> a -> Vector b -> Vector a

O(n) Left-to-right scan over a vector (strictly) with its index.

Since: 0.12.2.0

prescanr :: (Unbox a, Unbox b) => (a -> b -> b) -> b -> Vector a -> Vector b

O(n) Right-to-left prescan.

prescanr f z = reverse . prescanl (flip f) z . reverse

prescanr' :: (Unbox a, Unbox b) => (a -> b -> b) -> b -> Vector a -> Vector b

O(n) Right-to-left prescan with strict accumulator.

postscanr :: (Unbox a, Unbox b) => (a -> b -> b) -> b -> Vector a -> Vector b

O(n) Right-to-left postscan.

postscanr' :: (Unbox a, Unbox b) => (a -> b -> b) -> b -> Vector a -> Vector b

O(n) Right-to-left postscan with strict accumulator.

scanr :: (Unbox a, Unbox b) => (a -> b -> b) -> b -> Vector a -> Vector b

O(n) Right-to-left scan.

scanr' :: (Unbox a, Unbox b) => (a -> b -> b) -> b -> Vector a -> Vector b

O(n) Right-to-left scan with strict accumulator.

scanr1 :: Unbox a => (a -> a -> a) -> Vector a -> Vector a

O(n) Right-to-left, initial-value free scan over a vector.

Note: Since 0.13, application of this to an empty vector no longer results in an error; instead it produces an empty vector.

Examples

Expand

>>> import qualified Data.Vector.Unboxed as VU
>>> VU.scanr1 min $ VU.fromListN 5 [3,1,4,2,4 :: Int]
[1,1,2,2,4]
>>> VU.scanr1 max $ VU.fromListN 5 [4,5,2,3,1 :: Int]
[5,5,3,3,1]
>>> VU.scanr1 min (VU.empty :: VU.Vector Int)
[]

scanr1' :: Unbox a => (a -> a -> a) -> Vector a -> Vector a

O(n) Right-to-left, initial-value free scan over a vector with a strict accumulator.

Note: Since 0.13, application of this to an empty vector no longer results in an error; instead it produces an empty vector.

Examples

Expand

>>> import qualified Data.Vector.Unboxed as VU
>>> VU.scanr1' min $ VU.fromListN 5 [3,1,4,2,4 :: Int]
[1,1,2,2,4]
>>> VU.scanr1' max $ VU.fromListN 5 [4,5,2,3,1 :: Int]
[5,5,3,3,1]
>>> VU.scanr1' min (VU.empty :: VU.Vector Int)
[]

iscanr :: (Unbox a, Unbox b) => (Int -> a -> b -> b) -> b -> Vector a -> Vector b

O(n) Right-to-left scan over a vector with its index.

Since: 0.12.2.0

iscanr' :: (Unbox a, Unbox b) => (Int -> a -> b -> b) -> b -> Vector a -> Vector b

O(n) Right-to-left scan over a vector (strictly) with its index.

@sinqce 0.12.2.0

Comparisons

eqBy :: (Unbox a, Unbox b) => (a -> b -> Bool) -> Vector a -> Vector b -> Bool

O(n) Check if two vectors are equal using the supplied equality predicate.

Since: 0.12.2.0

cmpBy :: (Unbox a, Unbox b) => (a -> b -> Ordering) -> Vector a -> Vector b -> Ordering

O(n) Compare two vectors using the supplied comparison function for vector elements. Comparison works the same as for lists.

cmpBy compare == compare

Since: 0.12.2.0

Conversions

Lists

toList :: Unbox a => Vector a -> [a]

O(n) Convert a vector to a list.

fromList :: Unbox a => [a] -> Vector a

O(n) Convert a list to a vector.

fromListN :: Unbox a => Int -> [a] -> Vector a

O(n) Convert the first n elements of a list to a vector. It's expected that the supplied list will be exactly n elements long. As an optimization, this function allocates a buffer for n elements, which could be used for DoS-attacks by exhausting the memory if an attacker controls that parameter.

fromListN n xs = fromList (take n xs)

Examples

Expand

>>> import qualified Data.Vector.Unboxed as VU
>>> VU.fromListN 3 [1,2,3,4,5 :: Int]
[1,2,3]
>>> VU.fromListN 3 [1 :: Int]
[1]

Other vector types

convert :: (Vector v a, Vector w a) => v a -> w a

O(n) Convert between different vector types.

Mutable vectors

freeze :: (Unbox a, PrimMonad m) => MVector (PrimState m) a -> m (Vector a)

O(n) Yield an immutable copy of the mutable vector.

thaw :: (Unbox a, PrimMonad m) => Vector a -> m (MVector (PrimState m) a)

O(n) Yield a mutable copy of an immutable vector.

copy :: (Unbox a, PrimMonad m) => MVector (PrimState m) a -> Vector a -> m ()

O(n) Copy an immutable vector into a mutable one. The two vectors must have the same length.

unsafeFreeze :: (Unbox a, PrimMonad m) => MVector (PrimState m) a -> m (Vector a)

O(1) Unsafely convert a mutable vector to an immutable one without copying. The mutable vector may not be used after this operation.

unsafeThaw :: (Unbox a, PrimMonad m) => Vector a -> m (MVector (PrimState m) a)

O(1) Unsafely convert an immutable vector to a mutable one without copying. Note that this is a very dangerous function and generally it's only safe to read from the resulting vector. In this case, the immutable vector could be used safely as well.

Problems with mutation happen because GHC has a lot of freedom to introduce sharing. As a result mutable vectors produced by unsafeThaw may or may not share the same underlying buffer. For example:

foo = do
  let vec = V.generate 10 id
  mvec <- V.unsafeThaw vec
  do_something mvec

Here GHC could lift vec outside of foo which means that all calls to do_something will use same buffer with possibly disastrous results. Whether such aliasing happens or not depends on the program in question, optimization levels, and GHC flags.

All in all, attempts to modify a vector produced by unsafeThaw fall out of domain of software engineering and into realm of black magic, dark rituals, and unspeakable horrors. The only advice that could be given is: "Don't attempt to mutate a vector produced by unsafeThaw unless you know how to prevent GHC from aliasing buffers accidentally. We don't."

unsafeCopy :: (Unbox a, PrimMonad m) => MVector (PrimState m) a -> Vector a -> m ()

O(n) Copy an immutable vector into a mutable one. The two vectors must have the same length. This is not checked.

Deriving via

newtype UnboxViaPrim a

Newtype wrapper which allows to derive unboxed vector in term of primitive vectors using DerivingVia mechanism. This is mostly used as illustration of use of DerivingVia for vector, see examples below.

First is rather straightforward: we define newtype and use GND to derive Prim instance. Newtype instances should be defined manually. Then we use deriving via to define necessary instances.

>>> :set -XTypeFamilies -XStandaloneDeriving -XDerivingVia -XMultiParamTypeClasses
>>> -- Needed to derive Prim
>>> :set -XGeneralizedNewtypeDeriving -XDataKinds -XUnboxedTuples -XPolyKinds
>>> 
>>> import qualified Data.Vector.Generic         as VG
>>> import qualified Data.Vector.Generic.Mutable as VGM
>>> import qualified Data.Vector.Primitive       as VP
>>> import qualified Data.Vector.Unboxed         as VU
>>> 
>>> newtype Foo = Foo Int deriving VP.Prim
>>> 
>>> newtype instance VU.MVector s Foo = MV_Int (VP.MVector s Foo)
>>> newtype instance VU.Vector    Foo = V_Int  (VP.Vector    Foo)
>>> deriving via (VU.UnboxViaPrim Foo) instance VGM.MVector VU.MVector Foo
>>> deriving via (VU.UnboxViaPrim Foo) instance VG.Vector   VU.Vector  Foo
>>> instance VU.Unbox Foo

Second example is essentially same but with a twist. Instead of using Prim instance of data type, we use underlying instance of Int:

>>> :set -XTypeFamilies -XStandaloneDeriving -XDerivingVia -XMultiParamTypeClasses
>>> 
>>> import qualified Data.Vector.Generic         as VG
>>> import qualified Data.Vector.Generic.Mutable as VGM
>>> import qualified Data.Vector.Primitive       as VP
>>> import qualified Data.Vector.Unboxed         as VU
>>> 
>>> newtype Foo = Foo Int
>>> 
>>> newtype instance VU.MVector s Foo = MV_Int (VP.MVector s Int)
>>> newtype instance VU.Vector    Foo = V_Int  (VP.Vector    Int)
>>> deriving via (VU.UnboxViaPrim Int) instance VGM.MVector VU.MVector Foo
>>> deriving via (VU.UnboxViaPrim Int) instance VG.Vector   VU.Vector  Foo
>>> instance VU.Unbox Foo

Since: 0.13.0.0

Constructors

UnboxViaPrim a

newtype As (a :: Type) (b :: Type)

Newtype which allows to derive unbox instances for type a which uses b as underlying representation (usually tuple). Type a and its representation b are connected by type class IsoUnbox. Here's example which uses explicit IsoUnbox instance:

>>> :set -XTypeFamilies -XStandaloneDeriving -XDerivingVia
>>> :set -XMultiParamTypeClasses -XTypeOperators -XFlexibleInstances
>>> import qualified Data.Vector.Unboxed         as VU
>>> import qualified Data.Vector.Unboxed.Mutable as MVU
>>> import qualified Data.Vector.Generic         as VG
>>> import qualified Data.Vector.Generic.Mutable as VGM
>>> :{
data Foo a = Foo Int a
  deriving Show
instance VU.IsoUnbox (Foo a) (Int,a) where
  toURepr (Foo i a) = (i,a)
  fromURepr (i,a) = Foo i a
  {-# INLINE toURepr #-}
  {-# INLINE fromURepr #-}
newtype instance VU.MVector s (Foo a) = MV_Foo (VU.MVector s (Int, a))
newtype instance VU.Vector    (Foo a) = V_Foo  (VU.Vector    (Int, a))
deriving via (Foo a `VU.As` (Int, a)) instance VU.Unbox a => VGM.MVector MVU.MVector (Foo a)
deriving via (Foo a `VU.As` (Int, a)) instance VU.Unbox a => VG.Vector   VU.Vector   (Foo a)
instance VU.Unbox a => VU.Unbox (Foo a)
:}

It's also possible to use generic-based instance for IsoUnbox which should work for all product types.

>>> :set -XMultiParamTypeClasses -XTypeOperators -XFlexibleInstances -XDeriveGeneric
>>> :set -XDerivingVia
>>> import qualified Data.Vector.Unboxed         as VU
>>> import qualified Data.Vector.Generic         as VG
>>> import qualified Data.Vector.Generic.Mutable as VGM
>>> :{
data Bar a = Bar Int a
  deriving (Show,Generic)
instance VU.IsoUnbox (Bar a) (Int,a) where
newtype instance VU.MVector s (Bar a) = MV_Bar (VU.MVector s (Int, a))
newtype instance VU.Vector    (Bar a) = V_Bar  (VU.Vector    (Int, a))
deriving via (Bar a `VU.As` (Int, a)) instance VU.Unbox a => VGM.MVector VU.MVector (Bar a)
deriving via (Bar a `VU.As` (Int, a)) instance VU.Unbox a => VG.Vector  VU.Vector  (Bar a)
instance VU.Unbox a => VU.Unbox (Bar a)
:}

Since: 0.13.0.0

Constructors

As a

class IsoUnbox a b where

Isomorphism between type a and its representation in unboxed vector b. Default instance coerces between generic representations of a and b which means they have same shape and corresponding fields could be coerced to each other. Note that this means it's possible to have fields that have different types:

>>> :set -XMultiParamTypeClasses -XDeriveGeneric -XFlexibleInstances
>>> import GHC.Generics (Generic)
>>> import Data.Monoid
>>> import qualified Data.Vector.Unboxed as VU
>>> :{
data Foo a = Foo Int a
  deriving (Show,Generic)
instance VU.IsoUnbox (Foo a) (Int, a)
instance VU.IsoUnbox (Foo a) (Sum Int, Product a)
:}

Since: 0.13.0.0

Minimal complete definition

Nothing

Methods

toURepr :: a -> b

Convert value into it representation in unboxed vector.

default toURepr :: (Generic a, Generic b, Coercible (Rep a ()) (Rep b ())) => a -> b

fromURepr :: b -> a

Convert value representation in unboxed vector back to value.

default fromURepr :: (Generic a, Generic b, Coercible (Rep b ()) (Rep a ())) => b -> a

Orphan instances

Unbox a => Monoid (Vector a)

Instance details

Methods

mempty :: Vector a

mappend :: Vector a -> Vector a -> Vector a

mconcat :: [Vector a] -> Vector a

Unbox a => Semigroup (Vector a)

Instance details

Methods

(<>) :: Vector a -> Vector a -> Vector a

sconcat :: NonEmpty (Vector a) -> Vector a

stimes :: Integral b => b -> Vector a -> Vector a

Unbox e => IsList (Vector e)

Instance details

Associated Types

type Item (Vector e)

Methods

fromList :: [Item (Vector e)] -> Vector e

fromListN :: Int -> [Item (Vector e)] -> Vector e

toList :: Vector e -> [Item (Vector e)]

(Read a, Unbox a) => Read (Vector a)

Instance details

Methods

readsPrec :: Int -> ReadS (Vector a)

readList :: ReadS [Vector a]

readPrec :: ReadPrec (Vector a)

readListPrec :: ReadPrec [Vector a]

(Show a, Unbox a) => Show (Vector a)

Instance details

Methods

showsPrec :: Int -> Vector a -> ShowS

show :: Vector a -> String

showList :: [Vector a] -> ShowS

(Unbox a, Eq a) => Eq (Vector a)

Instance details

Methods

(==) :: Vector a -> Vector a -> Bool

(/=) :: Vector a -> Vector a -> Bool

(Unbox a, Ord a) => Ord (Vector a)

Instance details

Methods

compare :: Vector a -> Vector a -> Ordering

(<) :: Vector a -> Vector a -> Bool

(<=) :: Vector a -> Vector a -> Bool

(>) :: Vector a -> Vector a -> Bool

(>=) :: Vector a -> Vector a -> Bool

max :: Vector a -> Vector a -> Vector a

min :: Vector a -> Vector a -> Vector a