Data.Vector.Generic.Mutable.Base

Description

Class of mutable vectors

Documentation

class MVector v a where

Class of mutable vectors parametrised with a primitive state token.

Minimal complete definition

basicLength, basicUnsafeSlice, basicOverlaps, basicUnsafeNew, basicInitialize, basicUnsafeRead, basicUnsafeWrite

Methods

basicLength :: v s a -> Int

Length of the mutable vector. This method should not be called directly, use length instead.

basicUnsafeSlice

Arguments

:: Int
starting index

-> Int
length of the slice

-> v s a

-> v s a

Yield a part of the mutable vector without copying it. This method should not be called directly, use unsafeSlice instead.

basicOverlaps :: v s a -> v s a -> Bool

Check whether two vectors overlap. This method should not be called directly, use overlaps instead.

basicUnsafeNew :: PrimMonad m => Int -> m (v (PrimState m) a)

Create a mutable vector of the given length. This method should not be called directly, use unsafeNew instead.

basicInitialize :: PrimMonad m => v (PrimState m) a -> m ()

Initialize a vector to a standard value. This is intended to be called as part of the safe new operation (and similar operations), to properly blank the newly allocated memory if necessary.

Vectors that are necessarily initialized as part of creation may implement this as a no-op.

basicUnsafeReplicate :: PrimMonad m => Int -> a -> m (v (PrimState m) a)

Create a mutable vector of the given length and fill it with an initial value. This method should not be called directly, use replicate instead.

basicUnsafeRead :: PrimMonad m => v (PrimState m) a -> Int -> m a

Yield the element at the given position. This method should not be called directly, use unsafeRead instead.

basicUnsafeWrite :: PrimMonad m => v (PrimState m) a -> Int -> a -> m ()

Replace the element at the given position. This method should not be called directly, use unsafeWrite instead.

basicClear :: PrimMonad m => v (PrimState m) a -> m ()

Reset all elements of the vector to some undefined value, clearing all references to external objects. This is usually a noop for unboxed vectors. This method should not be called directly, use clear instead.

basicSet :: PrimMonad m => v (PrimState m) a -> a -> m ()

Set all elements of the vector to the given value. This method should not be called directly, use set instead.

basicUnsafeCopy

Arguments

:: PrimMonad m

=> v (PrimState m) a
target

-> v (PrimState m) a
source

-> m ()

Copy a vector. The two vectors may not overlap. This method should not be called directly, use unsafeCopy instead.

basicUnsafeMove

Arguments

:: PrimMonad m

=> v (PrimState m) a
target

-> v (PrimState m) a
source

-> m ()

Move the contents of a vector. The two vectors may overlap. This method should not be called directly, use unsafeMove instead.

basicUnsafeGrow :: PrimMonad m => v (PrimState m) a -> Int -> m (v (PrimState m) a)

Grow a vector by the given number of elements. This method should not be called directly, use unsafeGrow instead.