Data.Vector.Storable

Description

Storable-based vectors.

Storable vectors

data Vector a

Storable-based vectors

data MVector s a

Mutable Storable-based vectors

Constructors

MVector !Int !(ForeignPtr a)

class Storable a

The member functions of this class facilitate writing values of primitive types to raw memory (which may have been allocated with the above mentioned routines) and reading values from blocks of raw memory. The class, furthermore, includes support for computing the storage requirements and alignment restrictions of storable types.

Memory addresses are represented as values of type Ptr a, for some a which is an instance of class Storable. The type argument to Ptr helps provide some valuable type safety in FFI code (you can't mix pointers of different types without an explicit cast), while helping the Haskell type system figure out which marshalling method is needed for a given pointer.

All marshalling between Haskell and a foreign language ultimately boils down to translating Haskell data structures into the binary representation of a corresponding data structure of the foreign language and vice versa. To code this marshalling in Haskell, it is necessary to manipulate primitive data types stored in unstructured memory blocks. The class Storable facilitates this manipulation on all types for which it is instantiated, which are the standard basic types of Haskell, the fixed size Int types (Int8, Int16, Int32, Int64), the fixed size Word types (Word8, Word16, Word32, Word64), StablePtr, all types from Foreign.C.Types, as well as Ptr.

Minimal complete definition

sizeOf, alignment, (peek | peekElemOff | peekByteOff), (poke | pokeElemOff | pokeByteOff)

Accessors

Length information

length :: Storable a => Vector a -> Int

O(1) Yield the length of the vector

null :: Storable a => Vector a -> Bool

O(1) Test whether a vector is empty

Indexing

(!) :: Storable a => Vector a -> Int -> a

O(1) Indexing

(!?) :: Storable a => Vector a -> Int -> Maybe a

O(1) Safe indexing

head :: Storable a => Vector a -> a

O(1) First element

last :: Storable a => Vector a -> a

O(1) Last element

unsafeIndex :: Storable a => Vector a -> Int -> a

O(1) Unsafe indexing without bounds checking

unsafeHead :: Storable a => Vector a -> a

O(1) First element without checking if the vector is empty

unsafeLast :: Storable a => Vector a -> a

O(1) Last element without checking if the vector is empty

Monadic indexing

indexM :: (Storable a, Monad m) => Vector a -> Int -> m a

O(1) Indexing in a monad.

The monad allows operations to be strict in the vector when necessary. Suppose vector copying is implemented like this:

copy mv v = ... write mv i (v ! i) ...

For lazy vectors, v ! i would not be evaluated which means that mv would unnecessarily retain a reference to v in each element written.

With indexM, copying can be implemented like this instead:

copy mv v = ... do
                  x <- indexM v i
                  write mv i x

Here, no references to v are retained because indexing (but not the elements) is evaluated eagerly.

headM :: (Storable a, Monad m) => Vector a -> m a

O(1) First element of a vector in a monad. See indexM for an explanation of why this is useful.

lastM :: (Storable a, Monad m) => Vector a -> m a

O(1) Last element of a vector in a monad. See indexM for an explanation of why this is useful.

unsafeIndexM :: (Storable a, Monad m) => Vector a -> Int -> m a

O(1) Indexing in a monad without bounds checks. See indexM for an explanation of why this is useful.

unsafeHeadM :: (Storable a, Monad m) => Vector a -> m a

O(1) First element in a monad without checking for empty vectors. See indexM for an explanation of why this is useful.

unsafeLastM :: (Storable a, Monad m) => Vector a -> m a

O(1) Last element in a monad without checking for empty vectors. See indexM for an explanation of why this is useful.

Extracting subvectors (slicing)

slice

Arguments

:: Storable a

=> Int
i starting index

-> Int
n length

-> Vector a

-> Vector a

O(1) Yield a slice of the vector without copying it. The vector must contain at least i+n elements.

init :: Storable a => Vector a -> Vector a

O(1) Yield all but the last element without copying. The vector may not be empty.

tail :: Storable a => Vector a -> Vector a

O(1) Yield all but the first element without copying. The vector may not be empty.

take :: Storable a => Int -> Vector a -> Vector a

O(1) Yield at the first n elements without copying. The vector may contain less than n elements in which case it is returned unchanged.

drop :: Storable a => Int -> Vector a -> Vector a

O(1) Yield all but the first n elements without copying. The vector may contain less than n elements in which case an empty vector is returned.

splitAt :: Storable a => Int -> Vector a -> (Vector a, Vector a)

O(1) Yield the first n elements paired with the remainder without copying.

Note that splitAt n v is equivalent to (take n v, drop n v) but slightly more efficient.

unsafeSlice

Arguments

:: Storable a

=> Int
i starting index

-> Int
n length

-> Vector a

-> Vector a

O(1) Yield a slice of the vector without copying. The vector must contain at least i+n elements but this is not checked.

unsafeInit :: Storable a => Vector a -> Vector a

O(1) Yield all but the last element without copying. The vector may not be empty but this is not checked.

unsafeTail :: Storable a => Vector a -> Vector a

O(1) Yield all but the first element without copying. The vector may not be empty but this is not checked.

unsafeTake :: Storable a => Int -> Vector a -> Vector a

O(1) Yield the first n elements without copying. The vector must contain at least n elements but this is not checked.

unsafeDrop :: Storable a => Int -> Vector a -> Vector a

O(1) Yield all but the first n elements without copying. The vector must contain at least n elements but this is not checked.

Construction

Initialisation

empty :: Storable a => Vector a

O(1) Empty vector

singleton :: Storable a => a -> Vector a

O(1) Vector with exactly one element

replicate :: Storable a => Int -> a -> Vector a

O(n) Vector of the given length with the same value in each position

generate :: Storable a => Int -> (Int -> a) -> Vector a

O(n) Construct a vector of the given length by applying the function to each index

iterateN :: Storable a => Int -> (a -> a) -> a -> Vector a

O(n) Apply function n times to value. Zeroth element is original value.

Monadic initialisation

replicateM :: (Monad m, Storable a) => Int -> m a -> m (Vector a)

O(n) Execute the monadic action the given number of times and store the results in a vector.

generateM :: (Monad m, Storable a) => Int -> (Int -> m a) -> m (Vector a)

O(n) Construct a vector of the given length by applying the monadic action to each index

iterateNM :: (Monad m, Storable a) => Int -> (a -> m a) -> a -> m (Vector a)

O(n) Apply monadic function n times to value. Zeroth element is original value.

create :: Storable a => (forall s. ST s (MVector s a)) -> Vector a

Execute the monadic action and freeze the resulting vector.

create (do { v <- new 2; write v 0 'a'; write v 1 'b'; return v }) = <a,b>

createT :: (Traversable f, Storable a) => (forall s. ST s (f (MVector s a))) -> f (Vector a)

Execute the monadic action and freeze the resulting vectors.

Unfolding

unfoldr :: Storable a => (b -> Maybe (a, b)) -> b -> Vector a

O(n) Construct a vector by repeatedly applying the generator function to a seed. The generator function yields Just the next element and the new seed or Nothing if there are no more elements.

unfoldr (\n -> if n == 0 then Nothing else Just (n,n-1)) 10
 = <10,9,8,7,6,5,4,3,2,1>

unfoldrN :: Storable a => Int -> (b -> Maybe (a, b)) -> b -> Vector a

O(n) Construct a vector with at most n elements by repeatedly applying the generator function to a seed. The generator function yields Just the next element and the new seed or Nothing if there are no more elements.

unfoldrN 3 (\n -> Just (n,n-1)) 10 = <10,9,8>

unfoldrM :: (Monad m, Storable a) => (b -> m (Maybe (a, b))) -> b -> m (Vector a)

O(n) Construct a vector by repeatedly applying the monadic generator function to a seed. The generator function yields Just the next element and the new seed or Nothing if there are no more elements.

unfoldrNM :: (Monad m, Storable a) => Int -> (b -> m (Maybe (a, b))) -> b -> m (Vector a)

O(n) Construct a vector by repeatedly applying the monadic generator function to a seed. The generator function yields Just the next element and the new seed or Nothing if there are no more elements.

constructN :: Storable a => Int -> (Vector a -> a) -> Vector a

O(n) Construct a vector with n elements by repeatedly applying the generator function to the already constructed part of the vector.

constructN 3 f = let a = f <> ; b = f <a> ; c = f <a,b> in f <a,b,c>

constructrN :: Storable a => Int -> (Vector a -> a) -> Vector a

O(n) Construct a vector with n elements from right to left by repeatedly applying the generator function to the already constructed part of the vector.

constructrN 3 f = let a = f <> ; b = f<a> ; c = f <b,a> in f <c,b,a>

Enumeration

enumFromN :: (Storable a, Num a) => a -> Int -> Vector a

O(n) Yield a vector of the given length containing the values x, x+1 etc. This operation is usually more efficient than enumFromTo.

enumFromN 5 3 = <5,6,7>

enumFromStepN :: (Storable a, Num a) => a -> a -> Int -> Vector a

O(n) Yield a vector of the given length containing the values x, x+y, x+y+y etc. This operations is usually more efficient than enumFromThenTo.

enumFromStepN 1 0.1 5 = <1,1.1,1.2,1.3,1.4>

enumFromTo :: (Storable a, Enum a) => a -> a -> Vector a

O(n) Enumerate values from x to y.

WARNING: This operation can be very inefficient. If at all possible, use enumFromN instead.

enumFromThenTo :: (Storable a, Enum a) => a -> a -> a -> Vector a

O(n) Enumerate values from x to y with a specific step z.

WARNING: This operation can be very inefficient. If at all possible, use enumFromStepN instead.

Concatenation

cons :: Storable a => a -> Vector a -> Vector a

O(n) Prepend an element

snoc :: Storable a => Vector a -> a -> Vector a

O(n) Append an element

(++) :: Storable a => Vector a -> Vector a -> Vector a infixr 5

O(m+n) Concatenate two vectors

concat :: Storable a => [Vector a] -> Vector a

O(n) Concatenate all vectors in the list

Restricting memory usage

force :: Storable a => Vector a -> Vector a

O(n) Yield the argument but force it not to retain any extra memory, possibly by copying it.

This is especially useful when dealing with slices. For example:

force (slice 0 2 <huge vector>)

Here, the slice retains a reference to the huge vector. Forcing it creates a copy of just the elements that belong to the slice and allows the huge vector to be garbage collected.

Modifying vectors

Bulk updates

(//)

Arguments

:: Storable a

=> Vector a
initial vector (of length m)

-> [(Int, a)]
list of index/value pairs (of length n)

-> Vector a

O(m+n) For each pair (i,a) from the list, replace the vector element at position i by a.

<5,9,2,7> // [(2,1),(0,3),(2,8)] = <3,9,8,7>

update_

Arguments

:: Storable a

=> Vector a
initial vector (of length m)

-> Vector Int
index vector (of length n1)

-> Vector a
value vector (of length n2)

-> Vector a

O(m+min(n1,n2)) For each index i from the index vector and the corresponding value a from the value vector, replace the element of the initial vector at position i by a.

update_ <5,9,2,7>  <2,0,2> <1,3,8> = <3,9,8,7>

unsafeUpd :: Storable a => Vector a -> [(Int, a)] -> Vector a

Same as (//) but without bounds checking.

unsafeUpdate_ :: Storable a => Vector a -> Vector Int -> Vector a -> Vector a

Same as update_ but without bounds checking.

Accumulations

accum

Arguments

:: Storable a

=> (a -> b -> a)
accumulating function f

-> Vector a
initial vector (of length m)

-> [(Int, b)]
list of index/value pairs (of length n)

-> Vector a

O(m+n) For each pair (i,b) from the list, replace the vector element a at position i by f a b.

accum (+) <5,9,2> [(2,4),(1,6),(0,3),(1,7)] = <5+3, 9+6+7, 2+4>

accumulate_

Arguments

:: (Storable a, Storable b)

=> (a -> b -> a)
accumulating function f

-> Vector a
initial vector (of length m)

-> Vector Int
index vector (of length n1)

-> Vector b
value vector (of length n2)

-> Vector a

O(m+min(n1,n2)) For each index i from the index vector and the corresponding value b from the the value vector, replace the element of the initial vector at position i by f a b.

accumulate_ (+) <5,9,2> <2,1,0,1> <4,6,3,7> = <5+3, 9+6+7, 2+4>

unsafeAccum :: Storable a => (a -> b -> a) -> Vector a -> [(Int, b)] -> Vector a

Same as accum but without bounds checking.

unsafeAccumulate_ :: (Storable a, Storable b) => (a -> b -> a) -> Vector a -> Vector Int -> Vector b -> Vector a

Same as accumulate_ but without bounds checking.

Permutations

reverse :: Storable a => Vector a -> Vector a

O(n) Reverse a vector

backpermute :: Storable a => Vector a -> Vector Int -> Vector a

O(n) Yield the vector obtained by replacing each element i of the index vector by xs!i. This is equivalent to map (xs!) is but is often much more efficient.

backpermute <a,b,c,d> <0,3,2,3,1,0> = <a,d,c,d,b,a>

unsafeBackpermute :: Storable a => Vector a -> Vector Int -> Vector a

Same as backpermute but without bounds checking.

Safe destructive updates

modify :: Storable a => (forall s. MVector s a -> ST s ()) -> Vector a -> Vector a

Apply a destructive operation to a vector. The operation will be performed in place if it is safe to do so and will modify a copy of the vector otherwise.

modify (\v -> write v 0 'x') (replicate 3 'a') = <'x','a','a'>

Elementwise operations

Mapping

map :: (Storable a, Storable b) => (a -> b) -> Vector a -> Vector b

O(n) Map a function over a vector

imap :: (Storable a, Storable b) => (Int -> a -> b) -> Vector a -> Vector b

O(n) Apply a function to every element of a vector and its index

concatMap :: (Storable a, Storable b) => (a -> Vector b) -> Vector a -> Vector b

Map a function over a vector and concatenate the results.

Monadic mapping

mapM :: (Monad m, Storable a, Storable b) => (a -> m b) -> Vector a -> m (Vector b)

O(n) Apply the monadic action to all elements of the vector, yielding a vector of results

mapM_ :: (Monad m, Storable a) => (a -> m b) -> Vector a -> m ()

O(n) Apply the monadic action to all elements of a vector and ignore the results

forM :: (Monad m, Storable a, Storable b) => Vector a -> (a -> m b) -> m (Vector b)

O(n) Apply the monadic action to all elements of the vector, yielding a vector of results. Equivalent to flip mapM.

forM_ :: (Monad m, Storable a) => Vector a -> (a -> m b) -> m ()

O(n) Apply the monadic action to all elements of a vector and ignore the results. Equivalent to flip mapM_.

Zipping

zipWith :: (Storable a, Storable b, Storable c) => (a -> b -> c) -> Vector a -> Vector b -> Vector c

O(min(m,n)) Zip two vectors with the given function.

zipWith3 :: (Storable a, Storable b, Storable c, Storable d) => (a -> b -> c -> d) -> Vector a -> Vector b -> Vector c -> Vector d

Zip three vectors with the given function.

zipWith4 :: (Storable a, Storable b, Storable c, Storable d, Storable e) => (a -> b -> c -> d -> e) -> Vector a -> Vector b -> Vector c -> Vector d -> Vector e

zipWith5 :: (Storable a, Storable b, Storable c, Storable d, Storable e, Storable f) => (a -> b -> c -> d -> e -> f) -> Vector a -> Vector b -> Vector c -> Vector d -> Vector e -> Vector f

zipWith6 :: (Storable a, Storable b, Storable c, Storable d, Storable e, Storable f, Storable g) => (a -> b -> c -> d -> e -> f -> g) -> Vector a -> Vector b -> Vector c -> Vector d -> Vector e -> Vector f -> Vector g

izipWith :: (Storable a, Storable b, Storable c) => (Int -> a -> b -> c) -> Vector a -> Vector b -> Vector c

O(min(m,n)) Zip two vectors with a function that also takes the elements' indices.

izipWith3 :: (Storable a, Storable b, Storable c, Storable d) => (Int -> a -> b -> c -> d) -> Vector a -> Vector b -> Vector c -> Vector d

Zip three vectors and their indices with the given function.

izipWith4 :: (Storable a, Storable b, Storable c, Storable d, Storable e) => (Int -> a -> b -> c -> d -> e) -> Vector a -> Vector b -> Vector c -> Vector d -> Vector e

izipWith5 :: (Storable a, Storable b, Storable c, Storable d, Storable e, Storable f) => (Int -> a -> b -> c -> d -> e -> f) -> Vector a -> Vector b -> Vector c -> Vector d -> Vector e -> Vector f

izipWith6 :: (Storable a, Storable b, Storable c, Storable d, Storable e, Storable f, Storable g) => (Int -> a -> b -> c -> d -> e -> f -> g) -> Vector a -> Vector b -> Vector c -> Vector d -> Vector e -> Vector f -> Vector g

Monadic zipping

zipWithM :: (Monad m, Storable a, Storable b, Storable c) => (a -> b -> m c) -> Vector a -> Vector b -> m (Vector c)

O(min(m,n)) Zip the two vectors with the monadic action and yield a vector of results

zipWithM_ :: (Monad m, Storable a, Storable b) => (a -> b -> m c) -> Vector a -> Vector b -> m ()

O(min(m,n)) Zip the two vectors with the monadic action and ignore the results

Working with predicates

Filtering

filter :: Storable a => (a -> Bool) -> Vector a -> Vector a

O(n) Drop elements that do not satisfy the predicate

ifilter :: Storable a => (Int -> a -> Bool) -> Vector a -> Vector a

O(n) Drop elements that do not satisfy the predicate which is applied to values and their indices

uniq :: (Storable a, Eq a) => Vector a -> Vector a

O(n) Drop repeated adjacent elements.

mapMaybe :: (Storable a, Storable b) => (a -> Maybe b) -> Vector a -> Vector b

O(n) Drop elements when predicate returns Nothing

imapMaybe :: (Storable a, Storable b) => (Int -> a -> Maybe b) -> Vector a -> Vector b

O(n) Drop elements when predicate, applied to index and value, returns Nothing

filterM :: (Monad m, Storable a) => (a -> m Bool) -> Vector a -> m (Vector a)

O(n) Drop elements that do not satisfy the monadic predicate

takeWhile :: Storable a => (a -> Bool) -> Vector a -> Vector a

O(n) Yield the longest prefix of elements satisfying the predicate without copying.

dropWhile :: Storable a => (a -> Bool) -> Vector a -> Vector a

O(n) Drop the longest prefix of elements that satisfy the predicate without copying.

Partitioning

partition :: Storable a => (a -> Bool) -> Vector a -> (Vector a, Vector a)

O(n) Split the vector in two parts, the first one containing those elements that satisfy the predicate and the second one those that don't. The relative order of the elements is preserved at the cost of a sometimes reduced performance compared to unstablePartition.

unstablePartition :: Storable a => (a -> Bool) -> Vector a -> (Vector a, Vector a)

O(n) Split the vector in two parts, the first one containing those elements that satisfy the predicate and the second one those that don't. The order of the elements is not preserved but the operation is often faster than partition.

span :: Storable a => (a -> Bool) -> Vector a -> (Vector a, Vector a)

O(n) Split the vector into the longest prefix of elements that satisfy the predicate and the rest without copying.

break :: Storable a => (a -> Bool) -> Vector a -> (Vector a, Vector a)

O(n) Split the vector into the longest prefix of elements that do not satisfy the predicate and the rest without copying.

Searching

elem :: (Storable a, Eq a) => a -> Vector a -> Bool infix 4

O(n) Check if the vector contains an element

notElem :: (Storable a, Eq a) => a -> Vector a -> Bool infix 4

O(n) Check if the vector does not contain an element (inverse of elem)

find :: Storable a => (a -> Bool) -> Vector a -> Maybe a

O(n) Yield Just the first element matching the predicate or Nothing if no such element exists.

findIndex :: Storable a => (a -> Bool) -> Vector a -> Maybe Int

O(n) Yield Just the index of the first element matching the predicate or Nothing if no such element exists.

findIndices :: Storable a => (a -> Bool) -> Vector a -> Vector Int

O(n) Yield the indices of elements satisfying the predicate in ascending order.

elemIndex :: (Storable a, Eq a) => a -> Vector a -> Maybe Int

O(n) Yield Just the index of the first occurence of the given element or Nothing if the vector does not contain the element. This is a specialised version of findIndex.

elemIndices :: (Storable a, Eq a) => a -> Vector a -> Vector Int

O(n) Yield the indices of all occurences of the given element in ascending order. This is a specialised version of findIndices.

Folding

foldl :: Storable b => (a -> b -> a) -> a -> Vector b -> a

O(n) Left fold

foldl1 :: Storable a => (a -> a -> a) -> Vector a -> a

O(n) Left fold on non-empty vectors

foldl' :: Storable b => (a -> b -> a) -> a -> Vector b -> a

O(n) Left fold with strict accumulator

foldl1' :: Storable a => (a -> a -> a) -> Vector a -> a

O(n) Left fold on non-empty vectors with strict accumulator

foldr :: Storable a => (a -> b -> b) -> b -> Vector a -> b

O(n) Right fold

foldr1 :: Storable a => (a -> a -> a) -> Vector a -> a

O(n) Right fold on non-empty vectors

foldr' :: Storable a => (a -> b -> b) -> b -> Vector a -> b

O(n) Right fold with a strict accumulator

foldr1' :: Storable a => (a -> a -> a) -> Vector a -> a

O(n) Right fold on non-empty vectors with strict accumulator

ifoldl :: Storable b => (a -> Int -> b -> a) -> a -> Vector b -> a

O(n) Left fold (function applied to each element and its index)

ifoldl' :: Storable b => (a -> Int -> b -> a) -> a -> Vector b -> a

O(n) Left fold with strict accumulator (function applied to each element and its index)

ifoldr :: Storable a => (Int -> a -> b -> b) -> b -> Vector a -> b

O(n) Right fold (function applied to each element and its index)

ifoldr' :: Storable a => (Int -> a -> b -> b) -> b -> Vector a -> b

O(n) Right fold with strict accumulator (function applied to each element and its index)

Specialised folds

all :: Storable a => (a -> Bool) -> Vector a -> Bool

O(n) Check if all elements satisfy the predicate.

any :: Storable a => (a -> Bool) -> Vector a -> Bool

O(n) Check if any element satisfies the predicate.

and :: Vector Bool -> Bool

O(n) Check if all elements are True

or :: Vector Bool -> Bool

O(n) Check if any element is True

sum :: (Storable a, Num a) => Vector a -> a

O(n) Compute the sum of the elements

product :: (Storable a, Num a) => Vector a -> a

O(n) Compute the produce of the elements

maximum :: (Storable a, Ord a) => Vector a -> a

O(n) Yield the maximum element of the vector. The vector may not be empty.

maximumBy :: Storable a => (a -> a -> Ordering) -> Vector a -> a

O(n) Yield the maximum element of the vector according to the given comparison function. The vector may not be empty.

minimum :: (Storable a, Ord a) => Vector a -> a

O(n) Yield the minimum element of the vector. The vector may not be empty.

minimumBy :: Storable a => (a -> a -> Ordering) -> Vector a -> a

O(n) Yield the minimum element of the vector according to the given comparison function. The vector may not be empty.

minIndex :: (Storable a, Ord a) => Vector a -> Int

O(n) Yield the index of the minimum element of the vector. The vector may not be empty.

minIndexBy :: Storable a => (a -> a -> Ordering) -> Vector a -> Int

O(n) Yield the index of the minimum element of the vector according to the given comparison function. The vector may not be empty.

maxIndex :: (Storable a, Ord a) => Vector a -> Int

O(n) Yield the index of the maximum element of the vector. The vector may not be empty.

maxIndexBy :: Storable a => (a -> a -> Ordering) -> Vector a -> Int

O(n) Yield the index of the maximum element of the vector according to the given comparison function. The vector may not be empty.

Monadic folds

foldM :: (Monad m, Storable b) => (a -> b -> m a) -> a -> Vector b -> m a

O(n) Monadic fold

foldM' :: (Monad m, Storable b) => (a -> b -> m a) -> a -> Vector b -> m a

O(n) Monadic fold with strict accumulator

fold1M :: (Monad m, Storable a) => (a -> a -> m a) -> Vector a -> m a

O(n) Monadic fold over non-empty vectors

fold1M' :: (Monad m, Storable a) => (a -> a -> m a) -> Vector a -> m a

O(n) Monadic fold over non-empty vectors with strict accumulator

foldM_ :: (Monad m, Storable b) => (a -> b -> m a) -> a -> Vector b -> m ()

O(n) Monadic fold that discards the result

foldM'_ :: (Monad m, Storable b) => (a -> b -> m a) -> a -> Vector b -> m ()

O(n) Monadic fold with strict accumulator that discards the result

fold1M_ :: (Monad m, Storable a) => (a -> a -> m a) -> Vector a -> m ()

O(n) Monadic fold over non-empty vectors that discards the result

fold1M'_ :: (Monad m, Storable a) => (a -> a -> m a) -> Vector a -> m ()

O(n) Monadic fold over non-empty vectors with strict accumulator that discards the result

Prefix sums (scans)

prescanl :: (Storable a, Storable b) => (a -> b -> a) -> a -> Vector b -> Vector a

O(n) Prescan

prescanl f z = init . scanl f z

Example: prescanl (+) 0 <1,2,3,4> = <0,1,3,6>

prescanl' :: (Storable a, Storable b) => (a -> b -> a) -> a -> Vector b -> Vector a

O(n) Prescan with strict accumulator

postscanl :: (Storable a, Storable b) => (a -> b -> a) -> a -> Vector b -> Vector a

O(n) Scan

postscanl f z = tail . scanl f z

Example: postscanl (+) 0 <1,2,3,4> = <1,3,6,10>

postscanl' :: (Storable a, Storable b) => (a -> b -> a) -> a -> Vector b -> Vector a

O(n) Scan with strict accumulator

scanl :: (Storable a, Storable b) => (a -> b -> a) -> a -> Vector b -> Vector a

O(n) Haskell-style scan

scanl f z <x1,...,xn> = <y1,...,y(n+1)>
  where y1 = z
        yi = f y(i-1) x(i-1)

Example: scanl (+) 0 <1,2,3,4> = <0,1,3,6,10>

scanl' :: (Storable a, Storable b) => (a -> b -> a) -> a -> Vector b -> Vector a

O(n) Haskell-style scan with strict accumulator

scanl1 :: Storable a => (a -> a -> a) -> Vector a -> Vector a

O(n) Scan over a non-empty vector

scanl f <x1,...,xn> = <y1,...,yn>
  where y1 = x1
        yi = f y(i-1) xi

scanl1' :: Storable a => (a -> a -> a) -> Vector a -> Vector a

O(n) Scan over a non-empty vector with a strict accumulator

prescanr :: (Storable a, Storable b) => (a -> b -> b) -> b -> Vector a -> Vector b

O(n) Right-to-left prescan

prescanr f z = reverse . prescanl (flip f) z . reverse

prescanr' :: (Storable a, Storable b) => (a -> b -> b) -> b -> Vector a -> Vector b

O(n) Right-to-left prescan with strict accumulator

postscanr :: (Storable a, Storable b) => (a -> b -> b) -> b -> Vector a -> Vector b

O(n) Right-to-left scan

postscanr' :: (Storable a, Storable b) => (a -> b -> b) -> b -> Vector a -> Vector b

O(n) Right-to-left scan with strict accumulator

scanr :: (Storable a, Storable b) => (a -> b -> b) -> b -> Vector a -> Vector b

O(n) Right-to-left Haskell-style scan

scanr' :: (Storable a, Storable b) => (a -> b -> b) -> b -> Vector a -> Vector b

O(n) Right-to-left Haskell-style scan with strict accumulator

scanr1 :: Storable a => (a -> a -> a) -> Vector a -> Vector a

O(n) Right-to-left scan over a non-empty vector

scanr1' :: Storable a => (a -> a -> a) -> Vector a -> Vector a

O(n) Right-to-left scan over a non-empty vector with a strict accumulator

Conversions

Lists

toList :: Storable a => Vector a -> [a]

O(n) Convert a vector to a list

fromList :: Storable a => [a] -> Vector a

O(n) Convert a list to a vector

fromListN :: Storable a => Int -> [a] -> Vector a

O(n) Convert the first n elements of a list to a vector

fromListN n xs = fromList (take n xs)

Other vector types

convert :: (Vector v a, Vector w a) => v a -> w a

O(n) Convert different vector types

unsafeCast :: forall a b. (Storable a, Storable b) => Vector a -> Vector b

O(1) Unsafely cast a vector from one element type to another. The operation just changes the type of the underlying pointer and does not modify the elements.

The resulting vector contains as many elements as can fit into the underlying memory block.

Mutable vectors

freeze :: (Storable a, PrimMonad m) => MVector (PrimState m) a -> m (Vector a)

O(n) Yield an immutable copy of the mutable vector.

thaw :: (Storable a, PrimMonad m) => Vector a -> m (MVector (PrimState m) a)

O(n) Yield a mutable copy of the immutable vector.

copy :: (Storable a, PrimMonad m) => MVector (PrimState m) a -> Vector a -> m ()

O(n) Copy an immutable vector into a mutable one. The two vectors must have the same length.

unsafeFreeze :: (Storable a, PrimMonad m) => MVector (PrimState m) a -> m (Vector a)

O(1) Unsafe convert a mutable vector to an immutable one without copying. The mutable vector may not be used after this operation.

unsafeThaw :: (Storable a, PrimMonad m) => Vector a -> m (MVector (PrimState m) a)

O(1) Unsafely convert an immutable vector to a mutable one without copying. The immutable vector may not be used after this operation.

unsafeCopy :: (Storable a, PrimMonad m) => MVector (PrimState m) a -> Vector a -> m ()

O(n) Copy an immutable vector into a mutable one. The two vectors must have the same length. This is not checked.

Raw pointers

unsafeFromForeignPtr

Arguments

:: Storable a

=> ForeignPtr a
pointer

-> Int
offset

-> Int
length

-> Vector a

O(1) Create a vector from a ForeignPtr with an offset and a length.

The data may not be modified through the ForeignPtr afterwards.

If your offset is 0 it is more efficient to use unsafeFromForeignPtr0.

unsafeFromForeignPtr0

Arguments

:: Storable a

=> ForeignPtr a
pointer

-> Int
length

-> Vector a

O(1) Create a vector from a ForeignPtr and a length.

It is assumed the pointer points directly to the data (no offset). Use unsafeFromForeignPtr if you need to specify an offset.

The data may not be modified through the ForeignPtr afterwards.

unsafeToForeignPtr :: Storable a => Vector a -> (ForeignPtr a, Int, Int)

O(1) Yield the underlying ForeignPtr together with the offset to the data and its length. The data may not be modified through the ForeignPtr.

unsafeToForeignPtr0 :: Storable a => Vector a -> (ForeignPtr a, Int)

O(1) Yield the underlying ForeignPtr together with its length.

You can assume the pointer points directly to the data (no offset).

The data may not be modified through the ForeignPtr.

unsafeWith :: Storable a => Vector a -> (Ptr a -> IO b) -> IO b

Pass a pointer to the vector's data to the IO action. The data may not be modified through the 'Ptr.