Data.Vector.Storable.Mutable

Description

Mutable vectors based on Storable.

Mutable vectors of Storable types

data MVector s a

Mutable Storable-based vectors

Constructors

MVector !Int !(ForeignPtr a)

type IOVector = MVector RealWorld

type STVector s = MVector s

class Storable a

The member functions of this class facilitate writing values of primitive types to raw memory (which may have been allocated with the above mentioned routines) and reading values from blocks of raw memory. The class, furthermore, includes support for computing the storage requirements and alignment restrictions of storable types.

Memory addresses are represented as values of type Ptr a, for some a which is an instance of class Storable. The type argument to Ptr helps provide some valuable type safety in FFI code (you can't mix pointers of different types without an explicit cast), while helping the Haskell type system figure out which marshalling method is needed for a given pointer.

All marshalling between Haskell and a foreign language ultimately boils down to translating Haskell data structures into the binary representation of a corresponding data structure of the foreign language and vice versa. To code this marshalling in Haskell, it is necessary to manipulate primitive data types stored in unstructured memory blocks. The class Storable facilitates this manipulation on all types for which it is instantiated, which are the standard basic types of Haskell, the fixed size Int types (Int8, Int16, Int32, Int64), the fixed size Word types (Word8, Word16, Word32, Word64), StablePtr, all types from Foreign.C.Types, as well as Ptr.

Minimal complete definition

sizeOf, alignment, (peek | peekElemOff | peekByteOff), (poke | pokeElemOff | pokeByteOff)

Accessors

Length information

length :: Storable a => MVector s a -> Int

Length of the mutable vector.

null :: Storable a => MVector s a -> Bool

Check whether the vector is empty

Extracting subvectors

slice :: Storable a => Int -> Int -> MVector s a -> MVector s a

Yield a part of the mutable vector without copying it.

init :: Storable a => MVector s a -> MVector s a

tail :: Storable a => MVector s a -> MVector s a

take :: Storable a => Int -> MVector s a -> MVector s a

drop :: Storable a => Int -> MVector s a -> MVector s a

splitAt :: Storable a => Int -> MVector s a -> (MVector s a, MVector s a)

unsafeSlice

Arguments

:: Storable a

=> Int
starting index

-> Int
length of the slice

-> MVector s a

-> MVector s a

Yield a part of the mutable vector without copying it. No bounds checks are performed.

unsafeInit :: Storable a => MVector s a -> MVector s a

unsafeTail :: Storable a => MVector s a -> MVector s a

unsafeTake :: Storable a => Int -> MVector s a -> MVector s a

unsafeDrop :: Storable a => Int -> MVector s a -> MVector s a

Overlapping

overlaps :: Storable a => MVector s a -> MVector s a -> Bool

Check whether two vectors overlap.

Construction

Initialisation

new :: (PrimMonad m, Storable a) => Int -> m (MVector (PrimState m) a)

Create a mutable vector of the given length.

unsafeNew :: (PrimMonad m, Storable a) => Int -> m (MVector (PrimState m) a)

Create a mutable vector of the given length. The length is not checked.

replicate :: (PrimMonad m, Storable a) => Int -> a -> m (MVector (PrimState m) a)

Create a mutable vector of the given length (0 if the length is negative) and fill it with an initial value.

replicateM :: (PrimMonad m, Storable a) => Int -> m a -> m (MVector (PrimState m) a)

Create a mutable vector of the given length (0 if the length is negative) and fill it with values produced by repeatedly executing the monadic action.

clone :: (PrimMonad m, Storable a) => MVector (PrimState m) a -> m (MVector (PrimState m) a)

Create a copy of a mutable vector.

Growing

grow :: (PrimMonad m, Storable a) => MVector (PrimState m) a -> Int -> m (MVector (PrimState m) a)

Grow a vector by the given number of elements. The number must be positive.

unsafeGrow :: (PrimMonad m, Storable a) => MVector (PrimState m) a -> Int -> m (MVector (PrimState m) a)

Grow a vector by the given number of elements. The number must be positive but this is not checked.

Restricting memory usage

clear :: (PrimMonad m, Storable a) => MVector (PrimState m) a -> m ()

Reset all elements of the vector to some undefined value, clearing all references to external objects. This is usually a noop for unboxed vectors.

Accessing individual elements

read :: (PrimMonad m, Storable a) => MVector (PrimState m) a -> Int -> m a

Yield the element at the given position.

write :: (PrimMonad m, Storable a) => MVector (PrimState m) a -> Int -> a -> m ()

Replace the element at the given position.

modify :: (PrimMonad m, Storable a) => MVector (PrimState m) a -> (a -> a) -> Int -> m ()

Modify the element at the given position.

swap :: (PrimMonad m, Storable a) => MVector (PrimState m) a -> Int -> Int -> m ()

Swap the elements at the given positions.

unsafeRead :: (PrimMonad m, Storable a) => MVector (PrimState m) a -> Int -> m a

Yield the element at the given position. No bounds checks are performed.

unsafeWrite :: (PrimMonad m, Storable a) => MVector (PrimState m) a -> Int -> a -> m ()

Replace the element at the given position. No bounds checks are performed.

unsafeModify :: (PrimMonad m, Storable a) => MVector (PrimState m) a -> (a -> a) -> Int -> m ()

Modify the element at the given position. No bounds checks are performed.

unsafeSwap :: (PrimMonad m, Storable a) => MVector (PrimState m) a -> Int -> Int -> m ()

Swap the elements at the given positions. No bounds checks are performed.

Modifying vectors

Filling and copying

set :: (PrimMonad m, Storable a) => MVector (PrimState m) a -> a -> m ()

Set all elements of the vector to the given value.

copy :: (PrimMonad m, Storable a) => MVector (PrimState m) a -> MVector (PrimState m) a -> m ()

Copy a vector. The two vectors must have the same length and may not overlap.

move :: (PrimMonad m, Storable a) => MVector (PrimState m) a -> MVector (PrimState m) a -> m ()

Move the contents of a vector. The two vectors must have the same length.

If the vectors do not overlap, then this is equivalent to copy. Otherwise, the copying is performed as if the source vector were copied to a temporary vector and then the temporary vector was copied to the target vector.

unsafeCopy

Arguments

:: (PrimMonad m, Storable a)

=> MVector (PrimState m) a
target

-> MVector (PrimState m) a
source

-> m ()

Copy a vector. The two vectors must have the same length and may not overlap. This is not checked.

unsafeMove

Arguments

:: (PrimMonad m, Storable a)

=> MVector (PrimState m) a
target

-> MVector (PrimState m) a
source

-> m ()

Move the contents of a vector. The two vectors must have the same length, but this is not checked.

If the vectors do not overlap, then this is equivalent to unsafeCopy. Otherwise, the copying is performed as if the source vector were copied to a temporary vector and then the temporary vector was copied to the target vector.

Unsafe conversions

unsafeCast :: forall a b s. (Storable a, Storable b) => MVector s a -> MVector s b

O(1) Unsafely cast a mutable vector from one element type to another. The operation just changes the type of the underlying pointer and does not modify the elements.

The resulting vector contains as many elements as can fit into the underlying memory block.

Raw pointers

unsafeFromForeignPtr

Arguments

:: Storable a

=> ForeignPtr a
pointer

-> Int
offset

-> Int
length

-> MVector s a

Create a mutable vector from a ForeignPtr with an offset and a length.

Modifying data through the ForeignPtr afterwards is unsafe if the vector could have been frozen before the modification.

If your offset is 0 it is more efficient to use unsafeFromForeignPtr0.

unsafeFromForeignPtr0

Arguments

:: Storable a

=> ForeignPtr a
pointer

-> Int
length

-> MVector s a

O(1) Create a mutable vector from a ForeignPtr and a length.

It is assumed the pointer points directly to the data (no offset). Use unsafeFromForeignPtr if you need to specify an offset.

Modifying data through the ForeignPtr afterwards is unsafe if the vector could have been frozen before the modification.

unsafeToForeignPtr :: Storable a => MVector s a -> (ForeignPtr a, Int, Int)

Yield the underlying ForeignPtr together with the offset to the data and its length. Modifying the data through the ForeignPtr is unsafe if the vector could have frozen before the modification.

unsafeToForeignPtr0 :: Storable a => MVector s a -> (ForeignPtr a, Int)

O(1) Yield the underlying ForeignPtr together with its length.

You can assume the pointer points directly to the data (no offset).

Modifying the data through the ForeignPtr is unsafe if the vector could have frozen before the modification.

unsafeWith :: Storable a => IOVector a -> (Ptr a -> IO b) -> IO b

Pass a pointer to the vector's data to the IO action. Modifying data through the pointer is unsafe if the vector could have been frozen before the modification.